% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{approx_gam_spline}
\alias{approx_gam_spline}
\title{Approximate temporal patterns using GAM with Cubic Regression Splines}
\usage{
approx_gam_spline(x, y, xout)
}
\arguments{
\item{x}{A numeric vector representing the independent variable (coordinates) of the points to be interpolated.}

\item{y}{A numeric vector representing the dependent variable (values) corresponding to each coordinate in \code{x}.}

\item{xout}{A numeric vector specifying the locations where interpolation predictions should be made.}
}
\value{
A numeric vector of predicted values at the \code{xout} locations based on the GAM cubic spline interpolation.
}
\description{
This function uses Generalized Additive Models (GAM) with cubic regression splines
to interpolate the provided data. It then predicts values at specified locations.
}
\seealso{
\code{\link[mgcv]{bam}} and \code{\link[mgcv]{s}} for details on the GAM with cubic regression splines.
}
\keyword{internal}
