% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{buildSciDBDTWQuery2}
\alias{buildSciDBDTWQuery2}
\title{Build SciDB 2 query (Faster)}
\usage{
buildSciDBDTWQuery2(INPUTARRAY, OUTPUTSCHEMA, PATTERNNAMES, XMIN, XMAX, YMIN,
  YMAX, TMIN, TMAX, COLIDS, ROWIDS, CHUNKCOL, CHUNKROW, THRESHOLD, FILL,
  NORMALIZE, REALDAY, METHOD, THETA, ALPHA, BETA, DELAY, NCORES, LIBRARYPATH,
  EXPPRPATH)
}
\arguments{
\item{INPUTARRAY}{}

\item{OUTPUTSCHEMA}{}

\item{PATTERNNAMES}{}

\item{XMIN}{}

\item{XMAX}{}

\item{YMIN}{}

\item{YMAX}{}

\item{TMIN}{}

\item{TMAX}{}

\item{COLIDS}{}

\item{ROWIDS}{}

\item{CHUNKCOL}{}

\item{CHUNKROW}{}

\item{THRESHOLD}{}

\item{FILL}{}

\item{NORMALIZE}{}

\item{REALDAY}{}

\item{METHOD}{}

\item{THETA}{}

\item{ALPHA}{}

\item{BETA}{}

\item{DELAY}{}

\item{NCORES}{}

\item{LIBRARYPATH}{}

\item{EXPPRPATH}{}
}
\description{
The function builds the SciDB query for time
series analysis.
}

