% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{buildSciDBPostProcSecVegetationQuery}
\alias{buildSciDBPostProcSecVegetationQuery}
\title{Build SciDB query}
\usage{
buildSciDBPostProcSecVegetationQuery(INPUTARRAY, INPUTSCHEMA, OUTPUTSCHEMA,
  PATTERNNAMES, XMIN, XMAX, YMIN, YMAX, TMIN, TMAX, FILL, NCORES, EXPPRPATH)
}
\arguments{
\item{INPUTARRAY}{}

\item{INPUTSCHEMA}{}

\item{OUTPUTSCHEMA}{}

\item{PATTERNNAMES}{}

\item{XMIN}{}

\item{XMAX}{}

\item{YMIN}{}

\item{YMAX}{}

\item{TMIN}{}

\item{TMAX}{}

\item{FILL}{}

\item{NCORES}{}

\item{EXPPRPATH}{}
}
\description{
The function builds the SciDB query for time
series analysis.
}

