% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyTimeSeries.R
\docType{methods}
\name{classifyTimeSeries}
\alias{classifyTimeSeries}
\alias{classifyTimeSeries,twdtwMatches-method}
\alias{classifyTimeSeries-twdtwTimeSeries}
\title{Classify time series}
\usage{
classifyTimeSeries(x, patterns.labels = NULL, from = NULL, to = NULL,
  by = NULL, breaks = NULL, overlap = 0.5, threshold = Inf,
  fill = "unclassified", simplify = FALSE, ...)

\S4method{classifyTimeSeries}{twdtwMatches}(x, patterns.labels = NULL,
  from = NULL, to = NULL, by = NULL, breaks = NULL, overlap = 0.5,
  threshold = Inf, fill = "unclassified", simplify = FALSE, ...)
}
\arguments{
\item{x}{an object of class twdtw*. This is the target time series. 
Usually, it is a set of unclassified time series.}

\item{patterns.labels}{a vector with labels of the patterns.}

\item{from}{A character or \code{\link[base]{Dates}} object in the format "yyyy-mm-dd".}

\item{to}{A \code{\link[base]{character}} or \code{\link[base]{Dates}} object in the format "yyyy-mm-dd".}

\item{by}{A \code{\link[base]{character}} with the intevals size, \emph{e.g.} "6 month".}

\item{breaks}{A vector of class \code{\link[base]{Dates}}. This replaces the arguments \code{from},
\code{to}, and \code{by}.}

\item{overlap}{A number between 0 and 1. The minimum overlapping 
between one match and the interval of classification. Default is 0.5, 
\emph{i.e.} an overlap minimum of 50\%.}

\item{threshold}{A number. The TWDTW dissimilarity threshold, i.e. the maximum TWDTW 
cost for consideration in the classification. Default is \code{Inf}.}

\item{fill}{a character or value to fill the classification gaps. Default is \code{fill="unclassified"}.}

\item{simplify}{logical. TRUE returns a list of labels with the best pattern for 
each period of time defined by \code{breaks}. FALSE returns an object of class twdtwMatches 
with the best matches. Default is FALSE.}

\item{...}{arguments to pass to specifique methods for each twdtw* signature 
and other arguments for parallel processing to pass}
}
\value{
An object of class twdtw*.
}
\description{
This function classifies the intervals of a time series 
based on the TWDTW results.
}
\examples{
# Applying TWDTW analysis to objects of class twdtwTimeSeries
ts = twdtwTimeSeries(example_ts)
patt = twdtwTimeSeries(patterns.list)
mat = twdtwApply(x=ts, y=patt, weight.fun=logisticWeight(-0.1, 100), keep=TRUE)
best_mat = classifyTimeSeries(mat, from="2009-09-01", to="2013-09-01", by="6 month", overlap=0.5) 

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwApply}}, 
\code{\link[dtwSat]{twdtwMatches-class}}, 
\code{\link[dtwSat]{twdtwTimeSeries-class}}, and 
\code{\link[dtwSat]{twdtwRaster-class}},
}

