% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtwSat.R
\docType{methods}
\name{computeAccuracy}
\alias{computeAccuracy}
\title{Computes the accuracy of a classification}
\usage{
computeAccuracy(predicted, reference)
}
\arguments{
\item{predicted}{A vector of prediction.}

\item{reference}{A vector of reference for the predicted classes.}
}
\description{
This function computs the accuracy a classification based
on the a given dataset of controll points.
}

