% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{computeDTWForAllPatterns}
\alias{computeDTWForAllPatterns}
\title{Computes DTW distance}
\usage{
computeDTWForAllPatterns(template, TemporalPatterns.list, ...)
}
\arguments{
\item{template}{A zoo object with the template time series.
It must be larger than the query.}

\item{TemporalPatterns.list}{A list of patterns. Each node of the list
has two columns data.frame with the dates and the value.}
}
\description{
The function computes the whole possible alignments
for a set of patterns in one time series.
}

