% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{computeGridTopologyList}
\alias{computeGridTopologyList}
\title{Computes a list of GridTopology objects}
\usage{
computeGridTopologyList(cellcentre.offset, cellsize, cells.dim, thread.size,
  ncores = 1)
}
\arguments{
\item{cellcentre.offset}{see ?GridTopology.}

\item{cellsize}{see ?GridTopology.}

\item{cells.dim}{see ?GridTopology.}

\item{thread.size}{integer, vector with number of cells in each thread.}

\item{ncores}{An integer with the number of cores. Default is 1.}
}
\description{
This function computes a list of GridTopology
objects with the cells to be processed in each thread.
}

