% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{computeThreadsSize}
\alias{computeThreadsSize}
\title{Splits the total number of processes}
\usage{
computeThreadsSize(n, thread.size)
}
\arguments{
\item{n}{An integer for the total number of processes.}

\item{thread.size}{An integer for the thread size.}
}
\description{
A recursive function to splits the total
number of processes according to the threads size.
}

