% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_patterns.R
\name{create_pattern}
\alias{create_pattern}
\title{Create patterns}
\usage{
create_pattern(x, from, to, freq, attr, formula, ...)
}
\arguments{
\item{x}{an object of class \code{\link[dtwSat]{twdtwTimeSeries}}.}

\item{from}{A character or \code{\link[base]{Dates}} object in the format
"yyyy-mm-dd". If not provided it is equal to the smallest date of the
first element in x. See details.}

\item{to}{A \code{\link[base]{character}} or \code{\link[base]{Dates}}
object in the format "yyyy-mm-dd". If not provided it is equal to the
greatest date of the first element in x. See details.}

\item{freq}{An integer. The sampling frequency of the output patterns.}

\item{attr}{A vector character or numeric. The attributes in \code{x} to be used.
If not declared the function uses all attributes.}

\item{formula}{A formula. Argument to pass to \code{\link[mgcv]{gam}}.}

\item{...}{other arguments to pass to the function \code{\link[mgcv]{gam}} in the
package \pkg{mgcv}.}

\item{split}{A logical. If TRUE the samples are split by label. If FALSE
all samples are set to the same label.}
}
\value{
an object of class \code{\link[dtwSat]{twdtwTimeSeries}}
}
\description{
Create temporal patterns from objects of class twdtwTimeSeries.
}
\details{
The hidden assumption is that the temporal pattern is a cycle the repeats itself
within a given time interval. Therefore, all time series samples in \code{x} are aligned
with each other, keeping their respective sequence of days of the year. The function fits a
Generalized Additive Model (GAM) to the aligned set of samples.
}
\references{
\insertRef{Maus:2019}{dtwSat}

\insertRef{Maus:2016}{dtwSat}
}
\seealso{
\code{\link[dtwSat]{twdtwMatches-class}},
\code{\link[dtwSat]{twdtwTimeSeries-class}},
\code{\link[dtwSat]{getTimeSeries}}, and
\code{\link[dtwSat]{twdtwApply}}
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
