% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_patterns.R
\name{create_patterns}
\alias{create_patterns}
\title{Create a Pattern Using GAM}
\usage{
create_patterns(
  x,
  y,
  formula = band ~ s(time),
  start_column = "start_date",
  end_column = "end_date",
  label_colum = "label",
  sampling_freq = NULL,
  ...
)
}
\arguments{
\item{x}{A three dimensions stars object (x, y, time) with the satellite image time series.}

\item{y}{An sf object with the coordinates of the training points.}

\item{formula}{A formula for the GAM. Default is \code{band ~ \link[mgcv]{s}(time)}.}

\item{start_column}{Name of the column in y that indicates the start date. Default is 'start_date'.}

\item{end_column}{Name of the column in y that indicates the end date. Default is 'end_date'.}

\item{label_colum}{Name of the column in y that contains land use labels. Default is 'label'.}

\item{sampling_freq}{The time frequency for sampling including unit, e.g '16 day'. If NULL, the function will infer it.}

\item{...}{Additional arguments passed to the GAM function.}
}
\value{
A list containing the predicted values for each label.
}
\description{
This function creates a pattern based on Generalized Additive Models (GAM).
It uses the specified formula to fit the model and predict values.
}
