% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\docType{methods}
\name{createPattern}
\alias{createPattern}
\title{Create temporal patterns}
\usage{
createPattern(x, freq, mc.cores = 1, ...)
}
\arguments{
\item{x}{A \code{\link[base]{list}} of \code{\link[base]{data.frame}} such as 
retrived by \code{\link[dtwSat]{extractSampleTimeSeries}}.}

\item{freq}{An integer. The frequency in days of the output patterns}

\item{mc.cores}{The number of cores to use, See \code{\link[parallel]{mclapply}} 
for details}

\item{...}{other arguments to pass to the function \code{\link[mgcv]{gam}} in the 
packege \pkg{mgcv}}
}
\value{
A list temporal pattern as \code{\link[zoo]{zoo}} objects, 
one for each class in \code{x}
}
\description{
This function fits a gam model and retrieves a smoothed 
temporal pattern
}
\examples{

###


}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

