% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPatterns.R
\docType{methods}
\name{createPatterns}
\alias{createPatterns}
\title{Create temporal patterns}
\usage{
createPatterns(x, ref, from, to, freq = 1, attr, formula, ...)
}
\arguments{
\item{x}{A \code{\link[base]{list}} of \code{\link[zoo]{zoo}} such as 
retrived by \code{\link[dtwSat]{extractTimeSeries}}.}

\item{ref}{A vector identifying each time series. It must have the same length 
as \code{x}. If \code{ref} is not declared then all samples are taken as belonging 
to the same class.}

\item{from}{A character or \code{\link[base]{Dates}} object in the format 
"yyyy-mm-dd". If not informed it is equal to the smallest date of the 
first element in x. See details.}

\item{to}{A \code{\link[base]{character}} or \code{\link[base]{Dates}} 
object in the format "yyyy-mm-dd". If not informed it is equal to the 
greatest date of the first element in x. See details.}

\item{freq}{An integer. The sampling frequency of the output patterns.}

\item{attr}{A vector character or numeric. The attributes in \code{x} to be used. 
If not declared the function uses all attributes.}

\item{formula}{A formula. Argument to pass to \code{\link[mgcv]{gam}}.}

\item{...}{other arguments to pass to the function \code{\link[mgcv]{gam}} in the 
packege \pkg{mgcv}.}
}
\value{
A \code{\link[zoo]{zoo}} object
}
\description{
This function fits a Generalized Additive Model model 
to the samples and retrieves smoothed temporal patterns.
}
\details{
The hidden assumption is that the temporal pattern is a cycle the repeats itself 
within a given time interval. Therefore, all time series samples in \code{x} are aligned 
to each other keeping the time sequence of each sample and its respective day of the year. 
After that the function fits a Generalized Additive Model to the samples.
}
\examples{

\dontrun{
field_samples = read.csv(system.file("lucc_MT/data/samples.csv", package = "dtwSat"))
load(system.file("lucc_MT/field_samples_ts.RData", package = "dtwSat"))

# Sellect Soybean-cotton samples  
I = field_samples$class=="Soybean-cotton"
soybean_cotton_samples = field_samples_ts[I]

p_soybean_cotton = createPatterns(x = soybean_cotton_samples, 
                   formula = y ~ s(x))

plotPatterns(p_soybean_cotton)

p = createPatterns(x = field_samples_ts, ref = field_samples$class,
                   formula = y ~ s(x))
plotPatterns(p)

}
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

