% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{createRasterTS}
\alias{createRasterTS}
\title{Create raster time series}
\usage{
createRasterTS(..., x = list(...), timeline, doy)
}
\arguments{
\item{...}{\code{\link[raster]{RasterBrick-class}} or \code{\link[raster]{RasterStack-class}}}

\item{x}{A list of \code{\link[raster]{RasterBrick-class}} or 
\code{\link[raster]{RasterStack-class}}
Each layer of the Raster* object is a time step}

\item{timeline}{A vector of \code{\link[base]{Dates}}
It must have the length of the layers in the \code{\link[raster]{RasterBrick-class}} object}

\item{doy}{optional. A \code{\link[raster]{RasterBrick-class}} or 
\code{\link[raster]{RasterStack-class}} with the same extent as the 
objects in \code{x}}
}
\value{
A \code{\link[dtwSat]{twdtw-class}} object
}
\description{
This function creates a list of raster time series
}
\details{
a list of \code{\link[raster]{RasterBrick-class}} or 
\code{\link[raster]{RasterStack-class}} objects
}
\examples{

####

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{classifyIntervals}}
}

