% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{createTimeSequence}
\alias{createTimeSequence}
\title{Create time sequence}
\usage{
createTimeSequence(year = 2000:format(Sys.time(), "\%Y"), frequency = 16)
}
\arguments{
\item{year}{A vector with the years. Default
is form 2000 to the system time year \code{format(Sys.time(), ''\%Y'')}}

\item{frequency}{An integer with the frequency in days. Default is 16 days}
}
\value{
vector of class \code{\link[base]{Date}}
}
\description{
This function creates a sequence of dates for
each year. The sequences start on January 1st of each year.
}
\examples{
dates = createTimeSequence()
dates
}
\seealso{
\link[dtwSat]{getDatesFromDOY} and \link[dtwSat]{getModisTimeIndex}
}

