% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\name{dtwSat-class}
\alias{dtwSat-class}
\title{dtwSat-class}
\description{
Class for Multidimensional Time-Weighted DTW
}
\section{Slots }{

\describe{
 \item{\code{call}:}{Object of class \code{call} see
 \code{\link[base]{match.call}}.}
 \item{\code{alignments}:}{A named \code{list} where each node has one
     attribute whose lengths are identical to the number of
	    alignments. The attributes are:
	    \code{from} the alignment starts,
	    \code{to} the alignment ends,
	    \code{distance} the DTW distance, and
	    \code{normalizedDistance} the normalized DTW distance.
 }
 \item{\code{mapping}:}{ Object of class \code{list} where each node
 has the matching points between the query and the template time series.}
	\item{\code{internals}:}{ Object of class \code{list} similar to
	the \code{\link[dtw]{dtw}} class from package \pkg{dtw}.}
}
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{dtwSat}}, \code{\link[dtwSat]{twdtw}},
\code{\link[dtwSat]{plotPath}}, \code{\link[dtwSat]{plotAlignment}}
}

