% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtwSatClass.R
\docType{class}
\name{dtwSat}
\alias{dtwSat}
\alias{dtwSat-class}
\title{dtwSat class}
\description{
Class for Multidimensional Time-Weighted DTW results
}
\section{Slots }{

\describe{
\item{\code{call}:}{Object of class \code{call} see
\code{\link[base]{match.call}}}
\item{\code{alignments}:}{Object of class \code{list}. Each node of
the list has one attribute, whose lengths are identical to the number of
alignments. \code{from} the alignment starts, \code{to} the alignment ends,
\code{distance} the DTW distance, \code{normalizedDistance} the
normalized DTW distance}
\item{\code{mapping}:}{Object of class \code{list}. Each node of the list has
the matching points of the query to the template time series.}
\item{\code{internals}:}{Object of class \code{list} similar to
\code{\link[dtw]{dtw}} class from package \pkg{dtw}}
}
}
\author{
Victor Maus
}

