% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_ts.list}
\alias{example_ts.list}
\title{Data: A list of satellite time series}
\format{Each element in the \link[base]{list} has a \link[zoo]{zoo} object, 
whose indices are \code{\link[base]{Dates}} in the format ''yyyy-mm-dd''. 
Each node has 6 attributes: ''ndvi'', ''evi'', ''red'', ''nir'', ''blue'', 
and ''mir''.}
\usage{
example_ts.list
}
\description{
This dataset has a list of time series based 
on the MODIS product MOD13Q1 250 m 16 days. It is an irregular time 
series using the real date of each pixel.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{example_ts}, 
\link[dtwSat]{patterns.list}.

The package \pkg{rwtss} provides a client for Web Time 
Series Service \url{https://github.com/albhasan/rwtss.git}.

MOD13Q1 documentation: 
\url{https://lpdaac.usgs.gov/dataset_discovery/modis/modis_products_table/mod13q1}
}
\keyword{datasets}

