% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_patterns.R
\name{extract_time_series}
\alias{extract_time_series}
\title{Extract Time Series from a Stars Object for Specified Points}
\usage{
extract_time_series(x, y)
}
\arguments{
\item{x}{A stars object containing the raster time series data.}

\item{y}{An sf object containing the point geometries and their associated labels.}
}
\value{
A data.frame with the extracted time series for each point in the sf object,
with additional columns for the ID and label of each sample.
}
\description{
This function extracts a time series from a stars object for each specified point in the sf object.
Each extracted sample is then labeled with an ID and the label from the sf object.
}
