% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\docType{methods}
\name{extractSampleTimeSeries}
\alias{extractSampleTimeSeries}
\title{Extract time series from raster}
\usage{
extractSampleTimeSeries(x, y, proj4string = CRS(as.character(NA)),
  mc.cores = 1)
}
\arguments{
\item{x}{A list of \code{\link[raster]{RasterBrick-class}} or 
\code{\link[raster]{RasterStack-class}} objects}

\item{y}{A \code{\link[base]{data.frame}} whose attributes are: longitude, 
latitude, the start ''from'' and the end ''to'' of the time interval 
for each sample. This can also be a \code{\link[sp]{SpatialPointsDataFrame}} 
whose attributes are the start ''from'' and the end ''to'' of the time interval}

\item{proj4string}{projection string, see \code{\link[sp]{CRS-class}}. Used 
if \code{y} is a \code{\link[base]{data.frame}}}

\item{mc.cores}{The number of cores to use, See \code{\link[parallel]{mclapply}} 
for details}
}
\value{
A list of \code{\link[base]{data.frame}} objects, one for each class in \code{x}
}
\description{
This function extracts the time series from a list of 
raster given set of spatial location
}
\examples{

###


}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

