% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{get_dates_from_doy}
\alias{get_dates_from_doy}
\title{Get dates from year and day of the year}
\usage{
get_dates_from_doy(year, doy)
}
\arguments{
\item{year}{A vector with the years.}

\item{doy}{A vector with the day of the year.
It must have the same length as \code{year}.}
}
\value{
A \code{\link[base]{Dates}} object.
}
\description{
This function retrieves the date corresponding to the given
year and day of the year.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
