% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{get_time_series_freq}
\alias{get_time_series_freq}
\title{Compute the Most Common Sampling Frequency across all observations}
\usage{
get_time_series_freq(x)
}
\arguments{
\item{x}{A data frame including a column called `observations`` with the time series}
}
\value{
A difftime object representing the most common time difference between consecutive samples.
}
\description{
This function calculates the most common difference between consecutive time points.
This can be useful for determining the aproximate sampling frequency of the time series data.
}
