% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-getAlignments.R
\name{getAlignments-method}
\alias{getAlignments}
\alias{getAlignments-method}
\title{Get TWDTW alignments}
\usage{
getAlignments(object, ...)
}
\arguments{
\item{object}{an object of class twdtwTimeSeries.}

\item{...}{other arguments to be passed to methods}
}
\description{
Get TWDTW alignments from objects of class twdtwMatches
}
\examples{
matches = new("twdtwMatches")
getAlignments(matches)

}
\seealso{
\code{\link[dtwSat]{twdtwMatches-class}}, 
\code{\link[dtwSat]{twdtwTimeSeries-class}}, and
\code{\link[dtwSat]{twdtwRaster-class}}
}

