% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\docType{methods}
\name{getInternals}
\alias{getInternals}
\title{Get internals from dtwSat object}
\usage{
getInternals(object)
}
\arguments{
\item{object}{A \link[dtwSat]{dtwSat-class} object}
}
\value{
An object of class \code{\link[base]{list}} whose 
  elements have the internal structures used by \code{\link[dtwSat]{twdtw}}. 
  The elements are: 
      \cr\code{timeWeight}: time weight matrix,
      \cr\code{localMatrix}: local cost matrix,
      \cr\code{costMatrix}: cumulative cost matrix,
      \cr\code{directionMatrix}: directions of steps that would be taken in the alignments,
      \cr\code{stepPattern}: \code{\link[dtw]{stepPattern}} used for the 
      computation, see package \code{\link[dtw]{dtw}}
      \cr\code{query}: query time series, 
      \cr\code{timeseries}: satellite image time series,
      \cr\code{N}: \code{query} length, and 
      \cr\code{M}: \code{timeseries} length.
}
\description{
This function retrieves cost matrix, inputs, and other 
internal structures from dtwSat-class object
}
\examples{
names(query.list)
alig = dtwSat(query.list[["Soybean"]], timeseries=template, keep=TRUE)
getInternals(alig)
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{dtwSat}}, \code{\link[dtwSat]{twdtw}}
}

