% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\docType{methods}
\name{getMatches}
\alias{getMatches}
\title{Get matching points from dtwSat object}
\usage{
getMatches(object)
}
\arguments{
\item{object}{A \link[dtwSat]{dtwSat} object}
}
\value{
object of class \code{list}
}
\description{
This function retrieves the matching points
for each alignment between the \code{query} and the \code{template}.
}
\examples{
names(query.list)
alig = dtwSat(query.list[["Soybean"]], template, weight = "logistic",
       alpha = 0.1, beta = 50, alignments=4)
getMatches(alig)
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{dtwSat}}, \code{\link[dtwSat]{twdtw}}
}

