% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModisTimeIndex.R
\docType{methods}
\name{getModisTimeIndex}
\alias{getModisTimeIndex}
\title{Get Modis time index from date}
\usage{
getModisTimeIndex(date, frequency = 16)
}
\arguments{
\item{date}{A \code{\link[base]{Dates}} object}

\item{frequency}{An integer with the frequency in days. Default is 16 days}
}
\value{
An \code{\link[base]{integer}} object.
}
\description{
This function retrieves the nearest time 
index to a date
}
\examples{
i = getModisTimeIndex(date=as.Date("2000-01-01"))
i

}
\seealso{
\link[dtwSat]{getModisTimeSequence}, and
\link[dtwSat]{getDatesFromDOY}.
}

