% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModisTimeSequence.R
\docType{methods}
\name{getModisTimeSequence}
\alias{getModisTimeSequence}
\title{Create time sequence}
\usage{
getModisTimeSequence(year = 2000:format(Sys.time(), "\%Y"), frequency = 16)
}
\arguments{
\item{year}{A vector with the years. Default 
is form 2000 to the system time year \code{format(Sys.time(), ''\%Y'')}}

\item{frequency}{An integer with the frequency in days. Default is 16 days}
}
\value{
A vector of \code{\link[base]{Dates}}
}
\description{
This function creates a sequence of dates for 
each year. The sequences start on January 1st of each year.
}
\examples{
dates = getModisTimeSequence()
dates

}
\seealso{
\link[dtwSat]{getDatesFromDOY} and \link[dtwSat]{getModisTimeIndex}
}

