% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\docType{methods}
\name{getPatternNames}
\alias{getPatternNames}
\title{Get pattern names from dtwSat object}
\usage{
getPatternNames(object, p.names)
}
\arguments{
\item{object}{A \link[dtwSat]{dtwSat-class} object}

\item{p.names}{A \link[base]{character} or \link[base]{numeric}
vector with the patterns identification. If not declared the function 
retrieves the names for all patterns}
}
\value{
A \code{\link[base]{character}}
or \code{\link[base]{numeric}} vector
}
\description{
This function retrieves pattern names in the 
\link[dtwSat]{dtwSat-class} object
}
\examples{

alig = twdtw(patterns=patterns.list, timeseries=template)

getPatternNames(alig)

getPatternNames(alig, p.names=c(1,3))

getPatternNames(alig, p.names="Maize")

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{dtwSat-class}}, and
\code{\link[dtwSat]{twdtw}}
}

