% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-getTimeSeries.R
\name{getPatterns-method}
\alias{getPatterns}
\alias{getPatterns-method}
\title{Get time series}
\usage{
getPatterns(object, ...)
}
\arguments{
\item{object}{an twdtw* object.}

\item{...}{other arguments to be passed to methods.}
}
\description{
Generic method to get temporal patterns from objects of class twdtw*.
}
\seealso{
\code{\link[dtwSat]{twdtwMatches-class}}, 
\code{\link[dtwSat]{twdtwTimeSeries-class}}, and
\code{\link[dtwSat]{twdtwRaster-class}}
}

