% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPatterns.R
\docType{methods}
\name{getPatterns}
\alias{getPatterns}
\title{Get patterns from twdtw object}
\usage{
getPatterns(object, p.names)
}
\arguments{
\item{object}{A \link[dtwSat]{twdtw-class} object.}

\item{p.names}{A \link[base]{character} or \link[base]{numeric}
vector with the patterns identification. If not declared the function 
retrieves all patterns.}
}
\value{
A a list of \link[zoo]{zoo} objects
}
\description{
This function retrieves temporal patterns used in the
TWDTW analysis.
}
\examples{

log_fun = logisticWeight(alpha=-0.1, beta=100)
matches = twdtw(x=example_ts, patterns=patterns.list, 
             weight.fun = log_fun)

a = getPatterns(matches)
names(a) 

a = getPatterns(matches, p.names="Maize")
names(a) 

a = getPatterns(matches, p.names=c(1,2))
names(a) 


}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, and
\code{\link[dtwSat]{twdtw}}.
}

