% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeSeries.R
\name{getTimeSeries}
\alias{getTimeSeries}
\title{Get time series}
\usage{
getTimeSeries(object, labels = NULL)
}
\arguments{
\item{object}{an twdtw* object.}

\item{labels}{a vector with the time series labels. If not informed the 
function retrieves all time series.}
}
\description{
Generic method to get time series from objects of class twdtw*.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwMatches-class}}, 
\code{\link[dtwSat]{twdtwTimeSeries-class}}, and
\code{\link[dtwSat]{twdtwRaster-class}}
}

