% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\docType{methods}
\name{initialize,dtwSat-method}
\alias{initialize,dtwSat-method}
\title{dtwSat-initialize}
\usage{
\S4method{initialize}{dtwSat}(.Object, call, internals, alignments, mapping)
}
\arguments{
\item{.Object}{A character "dtwSat"}

\item{call}{Object of class \code{call} see \code{\link[base]{match.call}}}

\item{internals}{Object of class \code{list}}

\item{alignments}{Object of class \code{list}}

\item{mapping}{Object of class \code{list}}
}
\description{
Initialize object of class dtwSat
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

