% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{iqueryApply}
\alias{iqueryApply}
\title{iquery builder}
\usage{
iqueryApply(array, expr)
}
\arguments{
\item{array}{A character for the array name.}

\item{expr}{An R expression for processing. 
The expression must retrieve a list of 8 elements of type double.}
}
\description{
This function builds the r_exec call.
}
\examples{

in_iquery = iquerySubsetArray(array = "MOD13Q1", attributes=c("evi", "ndvi", "cdoy"), 
                               min.col = 60124, max.col = 60125, 
                               min.row = 48978, max.row = 48979,
                               min.time = 0, max.time = 500)

proc_iquery = iqueryApply(array=in_iquery, expr="as.list(as.double(1:8))")

proc_iquery

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{modisColRowFromLongLat}
}

