% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{iqueryBlock}
\alias{iqueryBlock}
\title{iquery builder}
\usage{
iqueryBlock(array.in, array.out = NULL, attributes, expr, outputschema,
  min.col, max.col, block.col, min.row, max.row, block.row, min.time = 0,
  max.time = 99, index.start, index.end, chunk, overlap = c(0, 0, 0))
}
\arguments{
\item{array.in}{A character for the input array name.}

\item{array.out}{A character for the output array name. If not declared the 
iquery will not same the final results.}

\item{attributes}{A character vector for the array attributes.}

\item{expr}{An R expression for processing. The expression must retrieve a 
list of 8 elements of type double.}

\item{outputschema}{the SciDB schema of the output array.}

\item{min.col}{minimum column index.}

\item{max.col}{maximum column index.}

\item{block.col}{col block size.}

\item{min.row}{minimum row index.}

\item{max.row}{maximun row index.}

\item{block.row}{row block size.}

\item{min.time}{minimum time index for subset.}

\item{max.time}{maximum time index for subset.}

\item{index.start}{the starting indices for each dimensions c(<col>, <row>, <time>).}

\item{index.end}{the ending indices for each dimensions c(<col>, <row>, <time>).}

\item{chunk}{the chunk size for each dimensions c(<col>, <row>, <time>).}

\item{overlap}{the overlap for each dimensions c(<col>, <row>, <time>).}
}
\description{
This function builds an afl query to redimension 
a subset of the input array for the TWDTW analysis.
}
\examples{

out_schema = "<ts_id:double,year:double,cdoy:double,distance:double> 
[col_id=59354:60132,256,1,row_id=48591:49096,256,1,time_id=0:99,100,0,class_id=0:255,256,0]"

str_iquery = iqueryBlock(array.in = "MOD13Q1", attributes=c("evi", "ndvi", "cdoy"), 
                        outputschema = out_schema, 
                        expr="as.list(as.double(c(60124, 48978,1,1,1:4)))", 
                        min.col = 60124, max.col = 60125, block.col = 256, 
                        min.row = 48978, max.row = 48979, block.row = 256,
                        min.time = 0, max.time = 99, index.start=c(48000,38400,0), 
                        index.end = c(72000, 62400, 14999), chunk=c(32,32,99))

str_iquery

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{twdtwApplyiquery}
}

