% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{iqueryBlockSize}
\alias{iqueryBlockSize}
\title{iquery builder}
\usage{
iqueryBlockSize(min.col, max.col, block.col, min.row, max.row, block.row)
}
\arguments{
\item{min.col}{minimum column index.}

\item{max.col}{maximum column index.}

\item{block.col}{col block size.}

\item{min.row}{minimum row index.}

\item{max.row}{maximun row index.}

\item{block.row}{row block size.}
}
\description{
This function splits the SciDB array in block for procesing.
}
\examples{

bolcks = iqueryBlockSize(min.col = 59354, max.col = 60132, block.col = 256,
                         min.row = 48591, max.row = 49096, block.row = 256)
                               
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{iqueryCreateArray}
}

