% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{iqueryCreateArray}
\alias{iqueryCreateArray}
\title{iquery builder}
\usage{
iqueryCreateArray(array, min.col, max.col, chunk.col = 256, overlap.col = 1,
  min.row, max.row, chunk.row = 256, overlap.row = 1, min.time = 0,
  max.time = 99, chunk.time = max.time + 1, overlap.time = 0,
  min.class = 0, max.class = 255, chunk.class = max.class + 1,
  overlap.class = 0)
}
\arguments{
\item{array}{A character for the array name.}

\item{min.col}{minimum column index.}

\item{max.col}{maximum column index.}

\item{chunk.col}{column chunk size.}

\item{overlap.col}{column overlap.}

\item{min.row}{minimum row index.}

\item{max.row}{maximun row index.}

\item{chunk.row}{row chunk size.}

\item{overlap.row}{row overlap.}

\item{min.time}{minimum time index.}

\item{max.time}{maximum time index.}

\item{chunk.time}{time chunk size.}

\item{overlap.time}{time overlap.}

\item{min.class}{minimum class index.}

\item{max.class}{maximum class index.}

\item{chunk.class}{class chunk size.}

\item{overlap.class}{class overlap.}
}
\description{
This function builds an aql query to create 
the output array for the TWDTW analysis.
}
\examples{

str_iquery = iqueryCreateArray(array = "TEST_ARRAY",
                               min.col = 59354, max.col = 60132, 
                               min.row = 48591, max.row = 49096)

str_iquery

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{modisColRowFromLongLat}
}

