% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iquery.R
\name{iqueryCreateTWDTWArray}
\alias{iqueryCreateTWDTWArray}
\title{iquery builder}
\usage{
iqueryCreateTWDTWArray(array, index.start, index.end, chunk, overlap)
}
\arguments{
\item{array}{A character for the array name.}

\item{index.start}{the starting indices for each dimensions c(<col>, <row>, <time>, <class>).}

\item{index.end}{the ending indices for each dimensions c(<col>, <row>, <time>, <class>).}

\item{chunk}{the chunk size for each dimensions c(<col>, <row>, <time>, <class>).}

\item{overlap}{the overlap for each dimensions c(<col>, <row>, <time>, <class>).}
}
\description{
This function builds an aql query to create 
the output array for the TWDTW analysis.
}
\examples{


}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{modisColRowFromLongLat}
}

