% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{iqueryRankTWDTWDistance}
\alias{iqueryRankTWDTWDistance}
\title{iquery builder}
\usage{
iqueryRankTWDTWDistance(array, min.col, max.col, min.row, max.row, min.time,
  max.time, min.class, max.class, index.start, index.end, chunk, overlap)
}
\arguments{
\item{array}{A character for the array name.}

\item{min.col}{minimum column index.}

\item{max.col}{maximum column index.}

\item{min.row}{minimum row index.}

\item{max.row}{maximun row index.}

\item{min.time}{minimum time index.}

\item{max.time}{maximum time index.}

\item{min.class}{minimum class index.}

\item{max.class}{maximum class index.}

\item{index.start}{the starting indices for each dimensions c(<col>, <row>, <time>, <class>).}

\item{index.end}{the ending indices for each dimensions c(<col>, <row>, <time>, <class>).}

\item{chunk}{the chunk size for each dimensions c(<col>, <row>, <time>, <class>).}

\item{overlap}{the overlap for each dimensions c(<col>, <row>, <time>, <class>).}
}
\description{
This function builds an afl query that creates 
an output array with the rank of TWDTW distances.
}
\examples{

str_iquery = iqueryCreateArray(array = "TEST_ARRAY",
                               min.col = 59354, max.col = 60132, 
                               min.row = 48591, max.row = 49096)

str_iquery

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{modisColRowFromLongLat}
}

