% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{iqueryRedimensionOutput}
\alias{iqueryRedimensionOutput}
\title{iquery builder}
\usage{
iqueryRedimensionOutput(array, schema)
}
\arguments{
\item{array}{A character for the array name.}

\item{schema}{A character with the output array schema. Usually the schema of
the array created by \link[dtwSat]{iqueryCreateArray}}
}
\description{
This function builds the r_exec call.
}
\examples{

in_iquery = iquerySubsetArray(array = "MOD13Q1", attributes=c("evi", "ndvi", "cdoy"), 
                               min.col = 60124, max.col = 60125, 
                               min.row = 48978, max.row = 48979,
                               min.time = 0, max.time = 500)

proc_iquery = iqueryApply(array = in_iquery, expr="as.list(as.double(c(60124,48978,1,1,1:4)))")

out_schema = "<ts_id:double,year:double,cdoy:double,distance:double> [col_id=59354:60132,254,1,row_id=48591:49096,254,1,time_id=0:100,100,0,class_id=0:255,255,0]"
out_array = iqueryRedimensionOutput(array = proc_iquery, schema = out_schema)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{iqueryCreateArray}
}

