% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iquery.R
\name{iquerySellectMatches}
\alias{iquerySellectMatches}
\title{iquery builder}
\usage{
iquerySellectMatches(array.in, min, max, k = 1, chunk, overlap = c(0, 0, 0,
  0))
}
\arguments{
\item{chunk}{the chunk size for each dimensions c(<col>, <row>, <time>, <class>).}

\item{overlap}{the overlap for each dimensions c(<col>, <row>, <time>, <class>).}
}
\description{
This function builds an afl query to redimension 
a subset of the input array for the TWDTW analysis.
}
\examples{
###

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{twdtwApplyiquery}
}

