% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{iquerySubsetArray}
\alias{iquerySubsetArray}
\title{iquery builder}
\usage{
iquerySubsetArray(array, attributes, min.col, max.col, min.row, max.row,
  min.time, max.time)
}
\arguments{
\item{array}{A character for the array name.}

\item{attributes}{A character vector for the array attributes.}

\item{min.col}{minimum column index.}

\item{max.col}{maximum column index.}

\item{min.row}{minimum row index.}

\item{max.row}{maximun row index.}

\item{min.time}{minimum time index.}

\item{max.time}{maximum time index.}
}
\description{
This function builds an afl query to subset 
an SciDB array.
}
\examples{

str_iquery = iquerySubsetArray(array = "MOD13Q1", attributes=c("evi", "ndvi", "cdoy"), 
                               min.col = 60124, max.col = 60125, 
                               min.row = 48978, max.row = 48979,
                               min.time = 0, max.time = 500)

str_iquery

str_iquery = iquerySubsetArray(array = "MOD13Q1", attributes=c("evi", "ndvi", "cdoy"), 
                               min.col = 55324, max.col = 55325, 
                               min.row = 48978, max.row = 48979,
                               min.time = 0, max.time = 500)

str_iquery

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{modisColRowFromLongLat}
}

