% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{iquerySubsetRedimensionInput}
\alias{iquerySubsetRedimensionInput}
\title{iquery builder}
\usage{
iquerySubsetRedimensionInput(array, attributes, min.col, max.col, min.row,
  max.row, min.time = 0, max.time = 99, index.start, index.end, chunk,
  overlap = c(0, 0, 0))
}
\arguments{
\item{array}{A character for the array name.}

\item{attributes}{A character vector for the array attributes.}

\item{min.col}{minimum column index for subset.}

\item{max.col}{maximum column index for subset.}

\item{min.row}{minimum row index for subset.}

\item{max.row}{maximun row index for subset.}

\item{min.time}{minimum time index for subset.}

\item{max.time}{maximum time index for subset.}

\item{index.start}{the starting indices for each dimensions c(<col>, <row>, <time>).}

\item{index.end}{the ending indices for each dimensions c(<col>, <row>, <time>).}

\item{chunk}{the chunk size for each dimensions c(<col>, <row>, <time>).}

\item{overlap}{the overlap for each dimensions c(<col>, <row>, <time>).}
}
\description{
This function builds an afl query to redimension 
a subset of the input array for the TWDTW analysis.
}
\examples{

str_iquery = iquerySubsetRedimensionInput(array = "MOD13Q1", attributes=c("evi", "ndvi", "cdoy"), 
                               min.col = 60124, max.col = 60125, min.row = 48978, max.row = 48979,
                               min.time = 0, max.time = 99, index.start=c(48000,38400,0), 
                               index.end = c(72000, 62400, 14999), chunk=c(32,32,99))

str_iquery

str_iquery = iquerySubsetRedimensionInput(array = "MOD13Q1", attributes=c("evi", "ndvi", "cdoy"), 
                               min.col = 55324, max.col = 55325, min.row = 48978, max.row = 48979,
                               min.time = 0, max.time = 99, index.start=c(48000,38400,0), 
                               index.end = c(72000, 62400, 14999), chunk=c(32,32,99))

str_iquery

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{twdtwApplyiquery}
}

