% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iquery.R
\name{iqueryTWDTWBlock}
\alias{iqueryTWDTWBlock}
\title{iquery builder}
\usage{
iqueryTWDTWBlock(array.in, array.out, attributes, expr, grid = NULL,
  pixelsize = NULL, blocksize, min, max, chunk, overlap = c(0, 0, 0))
}
\arguments{
\item{array.in}{A character for the input array name.}

\item{array.out}{A character for the output array name. If not declared the 
iquery will not same the final results.}

\item{attributes}{A character vector for the array attributes.}

\item{expr}{An R expression for processing. The expression must retrieve a 
list of 8 elements of type double.}

\item{grid}{An object of class \link[sp]{SpatialPolygons}. Each polygon is 
one processing block.}

\item{pixelsize}{A number. The MODIS product resolution.}

\item{blocksize}{processing block size c(<col>, <row>). Used instead 
of \code{grid} and \code{pixelsize}.}

\item{min}{minimum indices for subset c(<col>, <row>, <time>).}

\item{max}{maximum indices for subset c(<col>, <row>, <time>).}

\item{chunk}{scidb chunk size for input c(<col>, <row>, <time>).}

\item{overlap}{scidb overlap for for input c(<col>, <row>, <time>).}
}
\description{
This function builds an afl query to redimension 
a subset of the input array for the TWDTW analysis.
}
\examples{

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{twdtwApplyiquery}
}

