% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtwSat.R
\docType{methods}
\name{is.twdtw}
\alias{is.twdtw}
\title{Method for Time-Weighted DTW}
\usage{
is.twdtw(x)
}
\arguments{
\item{x}{an R object. See \code{\link[dtwSat]{twdtw}}}
}
\description{
Methods for Time-Weighted DTW objects
}
\examples{
alig = twdtw(query.list[["Soybean"]], template, weight = "logistic", alpha = 0.1, beta = 50)
is.twdtw(alig)
}

