% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinAlignments.R
\docType{methods}
\name{joinAlignments}
\alias{joinAlignments}
\alias{joinAlignments,twdtwMatches-method}
\alias{joinAlignments-twdtwMatches}
\title{Joint TWDTW alignments}
\usage{
joinAlignments(...)

\S4method{joinAlignments}{twdtwMatches}(...)
}
\arguments{
\item{...}{objects of class twdtwMatches}
}
\value{
an object of class \code{\link[dtwSat]{twdtwMatches}}
}
\description{
Join TWDTW alignments from two or more objects of class twdtwMatches.
}
\examples{
# Join TWDTW align from objects of class twdtwTimeSeries
patterns = twdtwTimeSeries(timeseries=patterns.list, labels=names(patterns.list))
ts1 = twdtwTimeSeries(timeseries=example_ts.list[1], labels="A")
ts2 = twdtwTimeSeries(timeseries=example_ts.list[2], labels="B")
mat1 = twdtwApply(x=ts1, y=patterns)
mat2 = twdtwApply(x=ts2, y=patterns)
matches = joinAlignments(mat1, mat2)
matches

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwMatches-class}}, and 
\code{\link[dtwSat]{joinTimeSeries}}
}

