% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinPatterns.R
\name{joinPatterns}
\alias{joinPatterns}
\title{Join patterns}
\usage{
joinPatterns(...)
}
\arguments{
\item{...}{objects of class twdtwTimeSeries}
}
\description{
Join patterns (time series) from two or more objects of class twdtwTimeSeries. 
This function removes patterns (time series) with duplicated label.
}
\examples{
# Joining temporal patterns 


}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwMatches-class}}, 
\code{\link[dtwSat]{twdtwTimeSeries-class}}
}

