% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinTimeSeries.R
\name{joinTimeSeries}
\alias{joinTimeSeries}
\title{Join time series}
\usage{
joinTimeSeries(...)
}
\arguments{
\item{...}{objects of class twdtwTimeSeries}
}
\description{
Join time series from two or more objects of class twdtwTimeSeries
}
\examples{
## 

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwMatches-class}}, 
\code{\link[dtwSat]{twdtwTimeSeries-class}}, and
\code{\link[dtwSat]{twdtwRaster-class}}
}

