% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{knn1_twdtw}
\alias{knn1_twdtw}
\title{Train a KNN-1 TWDTW model with optional GAM resampling}
\usage{
knn1_twdtw(
  x,
  y,
  formula = NULL,
  start_column = "start_date",
  end_column = "end_date",
  label_colum = "label",
  sampling_freq = NULL,
  ...
)
}
\arguments{
\item{x}{A two-dimensional stars object (x, y) with time and bands as attributes.}

\item{y}{An sf object with the coordinates of the training points.}

\item{formula}{Either NULL or a formula to reduce samples of the same label using Generalized Additive Models (GAM).
Default is \code{band ~ s(time)}. See details.}

\item{start_column}{Name of the column in y that indicates the start date. Default is 'start_date'.}

\item{end_column}{Name of the column in y that indicates the end date. Default is 'end_date'.}

\item{label_colum}{Name of the column in y containing land use labels. Default is 'label'.}

\item{sampling_freq}{The time frequency for sampling, including the unit (e.g., '16 day').
If NULL, the function will infer the frequency. This parameter is only used if a formula is provided.}

\item{...}{Additional arguments passed to the GAM function.}
}
\value{
A 'knn1_twdtw' model containing the trained model information and the data used.
}
\description{
This function prepares a KNN-1 model with the Time Warp Dynamic Time Warping (TWDTW) algorithm.
If a formula is provided, the training samples are resampled using Generalized Additive Models (GAM).
}
\details{
If \code{formula} is NULL, the KNN-1 model will retain all training samples. If a formula is passed (e.g., \code{band ~ \link[mgcv]{s}(time)}),
then samples of the same label (land cover class) will be resampled using GAM.
Resampling can significantly reduce prediction processing time.
}
\examples{

# Example will be added once the function is properly implemented and tested.

}
