% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtwSat.R
\docType{methods}
\name{kthbacktrack}
\alias{kthbacktrack}
\title{Computes the backtrack of dtw}
\usage{
kthbacktrack(alignment, jmin = NULL)
}
\arguments{
\item{alignment}{A dtw alignment object.}

\item{jmin}{An integer. index of the last line of the global
cost matrix.}
}
\description{
This function computs the backtrack starting from
a given index of the last line of the global cost matrix.
}

