% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twdtw-length.R
\docType{methods}
\name{length}
\alias{length}
\alias{length.twdtw}
\title{Length method for twdtw-class}
\usage{
\method{length}{twdtw}(x)
}
\arguments{
\item{x}{An \code{\link[dtwSat]{twdtw-class}} object.}
}
\value{
An \link[base]{integer}.
}
\description{
Retrieves the number of matches in the 
twdtw-class object.
}
\examples{

matches = twdtw(x=example_ts, patterns=patterns.list)
       
length(matches)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, and 
\code{\link[dtwSat]{twdtw}}
}

