% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearWeight.R
\docType{methods}
\name{linearWeight}
\alias{linearWeight}
\title{Linear weight function}
\usage{
linearWeight(a, b = 0)
}
\arguments{
\item{a}{numeric. The slop of the line}

\item{b}{numeric. The intercept of the line}
}
\value{
An \code{\link[base]{function}} object
}
\description{
Builds a linear time weight 
function to compute the TWDTW local cost matrix [1]
}
\examples{
weight.fun = linearWeight(a=0.1)
weight.fun

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\references{
[1] Maus  V,  C\^{a}mara  G,  Cartaxo  R,  Sanchez  A,  Ramos  FM,  de Queiroz, GR.
(2016). A Time-Weighted Dynamic Time Warping method for land use and land cover 
mapping. Selected Topics in Applied Earth Observations and Remote Sensing, 
IEEE Journal of, X, XX-XX.
}
\seealso{
\code{\link[dtwSat]{twdtw}}
}

