% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{longLatTomodisColRow}
\alias{longLatTomodisColRow}
\title{longitude and latitude to MODIS column and row.}
\usage{
longLatTomodisColRow(x, y = NULL, pixelsize, projCRS = NULL)
}
\arguments{
\item{x}{An SpatialPoints object or the longitude.}

\item{pixelsize}{A number. The MODIS resolution.}

\item{projCRS}{A CRS object.}
}
\description{
This function computes the MODIS grid
attributes from a pair of coordinates.
}

