% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOD13Q1.MT.yearly.patterns}
\alias{MOD13Q1.MT.yearly.patterns}
\title{Data: patterns time series}
\format{A \link[dtwSat]{twdtwTimeSeries} object.}
\usage{
MOD13Q1.MT.yearly.patterns
}
\description{
This dataset has a list of patterns with the phenological cycle of: Water,
Cotton-Fallow, Forest, Low vegetation, Pasture, Soybean-Cotton, Soybean-Maize, Soybean-Millet, 
Soybean-Sunflower, and Wetland. These time series are based on the MODIS product 
MOD13Q1 250 m 16 days [1]. The patterns were build from ground truth samples of each 
crop using Generalized Additive Models (GAM), see \link[dtwSat]{createPatterns}.
}
\references{
[1] Friedl MA, Sulla-Menashe D, Tan B, Schneider A, Ramankutty N, Sibley A, Huang X. (2010).
MODIS Collection 5 global land cover: Algorithm refinements and characterization of new
datasets. Remote Sensing of Environment, 114(1), 168 182.
}
\seealso{
MOD13Q1 documentation: See 
\url{https://lpdaac.usgs.gov/dataset_discovery/modis/modis_products_table/mod13q1_v006}.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\keyword{datasets}
