% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOD13Q1.patterns.list}
\alias{MOD13Q1.patterns.list}
\title{Data: patterns time series}
\format{
A named \code{list} of three \link[zoo]{zoo} objects, ''Soybean'', ''Cotton'', 
and ''Maize'', whose indices are \code{\link[base]{Dates}} in the format ''yyyy-mm-dd''.
Each node has 6 attributes: ''ndvi'', ''evi'', ''red'', ''nir'', ''blue'', 
and ''mir''.
}
\usage{
MOD13Q1.patterns.list
}
\description{
This dataset has a list of patterns with the phenological cycle of: Soybean,
Cotton, and Maize. These time series are based on the MODIS product 
MOD13Q1 250 m 16 days \insertCite{Didan:2015}{dtwSat}. The patterns were built 
from ground truth samples of each 
crop using Generalized Additive Models (GAM), see \link[dtwSat]{createPatterns}.
}
\references{
\insertAllCited{}
  
  \insertRef{Maus:2019}{dtwSat}
  
  \insertRef{Maus:2016}{dtwSat}
}
\seealso{
\link[dtwSat]{MOD13Q1.ts},
\link[dtwSat]{MOD13Q1.ts.list}, and 
\link[dtwSat]{createPatterns}.

For details about MOD13Q1 see \insertCite{Didan:2015}{dtwSat}.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\keyword{datasets}
