% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOD13Q1.ts.labels}
\alias{MOD13Q1.ts.labels}
\title{Data: Labels of the satellite time series in MOD13Q1.ts}
\format{An object of class \link[base]{data.frame}, whose attributas are: 
the label of the crop class ''label'', the start of the crop period ''from'',
and the end of the crop period ''to''. The dates are in the format ''yyyy-mm-dd''.}
\usage{
MOD13Q1.ts.labels
}
\description{
This labels are based on field work.
}
\seealso{
\link[dtwSat]{MOD13Q1.ts}.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\keyword{datasets}
