% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iquery.R
\name{modisColRowFromLongLat}
\alias{modisColRowFromLongLat}
\title{Get MODIS col row from longitude and latitude}
\usage{
modisColRowFromLongLat(x, pixelsize, proj4string = CRS(as.character(NA)))
}
\arguments{
\item{x}{An object of class SpatialPoints or data.frame with 
longitude and latitude.}

\item{pixelsize}{A number. The MODIS resolution.}

\item{proj4string}{A CRS object. See \link[sp]{CRS} for details.}
}
\description{
This function retrieves the MODIS col row  
from pair of coordinates.
}
\examples{
# Example for MOD13Q1 spatial resolution 231.6564 m 
coord = data.frame(-55.93715, -12.05729)
prj_str = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
modisColRowFromLongLat(x=coord, pixelsize=231.6564, proj4string=prj_str)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{modisColRowToLongLat}
}

