% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{modisColRowToLongLat}
\alias{modisColRowToLongLat}
\title{Get longitude and latitude from MODIS col row from}
\usage{
modisColRowToLongLat(col, row, h = NULL, v = NULL, pixelsize,
  proj4string = NULL)
}
\arguments{
\item{col}{An integer. A global index of the MODIS column
or a local index of the column within a specific h:v tile.}

\item{row}{An integer. A global index of the MODIS row
or a local index of the row within a specific h:v tile.}

\item{h}{An integer. Horizontal MODIS tile.}

\item{v}{An integer. Vertical MODIS tile.}

\item{pixelsize}{A number. The MODIS resolution.}

\item{proj4string}{A CRS object. See \link[sp]{CRS} for details.}
}
\description{
This function retrieves the longitude and latitude from 
MODIS col row.
}
\examples{
# Example for MOD13Q1 spatial resolution 231.6564 m  
modisColRowToLongLat(col=60124, row=48978, pixelsize=231.6564)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{modisColRowFromLongLat}
}

