% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{modisrasterFromSciDBArray}
\alias{modisrasterFromSciDBArray}
\title{Raster from scidb array.}
\usage{
modisrasterFromSciDBArray(arrayname, years, mincol, maxcol, minrow, maxrow,
  pixelsize, hostname, port, user, password, filepath, overwrite = TRUE)
}
\arguments{
\item{arrayname}{}

\item{years}{}

\item{mincol}{}

\item{maxcol}{}

\item{minrow}{}

\item{maxrow}{}

\item{pixelsize}{}

\item{port}{}

\item{user}{}

\item{password}{}

\item{filepath}{}

\item{overwrite}{}

\item{host}{}
}
\description{
This function creates and project MODIS
arrays from from scidb database.
}

