% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{modisrasterFromSciDBArrayRGB}
\alias{modisrasterFromSciDBArrayRGB}
\title{Raster from scidb array.}
\usage{
modisrasterFromSciDBArrayRGB(arrayname, slices, mincol, maxcol, minrow, maxrow,
  pixelsize, hostname, port, user, password, filepath, RGBcode,
  overwrite = FALSE)
}
\arguments{
\item{arrayname}{A carachter. SciDB array name.}

\item{slices}{A numeric vector. The array index for each desired time slices.}

\item{mincol}{An integer. The global index for the minimum MODIS column.}

\item{maxcol}{An integer. The global index for the maximum MODIS column.}

\item{minrow}{An integer. The global index for the minimum MODIS row.}

\item{maxrow}{An integer. The global index for the maximum MODIS row.}

\item{pixelsize}{A number. The MODIS resolution in meters.}

\item{port}{An integer.}

\item{user}{A carachter.}

\item{password}{A carachter.}

\item{filepath}{A carachter.}

\item{RGBcode}{an R data.frame with the class number and R,G,B values in the columns.}

\item{overwrite}{Logical. If TRUE, raster file will be overwritten if it exists. Default is FALSE.}

\item{host}{A carachter.}
}
\description{
This function creates and project MODIS grid from slices of
a scidb array.
}

