% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nalignments.R
\docType{methods}
\name{nalignments}
\alias{nalignments}
\title{Get the number of alignments from twdtw object}
\usage{
nalignments(object, ...)
}
\arguments{
\item{object}{A \link[dtwSat]{twdtw-class} object}

\item{...}{additional arguments passed to \code{\link[dtwSat]{getPatternNames}}}
}
\value{
A \link[base]{numeric}
}
\description{
This function retrieves the number of alignments 
from the object \link[dtwSat]{twdtw-class}
}
\examples{

weight.fun = logisticWeight(alpha=-0.1, beta=100)
alig = twdtw(x=template, patterns=patterns.list, weight.fun = weight.fun)

nalignments(alig)

nalignments(alig, p.names="Soybean")

nalignments(alig, p.names=c(2,3))

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, and
\code{\link[dtwSat]{twdtw}}
}

