% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmatches.R
\docType{methods}
\name{nmatches}
\alias{nmatches}
\title{Get the number of matches from twdtw object}
\usage{
nmatches(object, p.names)
}
\arguments{
\item{object}{A \link[dtwSat]{twdtw-class} object.}

\item{p.names}{A \link[base]{character} or \link[base]{numeric}
vector with the patterns identification. If not declared the function 
retrieves the total number of matches considering all patterns.}
}
\value{
A \link[base]{numeric}
}
\description{
This function retrieves the number of matches  
from a \link[dtwSat]{twdtw-class} object.
}
\examples{

log_fun = logisticWeight(alpha=-0.1, beta=100)
matches = twdtw(x=example_ts, patterns=patterns.list, weight.fun = log_fun)

nmatches(matches)

nmatches(matches, p.names="Soybean")

nmatches(matches, p.names=c(2,3))

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, and
\code{\link[dtwSat]{twdtw}}
}

