% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizePatterns.R
\docType{methods}
\name{normalizePatterns}
\alias{normalizePatterns}
\title{Normalize patterns length}
\usage{
normalizePatterns(..., patterns = list(...), patterns.length = NULL)
}
\arguments{
\item{...}{\link[zoo]{zoo} objects}

\item{patterns}{a list of \link[zoo]{zoo} objects}

\item{patterns.length}{An integer. Default is the length of the longest pattern}
}
\value{
A \link[base]{list} of \link[zoo]{zoo} objects with normalized 
length
}
\description{
This function normalizes the length of a pattern or 
list of patterns
}
\examples{
new.patterns.list = normalizePatterns(patterns = patterns.list, patterns.length = 23)
sapply(patterns.list, nrow)
sapply(new.patterns.list, nrow)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

