% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{normalizeQuery}
\alias{normalizeQuery}
\title{Normalize query length}
\usage{
normalizeQuery(..., query = list(...), query.length = NULL)
}
\arguments{
\item{...}{\link[zoo]{zoo} objects}

\item{query}{a list of \link[zoo]{zoo} objects}

\item{query.length}{An integer. Default is the length of the longest query}
}
\value{
A \link[base]{list} of \link[zoo]{zoo} objects with normalized
length
}
\description{
This function normalize the length of a query or
list of queries
}
\examples{
new.query.list = normalizeQuery(query = query.list, query.length = 23)
lapply(query.list, nrow)
lapply(new.query.list, nrow)
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

