% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patterns.vignette}
\alias{patterns.vignette}
\title{Data: patterns time series}
\format{A named \code{list} of 5 \link[zoo]{zoo} objects, ''Cotton-fallow''
''Forest'', ''Soybean-cotton'', ''Soybean-maize'', and ''Soybean-millet'', 
whose indices are \code{\link[base]{Dates}} in the format ''yyyy-mm-dd''.
Each node has 6 attributes: ''blue'', ''evi'', ''mir'', ''ndvi'', ''nir'', 
and ''red''.}
\usage{
patterns.vignette
}
\description{
This dataset has a list of queries form the crops: ''Cotton-fallow''
''Forest'', ''Soybean-cotton'', ''Soybean-maize'', and ''Soybean-millet''. 
These time series are based on the MODIS product 
MOD13Q1 250 m 16 days. The queries were extracted from field 
samples of each class using Generalized Additive Models (GAM), see 
package \link[mgcv]{gam}.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
MOD13Q1 documentation: See 
\url{https://lpdaac.usgs.gov/dataset_discovery/modis/modis_products_table/mod13q1}
}
\keyword{datasets}

