% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-dtwSat.R
\docType{methods}
\name{plot,dtwSat-method}
\alias{plot,dtwSat-method}
\title{Plotting dtwSat objects}
\usage{
\S4method{plot}{dtwSat}(x, type = "path", ...)
}
\arguments{
\item{x}{\code{\link[dtwSat]{dtwSat-class}} object}

\item{type}{A character for the plot type, "path" or "alignment"}

\item{...}{additional arguments passed to plotting functions
\code{\link[dtwSat]{plotPath}} and \code{\link[dtwSat]{plotAlignment}}
Default is "path"}
}
\value{
object of class \code{\link[ggplot2]{ggplot}}
}
\description{
Methods for plotting the results of the
Time-Weighted DTW analsys
}
\examples{
names(query.list)
alig = dtwSat(query.list[["Soybean"]], template, weight = "logistic",
       alpha = 0.1, beta = 50, alignments = 4, keep = TRUE)

## Plot path
gp1 = plot(alig, type="path", normalize=TRUE, show.dist = TRUE)
gp1

## Plot alignment
gp2 = plot(alig, type="alignment", dimension="evi", alignment=1, shift=0.5)
gp2
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{dtwSat}}, \code{\link[dtwSat]{twdtw}},
\code{\link[dtwSat]{plotPath}}, \code{\link[dtwSat]{plotAlignment}}
}

