% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtwSat.R
\docType{methods}
\name{plot.twdtw}
\alias{plot.twdtw}
\title{Plotting Time-Weighted DTW}
\usage{
\method{plot}{twdtw}(x, ylab = "Pattern", xlab = "Time series", ...)
}
\arguments{
\item{...}{additional arguments passed to plotting functions}

\item{obj}{twdtw object}
}
\description{
Methods for plotting Time-Weighted DTW objects
returned by twdtw.
}

