% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patterns.R
\name{plot_patterns}
\alias{plot_patterns}
\title{Plot Patterns from Time Series Data}
\usage{
plot_patterns(x, ...)
}
\arguments{
\item{x}{A list where each element is a data.frame representing a time series.
Each data.frame should have the same number of rows and columns,
with columns representing different attributes (e.g., bands or indices)
and rows representing time points.
The name of each element in the list will be used as the facet title.}

\item{...}{Not used.}
}
\value{
A ggplot object displaying the time series patterns.
}
\description{
This function takes a list of time series data and creates a multi-faceted plot
where each facet corresponds to a different time series from the list.
Within each facet, different attributes (columns of the time series) are
plotted as lines with different colors.
}
