% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignments.R
\docType{methods}
\name{plotAlignments}
\alias{plotAlignments}
\title{Plotting TWDTW alignments}
\usage{
plotAlignments(x, p.names, attr = 1, threshold = Inf)
}
\arguments{
\item{x}{An \code{\link[dtwSat]{twdtw-class}} object.}

\item{p.names}{A \link[base]{character} or \link[base]{integer}
vector with the patterns identification. If not declared the function 
will plot the alignments for all patterna in \code{x}.}

\item{attr}{An \link[base]{integer} or \link[base]{character} vector 
indicating the attribute for plotting. Default is 1.}

\item{threshold}{A number. The TWDTW dissimilarity threshold, \emph{i.e.} the 
maximum TWDTW cost for consideration. Default is \code{Inf}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Method for plotting the alignments and respective TWDTW 
dissimilarity measures.
}
\examples{

log_fun = logisticWeight(alpha=-0.1, beta=100)
matches = twdtw(x=example_ts, patterns=patterns.list, weight.fun = log_fun, 
        normalize.patterns=TRUE, patterns.length=23)

gp = plotAlignments(matches, attr=c("ndvi","evi"))
gp

gp = plotAlignments(matches, attr=c("ndvi","evi"), threshold=4)
gp

gp = plotAlignments(x=matches, p.names="Cotton", 
                   attr=c("ndvi","evi"), threshold=6)
gp

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, 
\code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{plotPaths}}, 
\code{\link[dtwSat]{plotCostMatrix}},
\code{\link[dtwSat]{plotMatches}}, 
\code{\link[dtwSat]{plotClassification}}, and 
\code{\link[dtwSat]{plotPatterns}}.
}

