% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClassification.R
\name{plotClassification}
\alias{plotClassification}
\title{Plotting subintervals classification}
\usage{
plotClassification(x, attr, ...)
}
\arguments{
\item{x}{A \code{\link[dtwSat]{twdtw-class}} object.}

\item{attr}{An \link[base]{integer} vector or \link[base]{character} vector 
indicating the attribute for plotting. If not declared the function will plot 
all attributes.}

\item{...}{additional arguments passed to \code{\link[dtwSat]{classifyIntervals}}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Method for plotting the classification of each 
subinterval of the time series based on TWDTW analysis.
}
\examples{
log_fun = logisticWeight(alpha=-0.1, beta=100)
matches = twdtw(x=example_ts, patterns=patterns.list, weight.fun = log_fun, 
        normalize.patterns=TRUE, patterns.length=23)

# Classify interval
from = as.Date("2009-09-01")
to = as.Date("2013-09-01")
by = "6 month"

# All classes
gp = plotClassification(x=matches, from=from, to=to, by=by, overlap=.4)
gp


}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, 
\code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{plotPaths}}, 
\code{\link[dtwSat]{plotCostMatrix}},
\code{\link[dtwSat]{plotAlignments}}, 
\code{\link[dtwSat]{plotMatches}}, and 
\code{\link[dtwSat]{plotPatterns}}.
}

