% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCrossValidation.R
\name{plotCrossValidation}
\alias{plotCrossValidation}
\title{Plotting cross-validation}
\usage{
plotCrossValidation(x)
}
\arguments{
\item{x}{An object of class \code{\link[dtwSat]{plotCrossValidation}}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Method for plotting cross-validation results.
}
\examples{
\dontrun{
# Data folder 
data_folder = system.file("lucc_MT/data", package = "dtwSat")

# Read dates 
dates = scan(paste(data_folder,"timeline", sep = "/"), what = "dates")

# Read raster time series 
evi = brick(paste(data_folder,"evi.tif", sep = "/"))
raster_timeseries = twdtwRaster(evi, timeline = dates)

# Read field samples 
field_samples = read.csv(paste(data_folder,"samples.csv", sep = "/")) 
table(field_samples[["label"]])

# Read field samples projection 
proj_str = scan(paste(data_folder,"samples_projection", sep = "/"), 
     what = "character")

# Get sample time series from raster time series 
field_samples_ts = getTimeSeries(raster_timeseries, 
     y = field_samples, proj4string = proj_str)
field_samples_ts

# Run cross validation
set.seed(1)
# Define TWDTW weight function 
log_fun = logisticWeight(alpha=-0.1, beta=50) 
cross_validation = twdtwCrossValidation(field_samples_ts, times=3, p=0.1, 
                          freq = 8, formula = y ~ s(x, bs="cc"), weight.fun = log_fun)

plot(cross_validation)

}

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwCrossValidation}}
}

