% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistance.R
\name{plotDistance}
\alias{plotDistance}
\title{Plotting distance maps}
\usage{
plotDistance(x, time.levels = 1, time.labels = 1, layers = NULL)
}
\arguments{
\item{x}{An object of class \code{\link[dtwSat]{twdtwRaster}}.}

\item{time.levels}{A \link[base]{character} or \link[base]{numeric}
vector with the layers to plot. For plot type ''change'' the minimum length 
is two.}

\item{time.labels}{A \link[base]{character} or \link[base]{numeric}
vector with the labels of the layers. It must have the same 
length as time.levels. Default is NULL.}

\item{layers}{A \link[base]{character} or \link[base]{numeric}
vector with the layers/bands of the raster time series.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Method for plotting TWDTW distance maps.
}
\examples{
\dontrun{

}
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwRaster-class}},
\code{\link[dtwSat]{twdtwApply}}, 
\code{\link[dtwSat]{plotArea}}, 
\code{\link[dtwSat]{plotChanges}}, and
\code{\link[dtwSat]{plotDistance}}.
}

