% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dtwSat.R
\name{plotGroup}
\alias{plotGroup}
\title{Plotting time interval groups}
\usage{
plotGroup(x, attr, ...)
}
\arguments{
\item{x}{A \code{\link[dtwSat]{dtwSat-class}} object}

\item{attr}{An \link[base]{integer} vector or \link[base]{character} vector 
indicating the attribute for plotting, \emph{i.e.} a column of the \code{timeseries}. 
If not declared the function will plot all attributes}

\item{...}{additional arguments passed to \code{\link[dtwSat]{classifyIntervals}}}
}
\value{
A \link[ggplot2]{ggplot} object
}
\description{
Method for plotting the group of each 
time intervals based on \code{\link[dtwSat]{twdtw}} analysis
}
\examples{
weight.fun = logisticWeight(alpha=-0.1, beta=100)
alig = twdtw(patterns=patterns.list, timeseries=template, weight.fun = weight.fun, 
        normalize.patterns=TRUE, patterns.length=23, keep=TRUE)

# Classify interval
from = as.Date("2009-09-01")
to = as.Date("2013-09-01")
by = "6 month"

# All classes
gp = plotGroup(x=alig, from=from, to=to, by=by, overlap=.3)
gp

# Cotton and Maize 
gp = plotGroup(x=alig, attr=c("ndvi","evi"), from=from, to=to, by=by, 
             overlap=.3, p.names=c("Cotton","Maize"))
gp


}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{dtwSat-class}}, 
\code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{plotPath}}, 
\code{\link[dtwSat]{plotCostMatrix}},
\code{\link[dtwSat]{plotAlignment}},
\code{\link[dtwSat]{plotMatch}}, and
\code{\link[dtwSat]{classifyIntervals}}
}

