% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLUCC.R
\docType{methods}
\name{plotLUCC}
\alias{plotLUCC}
\title{Plotting land use/cover changes (LUCC)}
\usage{
plotLUCC(x, type = "area", layer.levels, layer.labels = NULL,
  levels = NULL, labels = NULL, colors = NULL)
}
\arguments{
\item{x}{A \code{\link[raster]{Raster-class}}
\code{\link[raster]{brick}} or \code{\link[raster]{stack}} objects. 
A \link[base]{character} with the raster file name is also accepted.}

\item{type}{A \link[base]{character}. The plot type, ''map'', 
''area'', or ''change''. Default is ''map''}

\item{layer.levels}{A \link[base]{character} or \link[base]{numeric}
vector with the layers to compute the changes. The minimum length 
is two}

\item{layer.labels}{A \link[base]{character} or \link[base]{numeric}
vector with the labels of the layers. It must have the same 
length as layer.levels. Default is NULL}

\item{levels}{A \link[base]{character} or \link[base]{numeric}
vector with the levels of the raster values. Default is NULL}

\item{labels}{A \link[base]{character} or \link[base]{numeric}
vector with the labels of the raster values. It must have the same 
length as levels. Default is NULL}

\item{colors}{a set of aesthetic values to map data values to. See 
\link[ggplot2]{scale_fill_manual} for details}
}
\value{
A \link[ggplot2]{ggplot} object
}
\description{
Method for plotting the land use and land cover 
changes
}
\examples{

levels = c(seq_along(patterns.vignette), 255)
labels = c(names(patterns.vignette), "Unclassified")
colors = c("#996400", "#005500", "#D8B777", "#E6D219", "#E6BEC8", "#C8C8C8")
names(colors) = labels

x = system.file('lucc_MT.tif',  package = 'dtwSat')

### Plot maps
#gp1 = plotLUCC(x = x, type = "map", layer.labels = 2008:2013, 
#        levels = levels, labels = labels, colors = colors)
#gp1 

# Plot area 
#gp2 = plotLUCC(x = x, type = "area", layer.labels = 2008:2013, 
#        levels = levels, labels = labels, colors = colors) 
#gp2

# Plot land use changes 
#gp3 = plotLUCC(x = x, type = "change", layer.labels = 2008:2013, 
#        levels = levels, labels = labels, colors = colors) 
#gp3

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, 
\code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{plotCostMatrix}},
\code{\link[dtwSat]{plotAlignment}},
\code{\link[dtwSat]{plotMatch}}, and
\code{\link[dtwSat]{plotGroup}}
}

