% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMapSamples.R
\name{plotMapSamples}
\alias{plotMapSamples}
\title{Plotting maps}
\usage{
plotMapSamples(x, samples = "all", ...)
}
\arguments{
\item{x}{An object of class \code{\link[dtwSat]{twdtwAssessment}}.}

\item{samples}{a character defining the samples to plot 
"correct", "incorrect", "all". Default is "all".}

\item{...}{other arguments to pass to \code{\link[dtwSat]{twdtwRaster}}}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Method for plotting maps and samples.
}
\examples{
\dontrun{
#

}
}
\seealso{
\code{\link[dtwSat]{twdtwAssessment}},
\code{\link[dtwSat]{plotAccuracy}}, and
\code{\link[dtwSat]{plotAdjustedArea}}.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
