% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dtwSat.R
\docType{methods}
\name{plotMatch}
\alias{plotMatch}
\title{Plotting matching points of twdtw alignments}
\usage{
plotMatch(x, p.names, n, attr = 1, shift = 0.5)
}
\arguments{
\item{x}{An \code{\link[dtwSat]{dtwSat-class}} object}

\item{p.names}{A \link[base]{character} or \link[base]{integer}
vector with the patterns identification. If not declared the function 
will plot one alignment for each pattern in the 
\code{\link[dtwSat]{dtwSat-class}} object}

\item{n}{An \link[base]{integer} vector the same length as \code{p.names}.
The indices of the alignments for plotting. The alignments in the 
\code{\link[dtwSat]{dtwSat-class}} object are ordered by TWDTW distance in 
ascending order. If not declared the function will plot the best alignment 
for each pattern}

\item{attr}{An \link[base]{integer} or \link[base]{character} vector 
indicating the attribute for plotting, \emph{i.e.} a column of the \code{pattern}. 
Default is 1}

\item{shift}{A number, it shifts the pattern position in the \code{x}
direction. Default is 0.5}
}
\value{
A \link[ggplot2]{ggplot} object
}
\description{
Method for plotting the matching points of 
Time-Weighted DTW alignments
}
\examples{

weight.fun = logisticWeight(alpha=-0.1, beta=100)
alig = twdtw(patterns=patterns.list, timeseries=template, weight.fun = weight.fun, 
        normalize.patterns=TRUE, patterns.length=23, keep=TRUE)

gp = plotMatch(alig)
gp

gp = plotMatch(x=alig, p.names=1)
gp

gp = plotMatch(x=alig, p.names="Cotton")
gp

gp = plotMatch(x=alig, p.names=c("Cotton","Cotton","Cotton","Cotton"), 
     n = c(1:4))
gp

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{dtwSat-class}}, 
\code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{plotPath}}, 
\code{\link[dtwSat]{plotCostMatrix}},
\code{\link[dtwSat]{plotAlignment}}, and
\code{\link[dtwSat]{plotGroup}}
}

