% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMatches.R
\docType{methods}
\name{plotMatches}
\alias{plotMatches}
\title{Plotting TWDTW matching points}
\usage{
plotMatches(x, y, n, attr = 1, shift = 0.5, show.dist = FALSE)
}
\arguments{
\item{x}{A \code{\link[dtwSat]{twdtw-class}} object.}

\item{y}{A \link[base]{character} or \link[base]{integer}
vector with the patterns identification. If not declared the function 
will plot one alignment for each pattern in \code{x}.}

\item{n}{An \link[base]{integer} vector with the number of alignments to plot.
It can be use as vector of indices combined with \code{y} to plot 
specific matches. If not declared the function will plot the best match for 
each pattern.}

\item{attr}{An \link[base]{integer} or \link[base]{character} vector 
indicating the attribute for plotting. Default is 1.}

\item{shift}{A number, it shifts the pattern position in the \code{x}
direction. Default is 0.5.}

\item{show.dist}{show the distance for each alignment. Default is FALSE.}
}
\value{
A \link[ggplot2]{ggplot} object
}
\description{
Method for plotting the matching points from   
TWDTW analysis.
}
\examples{

log_fun = logisticWeight(alpha=-0.1, beta=100)
matches = twdtw(x=example_ts, patterns=patterns.list, weight.fun = log_fun, 
        normalize.patterns=TRUE, patterns.length=23, keep=TRUE)

gp = plotMatches(matches)
gp

gp = plotMatches(x=matches, y=1)
gp

gp = plotMatches(x=matches, y="Cotton")
gp

gp = plotMatches(x=matches, y="Soybean", n = 4)
gp

gp = plotMatches(x=matches, y=c("Soybean","Cotton"), 
     n = c(3,3))
gp

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, 
\code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{plotPaths}}, 
\code{\link[dtwSat]{plotCostMatrix}},
\code{\link[dtwSat]{plotAlignments}}, 
\code{\link[dtwSat]{plotClassification}}, 
\code{\link[dtwSat]{plotPatterns}}, and 
\code{\link[dtwSat]{plotTimeSeries}}.
}

