% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPaths.R
\docType{methods}
\name{plotPaths}
\alias{plotPaths}
\title{Plotting TWDTW paths}
\usage{
plotPaths(x, y, n = NULL)
}
\arguments{
\item{x}{An \code{\link[dtwSat]{twdtw-class}} object.}

\item{y}{A \link[base]{character} or \link[base]{numeric}
vector with the patterns identification. If not declared the function 
will plot the paths for all patterns.}

\item{n}{An \link[base]{integer} vector. The indices of the paths 
to plot. If not declared the function will plot all paths.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Method for plotting the minimum paths in the 
cost matrix of TWDTW.
}
\examples{

log_fun = logisticWeight(alpha=-0.1, beta=100)
matches = twdtw(x=example_ts, patterns=patterns.list, weight.fun = log_fun, 
        normalize.patterns=TRUE, patterns.length=23, keep=TRUE)
       
gp1 = plotPaths(matches, n=1:4)
gp1

gp2 = plotPaths(matches, y=c("Cotton","Maize"), n=1:4)
gp2

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, 
\code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{plotCostMatrix}},
\code{\link[dtwSat]{plotAlignments}}, 
\code{\link[dtwSat]{plotMatches}}, 
\code{\link[dtwSat]{plotClassification}},
\code{\link[dtwSat]{plotPatterns}}, and 
\code{\link[dtwSat]{plotTimeSeries}}.
}

