% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPatterns.R
\docType{methods}
\name{plotPatterns}
\alias{plotPatterns}
\title{Plotting temporal patterns}
\usage{
plotPatterns(x, p.names)
}
\arguments{
\item{x}{An \code{\link[dtwSat]{twdtw-class}} object or a list of 
\code{\link[zoo]{zoo}} objects}

\item{p.names}{A \link[base]{character} or \link[base]{numeric}
vector with the patterns identification. If not declared the function 
will plot the paths for all patterns}
}
\value{
A \link[ggplot2]{ggplot} object
}
\description{
Method for plotting the temporal patterns
}
\examples{

weight.fun = logisticWeight(alpha=-0.1, beta=100)
alig = twdtw(x=template, patterns=patterns.list, weight.fun = weight.fun, 
        normalize.patterns=TRUE, patterns.length=23, keep=TRUE)
       
gp1 = plotPatterns(alig)
gp1

gp2 = plotPatterns(patterns.list)
gp2



}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, 
\code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{plotCostMatrix}},
\code{\link[dtwSat]{plotAlignment}},
\code{\link[dtwSat]{plotMatch}}, and
\code{\link[dtwSat]{plotGroup}}
}

