% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\docType{methods}
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Plotting temporal patterns}
\usage{
plotTimeSeries(x, y)
}
\arguments{
\item{x}{A \code{\link[dtwSat]{twdtw-class}} object, a \code{\link[zoo]{zoo}},
or a list of \code{\link[zoo]{zoo}} objects.}

\item{y}{An integer. If x is a list \code{y} is an integer or character 
(list name(s)). If missing, all elements in the list will be plotted 
(up to a maximum of 6).}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Method for plotting the temporal patterns.
}
\examples{

log_fun = logisticWeight(alpha=-0.1, beta=100)
matches = twdtw(x=example_ts, patterns=patterns.list, weight.fun = log_fun, 
        normalize.patterns=TRUE, patterns.length=23)
       
gp1 = plotTimeSeries(matches)
gp1

gp2 = plotTimeSeries(patterns.list)
gp2

gp3 = plotTimeSeries(example_ts.list)
gp3

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, 
\code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{plotPaths}}, 
\code{\link[dtwSat]{plotCostMatrix}},
\code{\link[dtwSat]{plotAlignments}}, 
\code{\link[dtwSat]{plotMatches}}, 
\code{\link[dtwSat]{plotClassification}}, and 
\code{\link[dtwSat]{plotPatterns}}.
}

