% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtwSat.R
\docType{methods}
\name{plotTwdtw}
\alias{plotTwdtw}
\title{Plotting Time-Weighted DTW}
\usage{
plotTwdtw(x, n = 1, d = 1, type = "density", ...)
}
\arguments{
\item{x}{twdtw object}

\item{n}{An integer. The alignment. Default is 1}

\item{d}{An integer. The dimension for point-by-point
comparison, with matching lines \code{\link{dtwPlotThreeWay}}
in package \pkg{dtw}. Default is 1}

\item{type}{see \code{\link{plot.dtw}} in package \pkg{dtw}.
Default is "density"}

\item{...}{see \code{plot.dtw} in package \pkg{dtw}}
}
\description{
Methods for plotting Time-Weighted DTW objects
returned by twdtw.
}

