% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.knn1_twdtw}
\alias{predict.knn1_twdtw}
\title{Predict using the knn1_twdtw model}
\usage{
\method{predict}{knn1_twdtw}(object, newdata, ...)
}
\arguments{
\item{object}{A \code{knn1_twdtw} model object generated by the \code{knn1_twdtw} function.}

\item{newdata}{A data frame or similar object containing the new observations
(time series data) to be predicted.}

\item{...}{Additional arguments passed to the \link[twdtw]{twdtw} function.}
}
\value{
A vector of predicted classes for the \code{newdata}.
}
\description{
This function predicts the classes of new data using the Time Warped Dynamic Time Warping (TWDTW)
method with a 1-nearest neighbor approach. The prediction is based on the minimum TWDTW distance
to the known patterns stored in the \code{knn1_twdtw} model.
}
