% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{pretty_arguments}
\alias{pretty_arguments}
\title{Print Pretty Arguments}
\usage{
pretty_arguments(args)
}
\arguments{
\item{args}{A list of named arguments to display.}
}
\value{
Invisible NULL. The function is mainly used for its side effect of printing.
}
\description{
Display a list of arguments of a given function in a human-readable format.
}
\examples{
\dontrun{
pretty_arguments(formals(twdtw_knn1))
}

}
