% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtwSat.R
\docType{methods}
\name{print.twdtw}
\alias{print.twdtw}
\title{Method for Time-Weighted DTW}
\usage{
\method{print}{twdtw}(x, ...)
}
\arguments{
\item{x}{a \code{\link[dtwSat]{twdtw}} object}

\item{...}{additional arguments passed to print}
}
\description{
Methods for Time-Weighted DTW objects
}

