% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{print.twdtw_knn1}
\alias{print.twdtw_knn1}
\title{Print method for objects of class twdtw_knn1}
\usage{
\method{print}{twdtw_knn1}(x, ...)
}
\arguments{
\item{x}{An object of class \code{twdtw_knn1}.}

\item{...}{ignored}
}
\value{
Invisible \code{twdtw_knn1} object.
}
\description{
This method provides a structured printout of the important components
of a \code{twdtw_knn1} object.
}
