% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query.list}
\alias{query.list}
\title{Query time series}
\format{A named list of 3 zoo objects. Each node "Soybean",
"Cotton", and "Maize" has a zoo object with index of class Date
and 6 attributes: "ndvi", "evi", "red", "nir", "blue", "mir",
and "evi2".}
\usage{
query.list
}
\description{
This dataset has a list of queries form the
crops: Soybean, Cotton, and Maize. The time series are based on
the MODIS product MOD13Q1 250 m 16 days.
The queries were exttracted from more than 100 samples of each class
using Generalized Additive Models (GAM), see
package \link[mgcv]{gam}.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\keyword{datasets}

