% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{recoverDatesFromJulian}
\alias{recoverDatesFromJulian}
\title{Recover dates from year and julian number}
\usage{
recoverDatesFromJulian(years, n)
}
\arguments{
\item{n}{An integer with the julian number.}

\item{year}{An integers with the year.}
}
\description{
The function recovers the date from an year and
an julian  number.
}

