% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{recoverMODISDates}
\alias{recoverMODISDates}
\title{Recover MODIS dates}
\usage{
recoverMODISDates(years, frequency = 16)
}
\arguments{
\item{years}{An vector of integers with the years.Default
is form 2000 to the system time year format(Sys.time(), \"\%Y\").}

\item{frequency}{An integer with the frequency in days.}
}
\description{
The function recovers the MODIS image dates.
}

