% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{import}
\name{reexports}
\alias{arrangeGrob}
\alias{asymmetric}
\alias{autoplot}
\alias{grid.arrange}
\alias{rabinerJuangStepPattern}
\alias{reexports}
\alias{symmetric1}
\alias{symmetric2}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dtw}{\code{\link[dtw]{symmetric1}}, \code{\link[dtw]{symmetric2}}, \code{\link[dtw]{asymmetric}}, \code{\link[dtw]{rabinerJuangStepPattern}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{gridExtra}{\code{\link[gridExtra]{arrangeGrob}}, \code{\link[gridExtra]{grid.arrange}}}
}}
\keyword{internal}

