% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleTimeSeries.R
\docType{methods}
\name{resampleTimeSeries}
\alias{resampleTimeSeries}
\alias{resampleTimeSeries,twdtwTimeSeries-method}
\alias{resampleTimeSeries-twdtwMatches}
\title{Resample time series}
\usage{
resampleTimeSeries(object, length = NULL)

\S4method{resampleTimeSeries}{twdtwTimeSeries}(object, length = NULL)
}
\arguments{
\item{object}{an object of class twdtwTimeSeries.}

\item{length}{a number. If not informed the function will use the 
length of the longest pattern.}
}
\value{
An object of class \code{\link[dtwSat]{twdtwTimeSeries}} whose 
time series have the same number of samples (points).
}
\description{
resample time series in the same object to have the same 
the length.
}
\examples{
# Getting time series from objects of class twdtwTimeSeries
patterns = twdtwTimeSeries(timeseries=patterns.list, labels=names(patterns.list))
npatterns = resampleTimeSeries(patterns, length=46)
nrow(patterns)
nrow(npatterns)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwTimeSeries-class}}, and
\code{\link[dtwSat]{twdtwApply}}
}

