% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftDates.R
\docType{methods}
\name{shiftDates}
\alias{shiftDates}
\alias{shiftDates,twdtwTimeSeries-method}
\alias{shiftDates-twdtwMatches}
\title{Shift dates}
\usage{
shiftDates(object, year = NULL)

\S4method{shiftDates}{twdtwTimeSeries}(object, year = NULL)
}
\arguments{
\item{object}{an object of class twdtwTimeSeries.}

\item{year}{the base year to shit the time series.}
}
\value{
An object of class \code{\link[dtwSat]{twdtwTimeSeries}} whose 
time series are shifted to the same base year.
}
\description{
This function shifts the dates of the time series to a 
given base year.
}
\examples{
patterns = twdtwTimeSeries(timeseries=patterns.list, labels=names(patterns.list))
npatterns = shiftDates(patterns, year=2005)
index(patterns)
index(npatterns)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwTimeSeries-class}}, 
\code{\link[dtwSat]{getPatterns}}
}

