% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\docType{methods}
\name{show,dtwSat-method}
\alias{show,dtwSat-method}
\title{dtwSat-show}
\usage{
\S4method{show}{dtwSat}(object)
}
\arguments{
\item{object}{A \link[dtwSat]{dtwSat} object}
}
\description{
Show object of class dtwSat
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

