% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twdtw-show.R
\docType{methods}
\name{show,twdtw-method}
\alias{show,twdtw-method}
\title{Show method for twdtw-class}
\usage{
\S4method{show}{twdtw}(object)
}
\arguments{
\item{object}{An \code{\link[dtwSat]{twdtw-class}} object.}
}
\value{
Print the twdtw-class object.
}
\description{
Show the head of the matches ordered by 
TWDTW similarity.
}
\examples{

matches = twdtw(x=example_ts, patterns=patterns.list)
       
show(matches)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, and 
\code{\link[dtwSat]{twdtw}}.
}

