% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitDataset.R
\docType{methods}
\name{splitDataset}
\alias{splitDataset}
\title{split the time series samples and create patterns}
\usage{
splitDataset(timeseries, ref, times = 1, p = 0.1, mc.cores = 1, ...)
}
\arguments{
\item{timeseries}{A list of \code{\link[zoo]{zoo}} objects with the time series.}

\item{ref}{A vector with the identification of each time series. 
It must have the same length as \code{x}.}

\item{times}{Number of partitions to create.}

\item{...}{Other arguments to be passed to \code{\link[dtwSat]{createPattern}}.}
}
\value{
A \code{\link[base]{list}} of three attributes: the trained \code{pattrens},
the validation dataset \code{x}, and the reference for the validation dataset \code{ref}.
}
\description{
This function splits the set of samples for 
training and validation.
}
\examples{

set.seed(1)
field_samples = read.csv(system.file("lucc_MT/samples.csv", package="dtwSat")) 
proj_str = scan(file = system.file("lucc_MT/samples_projection", package="dtwSat"), 
                 what = "character")
reference = as.character(field_samples[["class"]])
load(system.file("lucc_MT/ts_list.RData", package="dtwSat"))

x = splitted_dataset = splitDataset(timeseries = ts_list, ref = reference, 
       times=2, p=0.1, mc.cores=1, freq=8, from="2007-09-01", to="2008-09-01", 
       formula = y ~ s(time, bs="cc"))

names(x)
names(x$Resample1)
plotPatterns(x$Resample1$patterns)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwAssessment}}
}

