% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtwSat.R
\docType{methods}
\name{statTimeSat}
\alias{statTimeSat}
\title{Function to compute phenological metrics.}
\usage{
statTimeSat(y, a = NULL, b = NULL)
}
\arguments{
\item{y}{A zoo object with the query time series.}

\item{a}{A vector. Indices in y of alignment starts.}

\item{b}{A vector. Indices in y of alignment ends.}
}
\description{
This compute phenological metrics from a time series.
}

