% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\docType{methods}
\name{summary,dtwSat-method}
\alias{summary,dtwSat-method}
\title{Summary method for dtwSat-class}
\usage{
\S4method{summary}{dtwSat}(object, ...)
}
\arguments{
\item{object}{An \code{\link[dtwSat]{dtwSat-class}} object}

\item{...}{additional arguments passed to summary}
}
\value{
A \link[base]{data.frame} object with the the summary 
for each pattern
}
\description{
Summary the alignments for each pattern in the 
dtwSat-class object
}
\examples{

alig = twdtw(patterns=patterns.list, timeseries=template)
       
show(alig)
summary(alig)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{dtwSat-class}}, and 
\code{\link[dtwSat]{twdtw}}
}

