% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twdtw-summary.R
\docType{methods}
\name{summary,twdtw-method}
\alias{summary,twdtw-method}
\title{Summary method for twdtw-class}
\usage{
\S4method{summary}{twdtw}(object, ...)
}
\arguments{
\item{object}{An \code{\link[dtwSat]{twdtw-class}} object}

\item{...}{additional arguments passed to summary}
}
\value{
A \link[base]{data.frame} object with the the summary 
for each pattern
}
\description{
Summary the alignments for each pattern in the 
twdtw-class object
}
\examples{

alig = twdtw(x=template, patterns=patterns.list)
       
show(alig)
summary(alig)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, and 
\code{\link[dtwSat]{twdtw}}
}

