% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{symmetric0}
\alias{symmetric0}
\title{dtw symmetric step with normalization N}
\format{symmetric0}
\usage{
symmetric0 # see ?stepPattern
}
\description{
A symmetric step with normalization N.
stepPattern object from package dtw, see ?stepPattern.
}
\keyword{datasets}

