% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template}
\alias{template}
\title{Template time series}
\format{A zoo objects with index of class Date and
6 attributes: "ndvi", "evi", "red", "nir", "blue", "mir",
and "evi2".}
\usage{
template
}
\description{
This dataset has a template time series based on the
MODIS product MOD13Q1 250 m 16 days. See package pkg{rwtss} that is an
R client for Web Time Series Service
\url{https://github.com/albhasan/rwtss.git}.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\keyword{datasets}

