% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtwSat.R
\docType{methods}
\name{timeSeriesAnalysis}
\alias{timeSeriesAnalysis}
\title{Satellite image time series analysis.}
\usage{
timeSeriesAnalysis(query, template, theta = 0, span = 2/3,
  normalize = TRUE, satStat = FALSE)
}
\arguments{
\item{query}{A zoo object with the query time series.}

\item{template}{A zoo object with the template time series.
It must be larger than the query.}

\item{theta}{A real between 1 and 0. It is a parameter for dtw local
cost matrix computation. For theta equal 0 the function computes a
normal dtw, for greater values it includes the time cost in the local
matrix. Default is 0.}

\item{span}{A real between 1 and 0. The span says how much the length
of a dtw match can be diferent from the length of the pattern. Default
is 2/3.}

\item{normalize}{A logical, default is TRUE. Setting TRUE it divides
the DTW distance by the pattern length.}

\item{satStat}{A logical, default is FALSE. Setting TRUE the function
also computes some statistics for each dtw match.}
}
\description{
This function applies an open boundary DTW analysis and
retrieves all possible alignments of a query within a template.
It also allows to compute some statistics for each dtw match.
}

