% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtwSat.R
\docType{methods}
\name{timeSeriesAnalysis2}
\alias{timeSeriesAnalysis2}
\title{Satellite image time series analysis.}
\usage{
timeSeriesAnalysis2(query, template, method = "dtw", threshold = Inf,
  normalize = TRUE, delay = NULL, theta = NULL, alpha = NULL,
  beta = NULL)
}
\arguments{
\item{query}{A zoo object with the query time series.}

\item{template}{A zoo object with the template time series.
It must be larger than the query.}

\item{method}{}

\item{threshold}{A real. The DTW distance threshold, i.e. the maximum DTW
distance for consideration. Default is Inf.}

\item{normalize}{A logical. It divides the DTW distance by the pattern
length. Default is TRUE.}

\item{delay}{A real. Parameter of petitjean method. The maximum delay for the dtw
alignment. Default is NULL.}

\item{theta}{A real. Parameter of lineartimeweight method. It is the slope
of the linear TWDTW. For theta equal 1 the time weight is equal to the
number of days. Default is NULL.}

\item{alpha}{A real. The steepness of logistictimeweight method. Default is NULL.}

\item{beta}{A real. The midpoint of logistictimeweight method. Default is NULL.}
}
\description{
This function applies an open boundary DTW analysis and
retrieves all possible alignments of a query within a template.
}

