% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{timeSeriesClassifier}
\alias{timeSeriesClassifier}
\title{Satellite image time series classification.}
\usage{
timeSeriesClassifier(dtwResults, from, to, by = 12, overlapping = 0.4,
  threshold = 4, sortBydtw = TRUE, aggregateByClass = TRUE)
}
\arguments{
\item{dtwResults}{A list of dtw .}

\item{from}{A character with the start date for the finalclassification.
The format is yyyy-mm-dd.}

\item{to}{A character with the end date for the finalclassification. The
format is yyyy-mm-dd.}

\item{by}{An integer. The length in months of the time subsequences
for the final classification. Default is 12 months.}

\item{overlapping}{A real between 0 and 1 representing the percentage of
minimum overlapping between the DTW mach and the subsequences for the
final classification. Default is 40\%.}

\item{threshold}{A real with the DTW threshold, i.e. the maximum DTW cost
for consideration.}

\item{sortBydtw}{A logical. If TRUE the function returns the best
classification for each period of time. If FALSE the functions
returns the lowest dtw for each class. Default is TRUE.}

\item{aggregateByClass}{A logical. Aggregate results by common class.
Default is TRUE}
}
\description{
This function classify the time segments based on the
open boundary DTW analysis results (i.e. results of timeSeriesAnalysis
function).
}
\examples{
##
}

