% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtwSat.R
\docType{methods}
\name{timeSeriesSmoothing}
\alias{timeSeriesSmoothing}
\title{Satellite image time series smoothing}
\usage{
timeSeriesSmoothing(x, y = NULL, timeline, frequency, method = c("wavelet",
  1))
}
\arguments{
\item{x}{Either a vector of dates or a zoo object with the
time series.}

\item{y}{A numeric vector.}

\item{timeline}{A time line vector for the internalsput.}

\item{frequency}{A numeric with the frequency for internalsput time series.}

\item{method}{A character vector for the smoothing methods, either
wavelet and level, eg. c(''wavelet'',''1'').}
}
\description{
This function applies a smoothing algorithm to
the satellite image time series. It allows to compute the
discreat wavelet smoothing for different levels. This function
only works for constant frequency data, then informa a constant
frequency time series. If it is not the case, either informe a
constant frequency time line or a frequency.
}
\examples{
sy = timeSeriesSmoothing(template$evi, frequency=16, method=c("wavelet",1))
plot(template$evi, xlab="Time", ylab="EVI")
lines(sy, col="red")
}

