% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_vignette.list}
\alias{ts_vignette.list}
\title{Data: A list of satellite time series}
\format{A named \code{list} of \link[zoo]{zoo} objects with a set of samples for: 
''Cotton-fallow'', 'Forest'', ''Soybean-cotton'', ''Soybean-maize'', and 
''Soybean-millet''. The indices of the \link[zoo]{zoo} objects are \code{\link[base]{Dates}} 
in the format ''yyyy-mm-dd''. Each node has 6 attributes: ''blue'', ''evi'', ''mir'', 
''ndvi'', ''nir'', and ''red''.}
\usage{
ts_vignette.list
}
\description{
This dataset has a list of time series based on the 
MODIS product MOD13Q1 250 m 16 days [1]. It is an irregularly sampled time series 
using the real date of each pixel from ''2009-08-05'' to ''2013-07-31''.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\references{
[1] Friedl MA, Sulla-Menashe D, Tan B, Schneider A, Ramankutty N, Sibley A, Huang X. (2010).
MODIS Collection 5 global land cover: Algorithm refinements and characterization of new
datasets. Remote Sensing of Environment, 114(1), 168 182.
}
\seealso{
\link[dtwSat]{example_ts}, and 
\link[dtwSat]{patterns.list}.

The package \pkg{rwtss} provides a client for Web Time 
Series Service \url{https://github.com/albhasan/rwtss.git}.

MOD13Q1 documentation: 
\url{https://lpdaac.usgs.gov/dataset_discovery/modis/modis_products_table/mod13q1}.
}
\keyword{datasets}

