% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twdtwApply.R
\docType{methods}
\name{twdtwApply}
\alias{twdtwApply}
\title{Apply Time-Weighted Dynamic Time Warping to raster time series}
\usage{
twdtwApply(x, patterns, win.fun, win.size = c(1, 1), chunk.size = 100,
  chunk.overlap, mc.cores = 1, ...)
}
\arguments{
\item{x}{A list of \code{\link[raster]{Raster-class}}
\code{\link[raster]{brick}} or \code{\link[raster]{stack}} object.
Each layer of the Raster* object is a time step. 
See \code{\link[dtwSat]{buildRasterTimeSeries}}.}

\item{patterns}{a list of \link[zoo]{zoo} objects. See \code{\link[dtwSat]{twdtw}} for
details.}

\item{win.fun}{A function. A function to be applied to the TWDTW results. See Details.}

\item{win.size}{A numeric vector. The size of a processing window in col x row order.
Default is a single pixel, \emph{i.e.} \code{win.size=c(1,1)}.}

\item{chunk.size}{An integer. Set the number of cells for each block, 
see \code{\link[raster]{blockSize}} for details.}

\item{chunk.overlap}{A numeric vector. The overlap of neighboring chunks of a 
raster. Overlap is specified for each dimension of the raster col x row. 
Usefull when \code{win.size} bigger than a single pixel. Normally it is the size of 
the window.}

\item{mc.cores}{The number of cores to use, See \code{\link[parallel]{mclapply}} 
for details.}

\item{...}{other arguments to pass to the functions \code{\link[dtwSat]{twdtw}} and 
\code{win.fun}. Note that the 'win.fun' should explicitly name the arguments.}
}
\value{
A \code{\link[raster]{brick}} object.
}
\description{
This function applies a multidimensional Time-Weighted DTW 
analysis for each pixel location in the raster time series and retrieves 
a classified raster time series.
}
\examples{

##### In the next example we build multi-band raster time series and run  
##### the TWDTW analyis. 

#### In this example we build a multi-band MOD13Q1 raster time series. 
#### The 'tif' files in 'lucc_MT/tif' have 999 EVI time series 
#### from 2007-01-01 to 2013-12-19, that means 160 points with temporal 
#### resolution of 16 days. 
blue = brick(system.file("lucc_MT/timeseries/blue.tif",  package = "dtwSat"))
red  = brick(system.file("lucc_MT/timeseries/red.tif",  package = "dtwSat"))
nir  = brick(system.file("lucc_MT/timeseries/nir.tif",  package = "dtwSat"))
mir  = brick(system.file("lucc_MT/timeseries/mir.tif",  package = "dtwSat"))
evi  = brick(system.file("lucc_MT/timeseries/evi.tif",  package = "dtwSat"))
ndvi = brick(system.file("lucc_MT/timeseries/ndvi.tif",  package = "dtwSat"))
dates = scan(system.file("lucc_MT/timeseries/timeline", package = "dtwSat"), what = "dates")
raster_timeseries = 
     buildRasterTimeSeries(blue, red, nir, mir, evi, ndvi, timeline = dates)
 
#### Run TWDTW analysis (~1.2 min using 4 core with 2.4 GHz clock)
# class_names = names(patterns_vignette.list)
# names(class_names) = class_names
# patterns_levels = c(seq_along(class_names), 255)
# patterns_labels = c(class_names, "Unclassified")
# system.time(
# land_use_maps <- 
#      twdtwApply(x = raster_timeseries, 
#      patterns = patterns_vignette.list, 
#      mc.cores = 4, win.fun = classifyIntervals, 
#      weight.fun = logisticWeight(alpha=-0.1, beta=50), 
#      from = as.Date("2007-09-01"),to = as.Date("2013-09-01"),
#      by = "12 month", overlap = 0.5, levels = patterns_levels,
#      labels = patterns_labels, simplify= TRUE)
#   )
#### Plot results 
# colors = c("#996400", "#005500", "#D8B777", "#E6D219", "#E6BEC8", "#C8C8C8")
# names(colors) = patterns_labels
## Land use maps
# gp1 = plotLUCC(x = land_use_maps, type = "map", layer.labels = 2008:2013, 
#          class.levels = patterns_levels, class.labels = patterns_labels, 
#          class.colors = colors)
# gp1 
## Accumulated area 
# gp2 = plotLUCC(x = land_use_maps, type = "area", layer.labels = 2008:2013, 
#          class.levels = patterns_levels, class.labels = patterns_labels, 
#          class.colors = colors)
# gp2
## Land use changes 
# gp3 = plotLUCC(x = land_use_maps, type = "change", layer.labels = 2008:2013, 
#          class.levels = patterns_levels, class.labels = patterns_labels, 
#          class.colors = colors)
# gp3

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{buildRasterTimeSeries}}, 
\code{\link[dtwSat]{classifyIntervals}}, and  
\code{\link[dtwSat]{plotLUCC}}.
}

