% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{twdtwApplyiquery}
\alias{twdtwApplyiquery}
\title{iquery builder}
\usage{
twdtwApplyiquery(str_iquery, logfile = NULL)
}
\arguments{
\item{str_iquery}{A list of iquery strings to be executed in SciBD.}

\item{logfile}{The path for the log file.}
}
\description{
This function runs the iquery string on SciDB.
}
\examples{

out_schema = "<ts_id:double,year:double,cdoy:double,distance:double> 
[col_id=59354:60132,256,1,row_id=48591:49096,256,1,time_id=0:99,100,0,class_id=0:255,256,0]"

str_iquery = iqueryBlock(array.in = "MOD13Q1", attributes=c("evi", "ndvi", "cdoy"), 
                        outputschema = out_schema, 
                        expr="as.list(as.double(c(60124, 48978,1,1,1:4)))", 
                        min.col = 60124, max.col = 60125, block.col = 256, 
                        min.row = 48978, max.row = 48979, block.row = 256,
                        min.time = 0, max.time = 99, index.start=c(48000,38400,0), 
                        index.end = c(72000, 62400, 14999), chunk=c(32,32,99))

str_iquery

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{iqueryBlock}
}

