% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twdtwApply.R
\name{twdtwApplySciDB}
\alias{twdtwApplySciDB}
\title{TWDTW analysis on SciDB}
\usage{
twdtwApplySciDB(x, y, resample = TRUE, length = NULL, weight.fun = NULL,
  dist.method = "Euclidean", step.matrix = symmetric1, n = NULL,
  span = NULL, min.length = 0.5, theta = 0.5, breaks = NULL,
  from = NULL, to = NULL, by = NULL, overlap = 0.5, ...)
}
\arguments{
\item{x}{an object of class twdtw*. This is the target time series. 
Usually, it is a set of unclassified time series.}

\item{y}{an object of class \link[dtwSat]{twdtwTimeSeries}. 
The temporal patterns.}

\item{resample}{resample the patterns to have the same length. Default is TRUE.
See \link[dtwSat]{resampleTimeSeries} for details.}

\item{length}{An integer. Patterns length used with \code{patterns.length}. 
If not declared the length of the output patterns will be the length of 
the longest pattern.}

\item{weight.fun}{A function. Any function that receive and performs a 
computation on a matrix. The function receives a matrix of time differences 
in days and returns a matrix of time-weights. If not declared the time-weight 
is zero. In this case the function runs the standard version of the dynamic 
time warping. See details.}

\item{dist.method}{A character. Method to derive the local cost matrix.
Default is ''Euclidean'' see \code{\link[proxy]{dist}} in package 
\pkg{proxy}.}

\item{step.matrix}{see \code{\link[dtw]{stepPattern}} in package \pkg{dtw} [2].}

\item{n}{An integer. The maximun number of matches to perform. 
NULL will return all matches.}

\item{span}{A number. Span between two matches, \emph{i.e.} the minimum
interval between two matches, for details see [3]. If not declared it removes
all overlapping matches of the same pattern. To include overlapping matches 
of the same pattern use \code{span=0}.}

\item{min.length}{A number between 0 an 1. This argument removes the over fittings.
Minimum length after warping. Percentage of the original pattern length. Default is 0.5, 
meaning that the matching cannot be shorter than half of the pattern length.}

\item{theta}{numeric between 0 and 1. The weight of the time 
for the TWDTW computation. Use \code{theta=0} to cancel the time-weight, 
\emph{i.e.} to run the original DTW algorithm. Default is 0.5, meaning that 
the time has the same weight as the curve shape in the TWDTW analysis.}

\item{breaks}{A vector of class \code{\link[base]{Dates}}. This replaces the arguments \code{from},
\code{to}, and \code{by}.}

\item{from}{A character or \code{\link[base]{Dates}} object in the format "yyyy-mm-dd".}

\item{to}{A \code{\link[base]{character}} or \code{\link[base]{Dates}} object in the format "yyyy-mm-dd".}

\item{by}{A \code{\link[base]{character}} with the intevals size, \emph{e.g.} "6 month".}

\item{overlap}{A number between 0 and 1. The minimum overlapping 
between one match and the interval of classification. Default is 0.5, 
\emph{i.e.} an overlap minimum of 50\%.}

\item{...}{arguments to pass to \code{\link[raster]{writeRaster}} and for parallel 
processing to pass to \code{\link[parallel]{mclapply}}.}
}
\value{
An object of class \code{\link[base]{list}} whose attributes are:
"time_id", "class_id", "distance", "ts_id", "row_id", "col_id", "cdoy", and "year".
}
\description{
This function calls the method twdtwApply and retrieves the results 
as a list.
}
\details{
The output of this function is a list with the attributes used by r_exec to save 
a SciDB array. The labels of the input time series must have the format "<row>_<col>", where 
"row" and "col" are the indices of the SciDB array where the time sereis come from.
}
\examples{
log_fun = logisticWeight(-0.1, 100)
ts = twdtwTimeSeries(example_ts.list, labels = paste(c(1,2), c(1,1), sep="_") )
patt = twdtwTimeSeries(patterns.list)
time_intervals = seq(from=as.Date("2007-09-01"), to=as.Date("2013-09-01"), by="6 month")
mat1 = twdtwApplySciDB(x=ts, y=patt, weight.fun=log_fun, breaks=time_intervals)
mat1

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{twdtwApply}
}

