% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twdtwAssessment.R
\docType{methods}
\name{twdtwAssessment}
\alias{twdtwAssessment}
\title{Assess TWDTW analyis using validation samples}
\usage{
twdtwAssessment(x, mc.cores = 1, overlap = 0.5, threshold = Inf, ...)
}
\arguments{
\item{x}{A \code{\link[base]{list}} of three attributes: the trained \code{pattrens},
the validation dataset \code{x}, and the reference for the validation dataset \code{ref}.
See \code{\link[dtwSat]{splitDataset}} for details.}

\item{mc.cores}{The number of cores to use, See \code{\link[parallel]{mclapply}} 
for details.}

\item{overlap}{A number between 0 and 1. The minimum overlapping between one match and the 
interval of classification. Default is 0.3, i.e. an overlap minimum of 30\%. 
See \code{\link[dtwSat]{classifyIntervals}} for details.}

\item{threshold}{A number. The TWDTW dissimilarity threshold, \emph{i.e.} the maximum TWDTW cost 
for consideration in the classification. Default is Inf.
See \code{\link[dtwSat]{classifyIntervals}} for details.}

\item{...}{Other arguments to be passed to \code{\link[dtwSat]{twdtw}}.}
}
\value{
A \code{\link[base]{data.frame}} of two attributes \code{Reference} 
and \code{Predicted}.
}
\description{
This function classifies the set of samples time series
}
\examples{

# set.seed(1)
# field_samples = read.csv(system.file("lucc_MT/samples.csv", package="dtwSat")) 
# proj_str = scan(file = system.file("lucc_MT/samples_projection", package="dtwSat"), 
#                  what = "character")
# reference = as.character(field_samples[["class"]])
# load(system.file("lucc_MT/ts_list.RData", package="dtwSat"))

# x = splitted_dataset = splitDataset(timeseries = ts_list, ref = reference, 
#        times=2, p=0.1, mc.cores=1, freq=8, from="2007-09-01", to="2008-09-01", 
#        formula = y ~ s(time, bs="cc"))

# assess_table = lapply(x, FUN=twdtwAssessment, overlap = 0.5, 
#             weight.fun = logisticWeight(alpha=-0.1, beta=50))

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{splitDataset}}
}

