% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-twdtwMatches.R, R/methods-getAlignments.R, R/methods-getTimeSeries.R, R/methods-is.R, R/methods-labels.R, R/methods-length.R, R/methods-show.R
\docType{methods}
\name{twdtwMatches-class}
\alias{getAlignments,twdtwMatches-method}
\alias{getPatterns,twdtwMatches-method}
\alias{getTimeSeries,twdtwMatches-method}
\alias{is.twdtwMatches}
\alias{is.twdtwMatches,twdtwMatches-method}
\alias{labels,twdtwMatches-method}
\alias{length,twdtwMatches-method}
\alias{show,twdtwMatches-method}
\alias{twdtwMatches-class}
\title{class "twdtwMatches"}
\usage{
\S4method{getAlignments}{twdtwMatches}(object)

\S4method{getTimeSeries}{twdtwMatches}(object, labels = NULL)

\S4method{getPatterns}{twdtwMatches}(object, labels = NULL)

\S4method{is.twdtwMatches}{twdtwMatches}(x)

\S4method{labels}{twdtwMatches}(object)

\S4method{length}{twdtwMatches}(x)

\S4method{show}{twdtwMatches}(object)
}
\arguments{
\item{object}{an object of class twdtwMatches.}

\item{labels}{a vector with labels of the time series.}

\item{x}{an object of class twdtwMatches.}
}
\description{
Class for Time-Weighted Dynamic Time Warping results.
}
\section{Methods (by generic)}{
\itemize{
\item \code{getAlignments}: Get TWDTW alignments from the object of class twdtwMatches

\item \code{getTimeSeries}: Get time series from objects of class twdtwMatches

\item \code{getPatterns}: Get patterns from objects of class twdtwMatches

\item \code{is.twdtwMatches}: Check if the object belongs to the class twdtwMatches

\item \code{labels}: Labels of the time series in object of class twdtwMatches

\item \code{length}: Number of time series in objects of class twdtwMatches

\item \code{show}: Show object of class twdtwMatches
}}
\section{Slots }{

\describe{
 \item{\code{timeseries}:}{A object of class \link[dtwSat]{twdtwTimeSeries} with the satellite time series.}
 \item{\code{pattern}:}{A object of class \link[dtwSat]{twdtwTimeSeries} with the temporal patterns.}
 \item{\code{alignments}:}{A \code{\link[base]{list}} of TWDTW results with the same length as 
 the \code{timeseries}. Each element in this list has the following results for each temporal pattern 
 in \code{patterns}:
      \cr\code{from}: a vector with the starting dates of each match in the format "YYYY-MM-DD",
      \cr\code{to}: a vector with the ending dates of each match in the format "YYYY-MM-DD", 
	     \cr\code{distance}: a vector with TWDTW dissimilarity measure, and
	     \cr\code{K}: the number of matches of the pattern.
 }
	\item{This list might have additional elements:}{ if \code{keep=TRUE} in the \code{twdtw} call 
	the list is extended to include internal structures used during the TWDTW computation: 
      \cr\code{costMatrix}: cumulative cost matrix,
      \cr\code{directionMatrix}: directions of steps that would be taken from each element of matrix,
      \cr\code{startingMatrix}: the starting points of each element of the matrix,
      \cr\code{stepPattern}: \code{\link[dtw]{stepPattern}} used for the 
      computation, see package \code{\link[dtw]{dtw}},
      \cr\code{N}: the length of the \code{pattern}, 
      \cr\code{M}: the length of the time series \code{timeseries}, 
      \cr\code{timeWeight}: time weight matrix,
      \cr\code{localMatrix}: local cost matrix,
	     \cr\code{matching}: A list whose elements have the matching points for 
	     each match between pattern the time series, such that:
         \cr--\code{index1}: a vector with matching points of the pattern, and
         \cr--\code{index2}: a vector with matching points of the time series.
      }
}
}
\examples{
# Creating new object of class twdtwTimeSeries  
new("twdtwMatches")

# Getting time series from objects of class twdtwTimeSeries
matches = new("twdtwMatches")
getAlignments(matches)

## Getting time series from objects of class twdtwMatches 
matches = new("twdtwMatches")
getTimeSeries(matches)

## Getting patterns from objects of class twdtwMatches 
matches = new("twdtwMatches")
getPatterns(matches)

# Checking if object belongs to the class twdtwMatches
matches = new("twdtwMatches")
is.twdtwMatches(matches)

# Getting labels of time series in objects of class twdtwMatches
matches = new("twdtwMatches")
labels(matches)

# Getting the number of time series in objects of class twdtwMatches
matches = new("twdtwMatches")
length(matches)

# Showing objects of class twdtwMatches 
matches = new("twdtwMatches")
show(matches)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

