% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-twdtwMatches.R
\name{twdtwMatches}
\alias{twdtwMatches}
\title{Create twdtwMatches object}
\usage{
twdtwMatches(timeseries, patterns, alignments)
}
\arguments{
\item{timeseries}{an object of class twdtwTimeSeries.}

\item{patterns}{an object of class twdtwTimeSeries.}

\item{alignments}{an object of class list with the TWDTW results. 
For more details see slot alignments of twdtwMatches-class.}
}
\description{
Create object of class twdtwMatches.
}
\examples{
# Creating objects of class twdtwMatches 
twdtwMatches(timeseries = twdtwTimeSeries(), 
             patterns = twdtwTimeSeries(), alignments = list())

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwMatches-class}} and 
\code{\link[dtwSat]{twdtwApply}}
}

