% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwtSat.R
\name{twdtwProxyRegister}
\alias{twdtwProxyRegister}
\title{TWDTW distance proxy register}
\usage{
twdtwProxyRegister()
}
\description{
This function performs a multidimensional Time-Weighted DTW 
analysis and retrieves the matches between the temporal patterns and 
a set of time series [1].
}
\details{
Register TWDTW as a distance function into package proxy
}
\seealso{
\code{\link[proxy]{pr_DB}},
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
