% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-twdtwRaster.R, R/labels.R, R/show.R, R/utils.R
\docType{methods}
\name{twdtwRaster-class}
\alias{dim,twdtwRaster-method}
\alias{extent,twdtwRaster-method}
\alias{index,twdtwRaster-method}
\alias{labels,twdtwRaster-method}
\alias{length,twdtwRaster-method}
\alias{ncol,twdtwRaster-method}
\alias{nlayers,twdtwRaster-method}
\alias{nrow,twdtwRaster-method}
\alias{projection,twdtwRaster-method}
\alias{res,twdtwRaster-method}
\alias{show,twdtwRaster-method}
\alias{twdtwRaster,ANY-method}
\alias{twdtwRaster-class}
\title{class "twdtwRaster"}
\usage{
\S4method{twdtwRaster}{ANY}(..., timeline, doy = NULL, layers = NULL,
  filepath = NULL)

\S4method{labels}{twdtwRaster}(object)

\S4method{show}{twdtwRaster}(object)

\S4method{dim}{twdtwRaster}(x)

\S4method{res}{twdtwRaster}(x)

\S4method{extent}{twdtwRaster}(x)

\S4method{projection}{twdtwRaster}(x)

\S4method{ncol}{twdtwRaster}(x)

\S4method{nrow}{twdtwRaster}(x)

\S4method{nlayers}{twdtwRaster}(x)

\S4method{index}{twdtwRaster}(x)

\S4method{length}{twdtwRaster}(x)
}
\arguments{
\item{...}{objects of class \code{\link[raster]{RasterBrick-class}} or 
\code{\link[raster]{RasterStack-class}}.}

\item{timeline}{a vector with the dates of the satellite images 
in the format of "YYYY-MM-DD".}

\item{doy}{A \code{\link[raster]{RasterBrick-class}} or 
\code{\link[raster]{RasterStack-class}} with a sequence of days of the year for each pixel. 
\code{doy} must have the same spatial and temporal extents as the Raster* objects passed to \code{...}.
If \code{doy} is not informed then at least one Raster* object must be passed through \code{...}.}

\item{layers}{a vector with the names of the \code{Raster*} objects 
passed to "\code{...}". If not informed the layers are set to the 
names of the Raster* objects.}

\item{filepath}{A character. The path to save the raster time series. If informed the 
function saves a raster file for each Raster* object in the list, \emph{i.e} one file 
for each time series. This way the function retrieves an list of 
\code{\link[raster]{RasterBrick-class}}. It is useful when the time series are 
originally stores in separated files. See details.}

\item{object}{an object of class twdtwRaster.}

\item{x}{an object of class twdtwRaster.}
}
\description{
Class for set of satellite time series.
}
\details{
The performance the functions \code{\link[dtwSat]{twdtwApply}} and 
\code{\link[dtwSat]{getTimeSeries}} is improved if the Raster* objects are connected 
to files with the whole time series for each attribute.
}
\section{Methods (by generic)}{
\itemize{
\item \code{twdtwRaster}: Create object of class twdtwRaster.
}}
\section{Slots }{

\describe{
 \item{\code{timeseries}:}{A list of multi-layers Raster* objects 
       with the satellite image time series.}
 \item{\code{timeline}:}{A vector of class \code{\link[base]{date}} 
       with dates of the satellite images in \code{timeseries}.}
 \item{\code{layers}:}{A vector of class \code{\link[base]{factor}} 
       with the names of the Raster* objects.}
}
}
\examples{
# Creating new object of class twdtwTimeSeries  

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

