% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-twdtwRaster.R
\name{twdtwRaster}
\alias{twdtwRaster}
\title{Create twdtwRaster object}
\usage{
twdtwRaster(..., timeline, doy = NULL, layers = NULL, labels = NULL,
  filepath = NULL)
}
\arguments{
\item{...}{objects of class \code{\link[raster]{RasterBrick-class}} or 
\code{\link[raster]{RasterStack-class}}.}

\item{timeline}{a vector with the dates of the satellite images 
in the format of "YYYY-MM-DD".}

\item{doy}{A \code{\link[raster]{RasterBrick-class}} or 
\code{\link[raster]{RasterStack-class}} with a sequence of days of the year for each pixel. 
\code{doy} must have the same spatial and temporal extents as the Raster* objects passed to \code{...}.
If \code{doy} is not informed then at least one Raster* object must be passed through \code{...}.}

\item{layers}{a vector with the names of the \code{Raster*} objects 
passed to "\code{...}". If not informed the layers are set to the 
names of objects in "\code{...}".}

\item{labels}{a vector of class \code{\link[base]{factor}} with 
levels and labels of the values in the Raster* objects. This is 
useful for categorical Raster* values of land use classes.}

\item{filepath}{A character. The path to save the raster time series. If informed the 
function saves a raster file for each Raster* object in the list, \emph{i.e} one file 
for each time series. This way the function retrieves an list of 
\code{\link[raster]{RasterBrick-class}}. It is useful when the time series are 
originally stores in separated files. See details.}
}
\description{
Create object of class twdtwRaster.
}
\details{
The performance the functions \code{\link[dtwSat]{twdtwApply}} and 
\code{\link[dtwSat]{getTimeSeries}} is improved if the Raster* objects are connected 
to files with the whole time series for each attribute.
}
\examples{
# Creating objects of class twdtwRaster 
evi = brick(system.file("lucc_MT/data/evi.tif", package="dtwSat"))
ndvi = brick(system.file("lucc_MT/data/ndvi.tif", package="dtwSat"))
blue = brick(system.file("lucc_MT/data/blue.tif", package="dtwSat"))
red = brick(system.file("lucc_MT/data/red.tif", package="dtwSat"))
nir = brick(system.file("lucc_MT/data/nir.tif", package="dtwSat"))
mir = brick(system.file("lucc_MT/data/mir.tif", package="dtwSat"))
doy = brick(system.file("lucc_MT/data/doy.tif", package="dtwSat"))
timeline = scan(system.file("lucc_MT/data/timeline", package="dtwSat"), what="date")
rts=twdtwRaster(evi, ndvi, blue, red, nir, mir, timeline=timeline, doy=doy)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwRaster-class}},
\code{\link[dtwSat]{getTimeSeries}}, and 
\code{\link[dtwSat]{twdtwApply}}
}

