% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-twdtwTimeSeries.R
\docType{methods}
\name{twdtwTimeSeries,list-method}
\alias{twdtwTimeSeries,list-method}
\title{twdtwTimeSeries-class}
\usage{
\S4method{twdtwTimeSeries}{list}(...)
}
\description{
twdtwTimeSeries-class
}
\examples{
# Creating objects of class twdtwTimeSeries from list of zoo objects 
patt = twdtwTimeSeries(patterns.list)
patt

}

