% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-twdtwTimeSeries.R
\name{twdtwTimeSeries}
\alias{twdtwTimeSeries}
\title{Create twdtwTimeSeries object}
\usage{
twdtwTimeSeries(..., timeseries, labels)
}
\arguments{
\item{...}{objects of class \code{\link[zoo]{zoo}}.}

\item{timeseries}{a list of \code{\link[zoo]{zoo}} objects.}

\item{labels}{a vector with labels of the time series.}
}
\description{
Create object of class twdtwTimeSeries.
}
\examples{
# Creating objects of class twdtwTimeSeries 
class(example_ts)
twdtwTimeSeries(example_ts)
sapply(example_ts.list, class)
twdtwTimeSeries(timeseries = example_ts.list)
twdtwTimeSeries(timeseries = example_ts.list, labels = c(1,2))
twdtwTimeSeries(example_ts.list[[1]], example_ts.list[[2]])

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

