% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{twdtwTimeSeriesFromVector}
\alias{twdtwTimeSeriesFromVector}
\title{Create time series from vectors}
\usage{
twdtwTimeSeriesFromVector(x, y, timeline, ...)
}
\arguments{
\item{x}{A vector of integers with columns indices.}

\item{y}{A vector of integers with rows indices.}

\item{timeline}{A vector of dates with the date in the format ''YYYY-MM-DD''.}

\item{...}{numeric vectors with the data. This are the attributes of the time series.}
}
\value{
An object of class \code{\link[dtwSat]{twdtwTimeSeries}}
}
\description{
This function creates a twdtwTimeSeries object using data stored in 
vectors.
}
\details{
All arguments passed to this function must have the same length.
}
\examples{
n = seq(100)
col_id = rep(c(1,2), each=length(n))
row_id = rep(1, length(col_id))
from = as.Date("2000-01-01")
dates = rep(seq(from, from+length(n)-1, by=1), 2)

t = seq(0,4*pi,,length(col_id))
attr1 = 3*sin(2*t)+runif(seq(length(col_id)))*2
attr2 = 3*sin(2*t)+rnorm(seq(length(col_id)))*2

ts = twdtwTimeSeriesFromVector(x=col_id, y=row_id, timeline=dates, attr1, attr2)
plot(ts, type="timeseries")

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{createTimeSequence}
}

