% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRasterTimeSeries.R
\docType{methods}
\name{buildRasterTimeSeries}
\alias{buildRasterTimeSeries}
\title{Build raster time series}
\usage{
buildRasterTimeSeries(x, timeline, doy, filepath = NULL, mc.cores = 1, ...)
}
\arguments{
\item{x}{A list of \code{\link[raster]{RasterBrick-class}} or 
\code{\link[raster]{RasterStack-class}}
Each layer of the Raster* object is a time step.}

\item{timeline}{A vector of \code{\link[base]{Dates}}
It must have the length of the layers in the Raster* object.}

\item{doy}{optional. A \code{\link[raster]{RasterBrick-class}} or 
\code{\link[raster]{RasterStack-class}} with the same spatial and temporal extent 
as the Raster* objects in \code{x}.}

\item{filepath}{A character. The path to save the raster time series. If informed the 
function saves a raster file for each Raster* object in the list, \emph{i.e} one file 
for each time series. This way the function retrieves an list of 
\code{\link[raster]{RasterBrick-class}} objects. See details.}

\item{mc.cores}{The number of cores to use, See \code{\link[parallel]{mclapply}} 
for details.}

\item{...}{other arguments to pass to the function \code{\link[raster]{writeRaster}}.}
}
\value{
details a list of \code{\link[raster]{RasterBrick-class}} or 
\code{\link[raster]{RasterStack-class}} objects.
}
\description{
This function creates a list of raster time series.
}
\details{
The performance the functions \code{\link[dtwSat]{twdtwApply}} and 
\code{\link[dtwSat]{extractTimeSeries}} is improved if the Raster* objects are connected 
to files with the whole time series for each attribute.
}
\examples{

####
 #require(raster)
 ## Read EVI raster file and timeline 
 #r = brick(system.file("lucc_MT/raster_ts/evi.tif",  package = "dtwSat"))
 #time = read.csv(system.file("lucc_MT/timeline.csv",  package = "dtwSat"), 
 #        as.is=TRUE)
         
 ## Build raster time series 
 #raster_ts = buildRasterTimeSeries(x = r, timeline = dates$date)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwApply}}, and
\code{\link[dtwSat]{plotLUCC}}.
}

