% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyIntervals.R
\docType{methods}
\name{classifyIntervals}
\alias{classifyIntervals}
\title{Classify time intervals}
\usage{
classifyIntervals(x, from = NULL, to = NULL, by = NULL, breaks = NULL,
  overlap = 0.3, threshold = Inf, simplify = FALSE, levels = NULL,
  labels = NULL, Unclassified = 255, p.names)
}
\arguments{
\item{x}{A \code{\link[dtwSat]{twdtw-class}} object or 
a \code{\link[base]{data.frame}} such as retrieved by \code{\link[dtwSat]{getAlignments}}.}

\item{from}{A character or \code{\link[base]{Dates}} object in the format "yyyy-mm-dd".}

\item{to}{A \code{\link[base]{character}} or \code{\link[base]{Dates}} object in the format "yyyy-mm-dd".}

\item{by}{A \code{\link[base]{character}} with the intevals size, \emph{e.g.} "6 month".}

\item{breaks}{A vector of class \code{\link[base]{Dates}}. This replaces the arguments \code{from},
\code{to}, and \code{by}.}

\item{overlap}{A number between 0 and 1. The minimum overlapping 
between one match and the interval of classification. Default is 0.3, 
\emph{i.e.} an overlap minimum of 30\%.}

\item{threshold}{A number. The TWDTW dissimilarity threshold, i.e. the maximum TWDTW 
cost for consideration in the classification. Default is \code{Inf}.}

\item{simplify}{return only the best pattern for each interval. Default is FALSE.}

\item{levels}{A character or numeric vector. The levels for the classification.}

\item{labels}{A character or numeric vector. The labels for each level.}

\item{Unclassified}{A numeric to fill classification gaps. Default is 255.}

\item{p.names}{A \link[base]{character} or \link[base]{numeric}
vector with the patterns identification. If not declared the function 
considers all patterns in the classification.}
}
\value{
A \code{\link[base]{data.frame}} with the best match 
for each interval of classification.
}
\description{
This function retrieves the best match within each 
predefined period based on the TWDTW dissimilarity measure.
}
\examples{

log_fun = logisticWeight(alpha=-0.1, beta=100)

matches = twdtw(x=example_ts, patterns=patterns.list, weight.fun = log_fun)
         
# Classify interval
from = as.Date("2009-09-01")
to = as.Date("2013-09-01")
by = "6 month"

# All classes
classifyIntervals(x=matches, from=from, to=to, by = by,
             overlap=.3, threshold=Inf)

# Cotton and Maize 
classifyIntervals(x=matches, from=from, to=to, by = by,
             overlap=.3, threshold=Inf, p.names=c("Cotton","Maize"))

# Simplify Cotton and Maize 
classifyIntervals(x=matches, from=from, to=to, by = by, simplify= TRUE,
             overlap=.3, threshold=Inf, p.names=c("Cotton","Maize"))
             
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

