% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{classifyIntervals}
\alias{classifyIntervals}
\title{Classify time intervals}
\usage{
classifyIntervals(x, from, to, by, breaks, overlap = 0.3, threshold = Inf,
  ...)
}
\arguments{
\item{x}{A \code{\link[dtwSat]{dtwSat-class}} object or 
a \code{\link[base]{data.frame}} such as retrieved by \code{\link[dtwSat]{getAlignments}}}

\item{from}{A character or \code{\link[base]{Dates}} object in the format "yyyy-mm-dd"}

\item{to}{A \code{\link[base]{character}} or \code{\link[base]{Dates}} object in the format "yyyy-mm-dd"}

\item{by}{A \code{\link[base]{character}} with the intevals size, \emph{e.g.} ''6 month''}

\item{breaks}{A vector of class \code{\link[base]{Dates}}}

\item{overlap}{A number between 0 and 1. The minimum overlapping 
between the one alignment and the interval of classification. 
Default is 1, \emph{i.e.} 100\%}

\item{threshold}{A number. The TWDTW threshold, i.e. the maximum TWDTW 
cost for consideration. Default is \code{Inf}}

\item{...}{additional arguments passed to \code{\link[dtwSat]{getPatternNames}}}
}
\value{
A \code{\link[base]{data.frame}} with the best alignment 
for each interval
}
\description{
This function retrieves the best alignment within each 
interval of classification based on the TWDTW distance
}
\examples{
weight.fun = logisticWeight(alpha=-0.1, beta=100, theta=0.5)
alig = twdtw(patterns=patterns.list, timeseries=template, weight.fun = weight.fun, 
        normalize=TRUE, patterns.length=23)
         
# Classify interval
from = as.Date("2009-09-01")
to = as.Date("2013-09-01")
by = "6 month"

# All classes
classifyIntervals(x=alig, from=from, to=to, by = by,
             overlap=.3, threshold=Inf)

# Cotton and Maize 
classifyIntervals(x=alig, from=from, to=to, by = by,
             overlap=.3, threshold=Inf, p.names=c("Cotton","Maize"))


}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

