% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{computeAccuracy}
\alias{computeAccuracy}
\title{Classification accuracy}
\usage{
computeAccuracy(predicted, reference)
}
\arguments{
\item{predicted}{A vector of prediction.}

\item{reference}{A vector of reference for the predicted classes.}
}
\description{
This function retrieves the accuracy of classification
based on the a given dataset of controll points.
}
\examples{
##
}

