% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_patterns.R
\name{create_patterns}
\alias{create_patterns}
\title{Create a Pattern Using GAM}
\usage{
create_patterns(
  x,
  y,
  formula,
  start_column = "start_date",
  end_column = "end_date",
  label_colum = "label",
  sampling_freq = NULL,
  ...
)
}
\arguments{
\item{x}{A stars object representing spatial-temporal data.}

\item{y}{An sf object representing spatial data with associated attributes.}

\item{formula}{A formula for the GAM.}

\item{start_column}{Name of the column in y that indicates the start date. Default is 'start_date'.}

\item{end_column}{Name of the column in y that indicates the end date. Default is 'end_date'.}

\item{label_colum}{Name of the column in y that contains labels. Default is 'label'.}

\item{sampling_freq}{The time frequency for sampling. If NULL, the function will infer it.}

\item{...}{Additional arguments passed to the GAM function.}
}
\value{
A list containing the predicted values for each label.
}
\description{
This function creates a pattern based on Generalized Additive Models (GAM).
It uses the specified formula to fit the model and predict values.
It operates on stars objects (for spatial-temporal data) and sf objects (for spatial data).
}
