% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPattern.R
\docType{methods}
\name{createPattern}
\alias{createPattern}
\title{Create temporal patterns}
\usage{
createPattern(x, from, to, freq = 1, attr, formula, ...)
}
\arguments{
\item{x}{A \code{\link[base]{list}} of \code{\link[zoo]{zoo}} such as 
retrived by \code{\link[dtwSat]{extractTimeSeries}}.}

\item{from}{A character or \code{\link[base]{Dates}} object in the format 
"yyyy-mm-dd". If not informed it is equal to the smallest date of the 
first element in x. See details.}

\item{to}{A \code{\link[base]{character}} or \code{\link[base]{Dates}} 
object in the format "yyyy-mm-dd". If not informed it is equal to the 
greatest date of the first element in x. See details.}

\item{freq}{An integer. The sampling frequency of the output patterns.}

\item{attr}{A vector character or numeric. The attributes in \code{x} to be used. 
If not declared the function uses all attributes.}

\item{formula}{A formula. Argument to pass to \code{\link[mgcv]{gam}}.}

\item{...}{other arguments to pass to the function \code{\link[mgcv]{gam}} in the 
packege \pkg{mgcv}.}
}
\value{
A \code{\link[zoo]{zoo}} object
}
\description{
This function fits a Generalized Additive Model model 
to the samples and retrieves a smoothed temporal pattern.
}
\details{
The hidden assumption is that the temporal pattern is a cycle the repeats itself 
within a given time interval. Therefore, all time series samples in \code{x} are aligned 
to each other keeping the time sequence of each sample and its respective day of the year. 
After that the function fits a Generalized Additive Model to the samples.
}
\examples{

field_samples = read.csv(system.file("lucc_MT/data/samples.csv", package = "dtwSat"))
proj_str = scan(system.file("lucc_MT/data/samples_projection", package = "dtwSat"), 
               what = "character")
load(system.file("lucc_MT/field_samples_ts.RData", package = "dtwSat"))

# Sellect Soybean-cotton samples  
I = field_samples$class=="Soybean-cotton"
soybean_cotton_samples = field_samples_ts[I]

p = createPattern(x = soybean_cotton_samples, 
                   formula = y ~ s(time, bs = "cc"))

plotPatterns(p)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

