% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\name{dtwSat-class}
\alias{dtwSat-class}
\title{dtwSat-class}
\description{
Class for Multidimensional Time-Weighted DTW alignments
}
\section{Slots }{

\describe{
 \item{\code{call}:}{An object of class \code{\link[base]{call}}, see 
 \code{\link[base]{match.call}}.}
 \item{\code{alignments}:}{A named \code{\link[base]{list}} whose elements 
 have length identical to the number of alignments.
 The elements are:  
      \cr\code{query}: query identification,
      \cr\code{from}: starting dates,
      \cr\code{to}: ending dates, and
	     \cr\code{distance}: TWDTW distances.
      }
 \item{\code{mapping}:}{An object of class \code{\link[base]{list}} whose 
 elements have the matching points for each alignment between 
 the query and the time series. 
 Each element has two vectors: 
      \cr\code{index1}: matching points of the query, and
      \cr\code{index2}: matching points of the time series.
      }
	\item{\code{internals}:}{An object of class \code{\link[base]{list}} whose 
  elements have the internal structures used by \code{\link[dtwSat]{twdtw}}. 
  The elements are: 
      \cr\code{timeWeight}: time weight matrix,
      \cr\code{localMatrix}: local cost matrix,
      \cr\code{costMatrix}: cumulative cost matrix,
      \cr\code{directionMatrix}: directions of steps that would be taken in the alignments,
      \cr\code{stepPattern}: \code{\link[dtw]{stepPattern}} used for the 
      computation, see package \code{\link[dtw]{dtw}}
      \cr\code{query}: query time series, 
      \cr\code{timeseries}: satellite image time series,
      \cr\code{N}: \code{query} length, and 
      \cr\code{M}: \code{timeseries} length.
      }
}
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{dtwSat}}, \code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{plotPath}}, \code{\link[dtwSat]{plotAlignment}},
\code{\link[dtw]{stepPattern}}, and \code{\link[dtw]{dtw}}.
}

