% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\name{dtwSat-class}
\alias{dtwSat-class}
\title{dtwSat-class}
\description{
Class for Multidimensional Time-Weighted DTW alignments
}
\section{Slots }{

\describe{
 \item{\code{call}:}{An object of class \code{\link[base]{call}}, see 
 \code{\link[base]{match.call}}.}
 \item{\code{alignments}:}{A named \code{\link[base]{list}} whose elements 
 have length identical to the number of alignments.
 The elements are:  
      \cr\code{pattern}: pattern identification,
      \cr\code{from}: starting date,
      \cr\code{to}: ending date, 
	     \cr\code{distance}: TWDTW distance, and
	     \cr\code{K}: the number of alignments of the pattern.
      }
	\item{\code{internals}:}{An object of class \code{\link[base]{list}} whose 
  elements have the internal structures used by \code{\link[dtwSat]{twdtw}}. 
  The elements are: 
      \cr\code{costMatrix}: cumulative cost matrix,
      \cr\code{directionMatrix}: directions of steps that would be taken in the alignments,
      \cr\code{stepPattern}: \code{\link[dtw]{stepPattern}} used for the 
      computation, see package \code{\link[dtw]{dtw}}
      \cr\code{N}: \code{pattern} length, 
      \cr\code{M}: \code{timeseries} length, 
      \cr\code{timeWeight}: time weight matrix,
      \cr\code{localMatrix}: local cost matrix,
      \cr\code{patterns}: temporal pattern, 
      \cr\code{timeseries}: satellite image time series,
      }
	\item{\code{matching}:}{An object of class \code{\link[base]{list}} whose 
  elements have the matching points for each alignment between the 
  \code{pattern} and the \code{timeseries}, such that
      \cr\code{index1}: matching points of the temporal pattern, and
      \cr\code{index2}: matching points of the time series.
      }
}
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{getPatternNames}},
\code{\link[dtwSat]{getAlignments}},
\code{\link[dtwSat]{getMatches}}, and 
\code{\link[dtwSat]{getInternals}}
}

