% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\docType{methods}
\name{dtwSat}
\alias{dtwSat}
\title{Multidimensional Time-Weighted DTW Alignment}
\usage{
dtwSat(query, template, ...)
}
\arguments{
\item{query}{A \link[zoo]{zoo} object with the multidimensional time series.}

\item{template}{A \link[zoo]{zoo} object with the template time series similar
to \code{query}. The \code{template} must have the same number of attributes
and be equal or longer than the \code{query},
\emph{i.e.} \code{nrow(query)<=nrow(template)}.}

\item{...}{see \code{\link[dtwSat]{twdtw}}}
}
\value{
object of class \code{\link[dtwSat]{dtwSat-class}}
}
\description{
This function performs a multidimensional Time-Weighted DTW
analysis and retrieves one or more possible alignments of a query within
a time series.
}
\examples{
names(query.list)
alig = dtwSat(query.list[["Soybean"]], template, weight = "logistic",
       alpha = 0.1, beta = 50, alignments=4)
alig
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw}}, \code{\link[dtwSat]{mtwdtw}}
}

