% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\docType{methods}
\name{dtwSat}
\alias{dtwSat}
\alias{dtwSat-class}
\title{dtwSat-class}
\usage{
dtwSat(query, template, ...)

dtwSat(query, template, ...)
}
\arguments{
\item{query}{A \link[zoo]{zoo} object with the multidimensional time series.}

\item{template}{A \link[zoo]{zoo} object with the template time series. The index of
the zoo object must be of class \code{\link[base]{Date}}.
It must be iguel or be equal or longer than the length of the query and
the same number of dimensions. The index of the zoo object must be of
class \code{\link[base]{Date}}.}

\item{...}{see \code{\link[dtwSat]{twdtw}}}
}
\value{
object of class \code{\link[dtwSat]{dtwSat}}
}
\description{
Class for Multidimensional Time-Weighted DTW

This function performs a multidimensional Time-Weighted DTW
analysis and retrieves one or more possible alignments of a query within
a time series.
}
\section{Slots }{

\describe{
	\item{\code{call}:}{Object of class \code{call} see
	\code{\link[base]{match.call}}.}
	\item{\code{alignments}:}{Object of class \code{list} where each node of
	the list has one attribute, whose lengths are identical to the number of
	alignments. The attributes are:}
	  \item{\code{from}}{ the alignment starts, }
	  \item{\code{to}}{ the alignment ends, }
	  \item{\code{distance}}{ the DTW distance, and}
	  \item{\code{normalizedDistance}}{ the normalized DTW distance.}
	\item{\code{mapping}:}{ Object of class \code{list}. Each node of the list has
	the matching points of the query to the template time series.}
	\item{\code{internals}:}{ Object of class \code{list} similar to
	\code{\link[dtw]{dtw}} class from package \pkg{dtw}.}
}
}
\examples{
names(query.list)
alig = dtwSat(query.list[["Soybean"]], template, weight = "logistic",
       alpha = 0.1, beta = 50, alignments=4)
alig
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}

Victor Maus, \email{vwmaus1@gmail.com}
}

