% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\docType{methods}
\name{extractSampleTimeSeries}
\alias{extractSampleTimeSeries}
\title{Extract time series from raster}
\usage{
extractSampleTimeSeries(x, proj4string = NA, raster.list, timeline, doy,
  year = 2000, nsamples = 10, mc.cores = 1)
}
\arguments{
\item{x}{A \code{\link[base]{data.frame}} informing: longitude, 
latitude, from, to, and classname, for each sample. This can also be 
a \code{\link[sp]{SpatialPointsDataFrame}} informing: from, to, and 
classname.}

\item{proj4string}{projection string, see \code{\link[sp]{CRS-class}}}

\item{raster.list}{A list of \code{\link[raster]{RasterBrick-class}} or 
\code{\link[raster]{RasterStack-class}}.}

\item{timeline}{A vector of class \code{\link[base]{Dates}}. 
It must have the length of the layers in the \code{\link[raster]{RasterBrick-class}} objects}

\item{doy}{A \code{\link[raster]{RasterBrick-class}} or \code{\link[raster]{RasterStack-class}} 
with the same extent as the objects in \code{raster.list}}

\item{year}{An integer. An base year to set all time series of the class to the same period}

\item{nsamples}{An integer. The number of samples for each class. Default extracts 
all samples given bx \code{x}}

\item{mc.cores}{The number of cores to use, See \code{\link[parallel]{mclapply}} 
for details}
}
\value{
A list of \code{\link[base]{data.frame}} objects, one for each class in \code{x}
}
\description{
This function extracts the time series from a list of 
raster given set of spatial location
}
\examples{

###


}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

