% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTimeSeries.R
\docType{methods}
\name{extractTimeSeries}
\alias{extractTimeSeries}
\title{Extract time series from raster}
\usage{
extractTimeSeries(x, y, proj4string = CRS(as.character(NA)), mc.cores = 1)
}
\arguments{
\item{x}{A list of Raster* objects \code{\link[raster]{Raster-class}}. To build the 
raster list use \code{\link[dtwSat]{buildRasterTimeSeries}}.}

\item{y}{A \code{\link[base]{data.frame}} whose attributes are: longitude, 
latitude, the start ''from'' and the end ''to'' of the time interval 
for each sample. This can also be a \code{\link[sp]{SpatialPointsDataFrame}} 
whose attributes are the start ''from'' and the end ''to'' of the time interval.}

\item{proj4string}{projection string, see \code{\link[sp]{CRS-class}}. Used 
if \code{y} is a \code{\link[base]{data.frame}}.}

\item{mc.cores}{The number of cores to use, see \code{\link[parallel]{mclapply}} 
for details.}
}
\value{
A list of \code{\link[zoo]{zoo}} objects, one for each location given by \code{y}.
}
\description{
This function extracts time series from a list of 
raster using the spatial location and the time period.
}
\examples{

\dontrun{
##### In the next example we build, extract and plot a multi-band time series. 

#### In this example we build a multi-band MOD13Q1 raster time series. 
#### The 'tif' files in 'lucc_MT/tif' have 999 EVI time series 
#### from 2007-01-01 to 2013-12-19, that means 160 points with temporal 
#### resolution of 16 days. 
data_folder = system.file("lucc_MT/data", package = "dtwSat")
blue = brick(paste(data_folder,"blue.tif", sep = "/"))
red  = brick(paste(data_folder,"red.tif", sep = "/"))
nir  = brick(paste(data_folder,"nir.tif", sep = "/"))
mir  = brick(paste(data_folder,"mir.tif", sep = "/"))
evi  = brick(paste(data_folder,"evi.tif", sep = "/"))
ndvi = brick(paste(data_folder,"ndvi.tif", sep = "/"))
dates = scan(paste(data_folder,"timeline", sep = "/"), what = "dates")
raster_timeseries = 
     buildRasterTimeSeries(blue, red, nir, mir, evi, ndvi, timeline = dates)
 
#### In the next example we extract and plot a multi-band time series from 
#### our raster time series built in the example above. 

## Location and time range 
ts_location = data.frame(longitude = -55.96957, latitude = -12.03864, 
                         from = "2007-09-01", to = "2013-09-01")
 
## Proj string 
crs_string = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"

## Extract time series 
ts = extractTimeSeries(x = raster_timeseries, y = ts_location, 
                       proj4string = crs_string)
 
autoplot(ts[[1]], facets = NULL) + xlab("Time") + ylab("Value")

}

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

