% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{get_stars_time_freq}
\alias{get_stars_time_freq}
\title{Compute the Most Common Sampling Frequency in a Stars Object}
\usage{
get_stars_time_freq(x)
}
\arguments{
\item{x}{A stars object containing time series data.}
}
\value{
A difftime object representing the most common time difference between consecutive samples.
}
\description{
This function calculates the most common difference between consecutive time points in a stars object.
This can be useful for determining the sampling frequency of the time series data.
}
