% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignments.R
\docType{methods}
\name{getAlignments}
\alias{getAlignments}
\title{Get alignments from twdtw object}
\usage{
getAlignments(object, p.names)
}
\arguments{
\item{object}{A \link[dtwSat]{twdtw-class} object.}

\item{p.names}{A \link[base]{character} or \link[base]{numeric}
vector with the patterns identification. If not declared the function 
retrieves the alignments for all patterns.}
}
\value{
A \link[base]{data.frame} with the following attributes:  
      \cr\code{pattern}: the pattern identification,
      \cr\code{from}: starting date,
      \cr\code{to}: ending date, and
	     \cr\code{distance}: TWDTW dissimilarity.
}
\description{
This function retrieves the alignments 
from the object \link[dtwSat]{twdtw-class}
}
\examples{

log_fun = logisticWeight(alpha=-0.1, beta=100)
matches = twdtw(x=example_ts, patterns=patterns.list, weight.fun = log_fun)

getAlignments(matches)

getAlignments(matches, p.names="Soybean")

getAlignments(matches, p.names=c(2,3))

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, and
\code{\link[dtwSat]{twdtw}}
}

