% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\docType{methods}
\name{getAlignments}
\alias{getAlignments}
\title{Get alignments from dtwSat object}
\usage{
getAlignments(object)
}
\arguments{
\item{object}{A \link[dtwSat]{dtwSat} object}
}
\value{
An object of class data.frame
}
\description{
This function retrieves the results of the alignments in the object
\link[dtwSat]{dtwSat}
}
\examples{
names(query.list)
alig = dtwSat(query.list[["Soybean"]], template, weight = "logistic",
             alpha = 0.1, beta = 50, alignments=4)
getAlignments(alig)
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

