% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignments.R
\name{getAlignments}
\alias{getAlignments}
\title{Get TWDTW alignments}
\usage{
getAlignments(object)
}
\arguments{
\item{object}{an object of class twdtwTimeSeries.}

\item{...}{other arguments to be passed to methods}
}
\description{
Generic method to get the TWDTW alignments from objects of class twdtwMatches
}
\examples{
matches = new("twdtwMatches")
getAlignments(matches)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwMatches-class}}, 
\code{\link[dtwSat]{twdtwTimeSeries-class}}, and
\code{\link[dtwSat]{twdtwRaster-class}}
}

