% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\docType{methods}
\name{getAlignments}
\alias{getAlignments}
\title{Get alignments from twdtw object}
\usage{
getAlignments(object, ...)
}
\arguments{
\item{object}{A \link[dtwSat]{twdtw-class} object}

\item{...}{additional arguments passed to \code{\link[dtwSat]{getPatternNames}}}
}
\value{
A \link[base]{data.frame} with the following columns:  
      \cr\code{pattern}: the pattern identification,
      \cr\code{from}: starting date,
      \cr\code{to}: ending date, and
	     \cr\code{distance}: TWDTW distances.
}
\description{
This function retrieves the alignments 
from the object \link[dtwSat]{twdtw-class}
}
\examples{

weight.fun = logisticWeight(alpha=-0.1, beta=100, theta=0.5)
alig = twdtw(x=template, patterns=patterns.list, weight.fun = weight.fun)

getAlignments(alig)

getAlignments(alig, p.names="Soybean")

getAlignments(alig, p.names=c(2,3))

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, and
\code{\link[dtwSat]{twdtw}}
}

