% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\docType{methods}
\name{getAlignments}
\alias{getAlignments}
\title{Get alignments from dtwSat object}
\usage{
getAlignments(object)
}
\arguments{
\item{object}{A \link[dtwSat]{dtwSat-class} object}
}
\value{
An object of class \link[base]{data.frame}
}
\description{
This function retrieves the alignments
in the object \link[dtwSat]{dtwSat-class}
}
\examples{
names(query.list)
alig = dtwSat(query.list[["Soybean"]], template)
getAlignments(alig)
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{dtwSat}} and \code{\link[dtwSat]{twdtw}}
}

