% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{getDatesFromDOY}
\alias{getDatesFromDOY}
\title{Get dates from year and day of the year}
\usage{
getDatesFromDOY(year, doy)
}
\arguments{
\item{year}{An vector with the years}

\item{doy}{An vector with the day of the year}
}
\value{
object of class \code{\link[base]{Date}}
}
\description{
This function retrieves the date corresponding to year
and day of the year
}
\examples{
year = c(2000, 2001)
doy = c(366, 365)
dates = getDatesFromDOY(year, doy)
dates
}
\seealso{
\link[dtwSat]{createTimeSequence} and \link[dtwSat]{getModisTimeIndex}
}

