% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\docType{methods}
\name{getInternals}
\alias{getInternals}
\title{Get internals from twdtw object}
\usage{
getInternals(object, ...)
}
\arguments{
\item{object}{A \link[dtwSat]{twdtw-class} object}

\item{...}{additional arguments passed to \code{\link[dtwSat]{getPatternNames}}}
}
\value{
A \code{\link[base]{list}} whose 
  elements have the internal structures used used in \code{\link[dtwSat]{twdtw}}. 
  The elements are: 
      \cr\code{timeWeight}: time weight matrix,
      \cr\code{localMatrix}: local cost matrix,
      \cr\code{costMatrix}: cumulative cost matrix,
      \cr\code{directionMatrix}: directions of steps that would be taken in the alignments,
      \cr\code{stepPattern}: \code{\link[dtw]{stepPattern}} used for the 
      computation, see package \code{\link[dtw]{dtw}}
      \cr\code{pattern}: pattern time series, 
      \cr\code{x}: satellite image time series,
      \cr\code{N}: \code{pattern} length, and 
      \cr\code{M}: \code{x} length.
}
\description{
This function retrieves cost matrix, inputs, and other 
internal structures from \link[dtwSat]{twdtw-class} object
}
\examples{

weight.fun = logisticWeight(alpha=-0.1, beta=100, theta=0.5)
alig = twdtw(x=template, patterns=patterns.list, weight.fun = weight.fun, keep=TRUE)

a = getInternals(alig)
names(a) 

a = getInternals(alig, p.names="Maize")
names(a) 

a = getInternals(alig, p.names=c(1,2))
names(a) 


}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, and
\code{\link[dtwSat]{twdtw}}
}

