% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInternals.R
\docType{methods}
\name{getInternals}
\alias{getInternals}
\title{Get internals from twdtw object}
\usage{
getInternals(object, y)
}
\arguments{
\item{object}{A \link[dtwSat]{twdtw-class} object.}

\item{y}{A \link[base]{character} or \link[base]{numeric}
vector with the patterns identification. If not declared the function 
retrieves the internals for all patterns.}
}
\value{
A \code{\link[base]{list}} whose 
  elements have the internal structures used used in \code{\link[dtwSat]{twdtw}}. 
  The elements are: 
      \cr\code{timeWeight}: time weight matrix,
      \cr\code{localMatrix}: local cost matrix,
      \cr\code{costMatrix}: cumulative cost matrix,
      \cr\code{directionMatrix}: directions of steps that would be taken in the alignments,
      \cr\code{startingMatrix}: the index of the first point that contributed for the cost computation,
      \cr\code{stepPattern}: \code{\link[dtw]{stepPattern}} used for the 
      computation, see package \code{\link[dtw]{dtw}},
      \cr\code{pattern}: pattern time series, 
      \cr\code{x}: satellite image time series,
      \cr\code{N}: \code{pattern} length, and 
      \cr\code{M}: \code{x} length.
}
\description{
This function retrieves cost matrix, inputs, and other 
internal structures from \link[dtwSat]{twdtw}.
}
\examples{

log_fun = logisticWeight(alpha=-0.1, beta=100)
matches = twdtw(x=example_ts, patterns=patterns.list, 
             weight.fun = log_fun, keep=TRUE)

a = getInternals(matches)
names(a) 

a = getInternals(matches, y="Maize")
names(a) 

a = getInternals(matches, y=c(1,2))
names(a) 


}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, and
\code{\link[dtwSat]{twdtw}}.
}

