% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatches.R
\docType{methods}
\name{getMatches}
\alias{getMatches}
\title{Get matching points from twdtw object}
\usage{
getMatches(object, p.names)
}
\arguments{
\item{object}{A \link[dtwSat]{twdtw-class} object.}

\item{p.names}{A \link[base]{character} or \link[base]{numeric}
vector with the patterns identification. If not declared the function 
retrieves the matching points for all patterns.}
}
\value{
A \code{\link[base]{list}} whose 
 elements have the information for each pattern. 
 Each pattern has then a sublist of matching points for each alignment, such as: 
      \cr\code{index1}: matching points of the pattern, and
      \cr\code{index2}: matching points of the time series.
}
\description{
This function retrieves the matching points 
for each alignment between the \code{pattern} and the 
time series \code{x}.
}
\examples{

log_fun = logisticWeight(alpha=-0.1, beta=100)
matches = twdtw(x=example_ts, patterns=patterns.list, weight.fun = log_fun, keep=TRUE)

getMatches(matches)

getMatches(matches, p.names="Maize")

getMatches(matches, p.names=1)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, and
\code{\link[dtwSat]{twdtw}}.
}

