% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\docType{methods}
\name{getMatches}
\alias{getMatches}
\title{Get matching points from dtwSat object}
\usage{
getMatches(object, ...)
}
\arguments{
\item{object}{A \link[dtwSat]{dtwSat-class} object}

\item{...}{additional arguments passed to \code{\link[dtwSat]{getPatternNames}}}
}
\value{
A \code{\link[base]{list}} whose 
 elements have the matching points for each alignment between 
 the temporal pattern and the time series. 
 Each element has two vectors: 
      \cr\code{index1}: matching points of the pattern, and
      \cr\code{index2}: matching points of the time series.
}
\description{
This function retrieves the matching points 
for each alignment between the \code{pattern} and the 
\code{timeseries}
}
\examples{

weight.fun = logisticWeight(alpha=-0.1, beta=100, theta=0.5)
alig = twdtw(patterns=patterns.list, timeseries=template, weight.fun = weight.fun)

getMatches(alig)

getMatches(alig, p.names="Maize")

getMatches(alig, p.names=1)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{dtwSat-class}}, and
\code{\link[dtwSat]{twdtw}}
}

