% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\docType{methods}
\name{getMatches}
\alias{getMatches}
\title{Get matching points from dtwSat object}
\usage{
getMatches(object)
}
\arguments{
\item{object}{A \link[dtwSat]{dtwSat} object}
}
\value{
object of class data.frame
}
\description{
This function retrieves the matching points from the alignments
in the object \link[dtwSat]{dtwSat}.
}
\examples{
names(query.list)
alig = dtwSat(query.list[["Soybean"]], template, weight = "logistic", alpha = 0.1, beta = 50, alignments=4)
getMatches(alig)
}

