% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPatternNames.R
\docType{methods}
\name{getPatternNames}
\alias{getPatternNames}
\title{Get pattern names from twdtw object}
\usage{
getPatternNames(object, y)
}
\arguments{
\item{object}{A \link[dtwSat]{twdtw-class} object.}

\item{y}{A \link[base]{character} or \link[base]{numeric}
vector with the patterns identification. If not declared the function 
retrieves the names for all patterns.}
}
\value{
A \code{\link[base]{character}}
or \code{\link[base]{numeric}} vector.
}
\description{
This function retrieves pattern names in the 
\link[dtwSat]{twdtw-class} object.
}
\examples{

matches = twdtw(x=example_ts, patterns=patterns.list)

getPatternNames(matches)

getPatternNames(matches, y=c(1,3))

getPatternNames(matches, y="Maize")

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, and
\code{\link[dtwSat]{twdtw}}.
}

