% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPatterns.R
\docType{methods}
\name{getPatterns}
\alias{getPatterns}
\alias{getPatterns,twdtwMatches-method}
\alias{getPatterns-twdtwMatches}
\title{Get patterns}
\usage{
getPatterns(object, labels = NULL)

\S4method{getPatterns}{twdtwMatches}(object, labels = NULL)
}
\arguments{
\item{object}{an twdtwMatches object.}

\item{labels}{a vector with the patterns labels. If not informed the 
function retrieves all patterns.}
}
\value{
a list of class \code{\link[zoo]{zoo}}
}
\description{
Generic method to get temporal patterns from objects of class twdtwMatches.
}
\examples{
# Get patterns from objects of class twdtwTimeSeries
patterns = twdtwTimeSeries(timeseries=patterns.list, labels=names(patterns.list))
ts = twdtwTimeSeries(timeseries=example_ts.list)
matches = twdtwApply(x=ts, y=patterns)
getPatterns(matches)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwMatches-class}}, 
\code{\link[dtwSat]{twdtwTimeSeries-class}}, and 
\code{\link[dtwSat]{getTimeSeries}}
}

