% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPatterns.R
\name{getPatterns}
\alias{getPatterns}
\title{Get patterns}
\usage{
getPatterns(object, labels = NULL)
}
\arguments{
\item{object}{an twdtwMatches object.}

\item{labels}{a vector with the patterns labels. If not informed the 
function retrieves all patterns.}
}
\description{
Generic method to get temporal patterns from objects of class twdtwMatches.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwMatches-class}}, 
\code{\link[dtwSat]{twdtwTimeSeries-class}}, and
\code{\link[dtwSat]{twdtwRaster-class}}
}

