% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeSeries.R
\docType{methods}
\name{getTimeSeries}
\alias{getTimeSeries}
\title{Get imput time series from twdtw object}
\usage{
getTimeSeries(object)
}
\arguments{
\item{object}{A \link[dtwSat]{twdtw-class} object.}
}
\value{
The satellite time series, a \link[zoo]{zoo} object.
}
\description{
This function retrieves time series used in the
TWDTW analysis.
}
\examples{

log_fun = logisticWeight(alpha=-0.1, beta=100)
matches = twdtw(x=example_ts, patterns=patterns.list, 
             weight.fun = log_fun)

getTimeSeries(matches)


}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, and
\code{\link[dtwSat]{twdtw}}.
}

