% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iquery.R
\name{iqueryApply}
\alias{iqueryApply}
\title{iquery builder}
\usage{
iqueryApply(str_iquery, logfile = NULL)
}
\arguments{
\item{str_iquery}{A list of iquery strings to be executed in SciBD.}

\item{logfile}{The path for the log file.}
}
\description{
This function runs the iquery string on SciDB.
}
\examples{
###

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{iqueryBlock}
}

