% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinAlignments.R
\name{joinAlignments}
\alias{joinAlignments}
\title{Joint TWDTW alignments}
\usage{
joinAlignments(...)
}
\arguments{
\item{...}{objects of class twdtwMatches}
}
\description{
Join TWDTW alignments from two or more objects of class twdtwMatches.
}
\examples{
## 

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwMatches-class}}, 
\code{\link[dtwSat]{twdtwTimeSeries-class}}, and
\code{\link[dtwSat]{twdtwRaster-class}}
}

