% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinTimeSeries.R
\docType{methods}
\name{joinTimeSeries}
\alias{joinTimeSeries}
\alias{joinTimeSeries,ANY-method}
\alias{joinTimeSeries-twdtwTimeSeries}
\title{Join time series}
\usage{
joinTimeSeries(...)

\S4method{joinTimeSeries}{ANY}(..., join.labels = TRUE)
}
\arguments{
\item{...}{objects of class twdtwTimeSeries}

\item{join.labels}{a logical. It TRUE the function joins labels that are identical 
to a factor. If FALSE a different label is kept for each samples.}
}
\value{
an object of class \code{\link[dtwSat]{twdtwTimeSeries}}
}
\description{
Join time series from two or more objects of class twdtwTimeSeries.
}
\examples{
## 

# Join time series from objects of class twdtwTimeSeries
ts1 = twdtwTimeSeries(timeseries=example_ts.list[1], labels="A")
ts2 = twdtwTimeSeries(timeseries=example_ts.list[2], labels="B")
joinTimeSeries(ts1, ts2, join.labels=TRUE)
joinTimeSeries(ts1, ts2, join.labels=FALSE)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwTimeSeries-class}}, and
\code{\link[dtwSat]{joinAlignments}}
}

