% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticWeight.R
\docType{methods}
\name{logisticWeight}
\alias{logisticWeight}
\title{Logistic weight function}
\usage{
logisticWeight(alpha, beta)
}
\arguments{
\item{alpha}{numeric. The steepness of logistic weight}

\item{beta}{numeric. The midpoint of logistic weight}
}
\value{
An \code{\link[base]{function}} object
}
\description{
Builds a logistic time weight 
function to compute the TWDTW local cost matrix [1]
}
\examples{
log_fun = logisticWeight(alpha=-0.1, beta=100)
log_fun

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\references{
[1] Maus  V,  C\^{a}mara  G,  Cartaxo  R,  Sanchez  A,  Ramos  FM,  de Queiroz, GR.
(2016). A Time-Weighted Dynamic Time Warping method for land use and land cover 
mapping. Selected Topics in Applied Earth Observations and Remote Sensing, 
IEEE Journal of, X, XX-XX.
}
\seealso{
\code{\link[dtwSat]{twdtw}}
}

