% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twdtw.R
\docType{methods}
\name{logisticWeight}
\alias{logisticWeight}
\title{Logistic weight function}
\usage{
logisticWeight(alpha, beta)
}
\arguments{
\item{alpha}{numeric. The steepness of logistic weight}

\item{beta}{numeric. The midpoint of logistic weight}
}
\value{
An \code{\link[base]{expression}} object
}
\description{
This function builds a logistic time weight 
function
}
\examples{
weight.fun = logisticWeight(alpha=-0.1, beta=100)
weight.fun

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw}}
}

