% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{modisColRowToLongLat}
\alias{modisColRowToLongLat}
\title{MODIS column and row to longitude and latitude.}
\usage{
modisColRowToLongLat(col, row, h = NULL, v = NULL, pixelsize,
  projCRS = NULL)
}
\arguments{
\item{col}{An integer. A global index of the MODIS column
or a local index of the column within a specific h:v tile}

\item{row}{An integer. A global index of the MODIS row
or a local index of the row within a specific h:v tile}

\item{h}{An integer. Horizontal MODIS tile}

\item{v}{An integer. Vertical MODIS tile}

\item{pixelsize}{A number. The MODIS resolution.}

\item{projCRS}{A CRS object.}
}
\description{
This function computes the longitude and latitude for a
MODIS column and row index.
}

