% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{mtwdtw}
\alias{mtwdtw}
\title{Performs multiple Time-Weighted DTW}
\usage{
mtwdtw(query, template, ...)
}
\arguments{
\item{query}{A list of zoo objects with the multidimensional
time series}

\item{template}{A zoo object with the template time series.
It must be iguel or be equal or longer than the length of the query and
the same number of dimensions}

\item{...}{See \code{\link[dtwSat]{twdtw}}}
}
\description{
The function performs the Time-Weighted DTW for a list
of queries
}
\examples{
# alig = mtwdtw(query.list, template, weight = "logistic", alpha = 0.1, beta = 50)
# alig
}

