% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twdtw.R
\docType{methods}
\name{mtwdtw}
\alias{mtwdtw}
\title{Performs multiple Time-Weighted DTW}
\usage{
mtwdtw(query, timeseries = NULL, template = NULL, normalize = FALSE,
  query.length = NULL, ...)
}
\arguments{
\item{query}{A \link[zoo]{zoo} object with a query time series.}

\item{timeseries}{A \link[zoo]{zoo} object with a time series similar
to \code{query}. \code{timeseries} must have the same number of attributes
and be equal to or longer than the \code{query},
\emph{i.e.} \code{nrow(query)<=nrow(timeseries)}.}

\item{template}{is deprecated, please use \code{timeseries} instead}

\item{normalize}{Normalize queries length. Default is FALSE}

\item{query.length}{An integer. Queries length used with normalize. If not declared
the length of the output queries will be the length of the longest query}

\item{...}{additional arguments passed to \code{\link[dtwSat]{twdtw}}}
}
\value{
An object of class \link[dtwSat]{dtwSat-class} without \code{internals}
or \code{mapping}
}
\description{
This function performs the Time-Weighted DTW for a list
of queries
}
\examples{
alig = mtwdtw(query.list, timeseries=template, weight = "logistic",
       alpha = 0.1, beta = 100)
alig
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw}}, \code{\link[dtwSat]{dtwSat-class}}
}

