% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/twdtw.R
\docType{methods}
\name{mtwdtw}
\alias{mtwdtw}
\title{Performs multiple Time-Weighted DTW}
\usage{
mtwdtw(query, template, ...)
}
\arguments{
\item{query}{A zoo object with the multidimensional time series.}

\item{template}{A zoo object with the template time series. The index of
the zoo object must be of class \code{\link[base]{Date}}.
It must be iguel or be equal or longer than the length of the query and
the same number of dimensions. The index of the zoo object must be of
class \code{\link[base]{Date}}.}

\item{...}{see \code{\link[dtwSat]{dtwSat}}}
}
\value{
obaject of class \code{\link[base]{data.frame}} see with attributes
from the slot alignment of the objeact \code{\link[dtwSat]{dtwSat}}
}
\description{
The function performs the Time-Weighted DTW for a list
of queries
}
\examples{
alig = mtwdtw(query.list, template, weight = "logistic",
       alpha = 0.1, beta = 50)
alig
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

