% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizePatterns.R
\name{normalizePatterns}
\alias{normalizePatterns}
\title{Normalize patterns}
\usage{
normalizePatterns(object, length = NULL)
}
\arguments{
\item{object}{an object of class twdtwTimeSeries.}

\item{length}{a number. If not informed the function will use the 
length of the longest pattern.}
}
\description{
Normalize the length of the temporal patterns. 
This function removes patterns (time series) with duplicated label.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwMatches-class}}, 
\code{\link[dtwSat]{twdtwTimeSeries-class}}, and
\code{\link[dtwSat]{twdtwRaster-class}}
}

