% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtwSatClass.R
\docType{methods}
\name{plot,dtwSat-method}
\alias{plot,dtwSat-method}
\title{Plotting Time-Weighted DTW}
\usage{
\S4method{plot}{dtwSat}(x, type = "path", ...)
}
\arguments{
\item{x}{\code{\link[dtwSat]{twdtw}} object}

\item{type}{A character for the plot type, "path" or "alignment".
Default is "path"}

\item{...}{additional arguments passed to plotting functions
\code{\link[dtwSat]{plotPath}} and \code{\link[dtwSat]{plotAlignment}}}
}
\description{
Plotting Time-Weighted DTW
}

