% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtwSat.R
\docType{methods}
\name{plot.twdtw}
\alias{plot.twdtw}
\title{Plotting Time-Weighted DTW}
\usage{
\method{plot}{twdtw}(x, ylab = "Pattern", xlab = "Time series",
  show.dist = FALSE, ...)
}
\arguments{
\item{x}{\code{\link[dtwSat]{twdtw}} object}

\item{ylab}{A character, y axis label}

\item{xlab}{A character, x axis label}

\item{show.dist}{Display dtw distance for each alignment}

\item{...}{additional arguments passed to plotting functions}
}
\description{
Methods for plotting Time-Weighted DTW objects
returned by twdtw.
}
\examples{
alig = twdtw(query.list[["Soybean"]], template, weight = "logistic", alpha = 0.1, beta = 50)
plot(alig)
}

