% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-dtwSat.R
\docType{methods}
\name{plotAlignment}
\alias{plotAlignment}
\title{Plotting Time-Weighted DTW alignment}
\usage{
plotAlignment(x, alignment, dimension, shift = 0.5)
}
\arguments{
\item{x}{\code{\link[dtwSat]{twdtw}} object}

\item{alignment}{An integer, the index of the alignment}

\item{dimension}{An integer or a character with the name of the
column in query and template use in the plot}

\item{shift}{A number, it shifts the pattern position.
Default is 0.5}
}
\value{
object of class \code{\link[ggplot2]{ggplot}}
}
\description{
Methods for plotting Time-Weighted DTW objects
returned by dtwSat.
}
\examples{
names(query.list)
alig = dtwSat(query.list[["Soybean"]], template, weight = "logistic",
       alpha = 0.1, beta = 100, alignments = 4, keep = TRUE)
gp = plotAlignment(alig, dimension="evi", alignment=1, shift=0.5)
gp
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

