% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignment.R
\docType{methods}
\name{plotAlignment}
\alias{plotAlignment}
\title{Plotting twdtw alignments}
\usage{
plotAlignment(x, p.names, attr = 1, threshold = Inf)
}
\arguments{
\item{x}{An \code{\link[dtwSat]{twdtw-class}} object}

\item{p.names}{A \link[base]{character} or \link[base]{integer}
vector with the patterns identification. If not declared the function 
will plot alignments for all patterna in the 
\code{\link[dtwSat]{twdtw-class}} object}

\item{attr}{An \link[base]{integer} vector or \link[base]{character} vector 
indicating the attribute for plotting, \emph{i.e.} a column of the \code{pattern}. 
Default is 1}

\item{threshold}{A number. The TWDTW threshold, i.e. the maximum TWDTW 
cost for consideration. Default is \code{Inf}}
}
\value{
A \link[ggplot2]{ggplot} object
}
\description{
Method for plotting the alignments and respective TWDTW 
distances
}
\examples{

weight.fun = logisticWeight(alpha=-0.1, beta=100, theta=0.5)
alig = twdtw(x=template, patterns=patterns.list, weight.fun = weight.fun, 
        normalize.patterns=TRUE, patterns.length=23, keep=TRUE)

gp = plotAlignment(alig, attr=c("ndvi","evi"))
gp

gp = plotAlignment(alig, attr=c("ndvi","evi"), threshold=4)
gp

gp = plotAlignment(x=alig, p.names="Cotton", 
                   attr=c("ndvi","evi"), threshold=6)
gp

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, 
\code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{plotPath}}, 
\code{\link[dtwSat]{plotCostMatrix}},
\code{\link[dtwSat]{plotMatch}}, and
\code{\link[dtwSat]{plotGroup}}
}

