% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCostMatrix.R
\name{plotCostMatrix}
\alias{plotCostMatrix}
\title{Plotting cost matrices}
\usage{
plotCostMatrix(x, matrix.name = "costMatrix", p.names)
}
\arguments{
\item{x}{An \code{\link[dtwSat]{twdtw-class}} object}

\item{matrix.name}{A character. The name of the matrix to plot,
"costMatrix", "localMatrix", or "timeWeight". 
Default is "costMatrix"}

\item{p.names}{A \link[base]{character} or \link[base]{numeric}
vector with the patterns identification. If not declared the function 
will plot the matrices for all patterns}
}
\value{
A \link[ggplot2]{ggplot} object
}
\description{
Method for plotting the internal matrices
}
\examples{

weight.fun = logisticWeight(alpha=-0.1, beta=100)
alig = twdtw(x=template, patterns=patterns.list, weight.fun = weight.fun, 
        normalize.patterns=TRUE, patterns.length=23, keep=TRUE)

# Plot local cost for all patterns
gp1 = plotCostMatrix(x=alig, matrix.name="localMatrix")
gp1

# Plot time weight for all patterns
gp2 = plotCostMatrix(x=alig, matrix.name="timeWeight")
gp2

# Plot accumulated cost for all patterns
gp3 = plotCostMatrix(x=alig, matrix.name="costMatrix")
gp3

# Plot accumulated cost for Cotton and Soybean
gp4 = plotCostMatrix(x=alig, matrix.name="costMatrix", 
     p.names=c("Cotton","Soybean"))

gp4

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, 
\code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{plotPath}}, 
\code{\link[dtwSat]{plotAlignment}},
\code{\link[dtwSat]{plotMatch}}, and
\code{\link[dtwSat]{plotGroup}}
}

