% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCostMatrix.R
\name{plotCostMatrix}
\alias{plotCostMatrix}
\title{Plotting cost matrices}
\usage{
plotCostMatrix(x, matrix.name = "costMatrix", p.names)
}
\arguments{
\item{x}{An \code{\link[dtwSat]{twdtw-class}} object.}

\item{matrix.name}{A character. The name of the matrix to plot,
"costMatrix" for accumulated cost, "localMatrix" for local cost, 
or "timeWeight" for time-weight. 
Default is "costMatrix".}

\item{p.names}{A \link[base]{character} or \link[base]{numeric}
vector with the patterns identification. If not declared the function 
will plot the matrices for all patterns.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Method for plotting the internal matrices used during the 
TWDTW computation.
}
\examples{

log_fun = logisticWeight(alpha=-0.1, beta=100)
matches = twdtw(x=example_ts, patterns=patterns.list, weight.fun = log_fun, 
        normalize.patterns=TRUE, patterns.length=23, keep=TRUE)

# Plot local cost for all patterns
gp1 = plotCostMatrix(x=matches, matrix.name="localMatrix")
gp1

# Plot time weight for all patterns
gp2 = plotCostMatrix(x=matches, matrix.name="timeWeight")
gp2

# Plot accumulated cost for all patterns
gp3 = plotCostMatrix(x=matches, matrix.name="costMatrix")
gp3

# Plot accumulated cost for Cotton and Soybean
gp4 = plotCostMatrix(x=matches, matrix.name="costMatrix", 
     p.names=c("Cotton","Soybean"))

gp4

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, 
\code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{plotPaths}}, 
\code{\link[dtwSat]{plotCostMatrix}},
\code{\link[dtwSat]{plotAlignments}}, 
\code{\link[dtwSat]{plotMatches}}, 
\code{\link[dtwSat]{plotClassification}}, and 
\code{\link[dtwSat]{plotPatterns}}.
}

