% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLUCC.R
\docType{methods}
\name{plotLUCC}
\alias{plotLUCC}
\title{Plotting land use/cover changes (LUCC)}
\usage{
plotLUCC(x, type = "area", layer.levels, layer.labels = NULL,
  class.levels = NULL, class.labels = NULL, class.colors = NULL)
}
\arguments{
\item{x}{A \code{\link[raster]{Raster-class}}
\code{\link[raster]{brick}} or \code{\link[raster]{stack}} objects. 
A \link[base]{character} with the raster file name is also accepted.}

\item{type}{A \link[base]{character}. The plot type, ''map'', 
''area'', or ''change''. Default is ''map''.}

\item{layer.levels}{A \link[base]{character} or \link[base]{numeric}
vector with the layers to plot. For plot type ''change'' the minimum length 
is two}

\item{layer.labels}{A \link[base]{character} or \link[base]{numeric}
vector with the labels of the layers. It must have the same 
length as layer.levels. Default is NULL.}

\item{class.levels}{A \link[base]{character} or \link[base]{numeric}
vector with the levels of the raster values. Default is NULL.}

\item{class.labels}{A \link[base]{character} or \link[base]{numeric}
vector with the labels of the raster values. It must have the same 
length as class.levels. Default is NULL.}

\item{class.colors}{a set of aesthetic values. It must have the same 
length as class.levels. Default is NULL. See 
\link[ggplot2]{scale_fill_manual} for details.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Method for plotting the land use and land cover 
changes
}
\examples{

levels = c(seq_along(patterns.vignette), 255)
labels = c(names(patterns.vignette), "Unclassified")
colors = c("#996400", "#005500", "#D8B777", "#E6D219", "#E6BEC8", "#C8C8C8")
names(colors) = labels

x = system.file('lucc_MT.tif',  package = 'dtwSat')

### Plot maps
#gp1 = plotLUCC(x = x, type = "map", layer.labels = 2008:2013, 
#        class.levels = levels, class.labels = labels, 
#        class.colors = colors)
#gp1 

# Plot area 
#gp2 = plotLUCC(x = x, type = "area", layer.labels = 2008:2013, 
#        class.levels = levels, class.labels = labels, 
#        class.colors = colors)
#gp2

# Plot land use changes 
#gp3 = plotLUCC(x = x, type = "change", layer.labels = 2008:2013, 
#        class.levels = levels, class.labels = labels, 
#        class.colors = colors)
#gp3

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwApply}}, 
\code{\link[raster]{brick}}, and
\code{\link[raster]{stack}}
}

