% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-dtwSat.R
\docType{methods}
\name{plotPath}
\alias{plotPath}
\title{Plotting Time-Weighted DTW paths}
\usage{
plotPath(x, normalize = FALSE, show.dist = FALSE)
}
\arguments{
\item{x}{\code{\link[dtwSat]{twdtw}} object}

\item{normalize}{Normalized distance. Default is FALSE}

\item{show.dist}{Display dtw distance for each alignment}
}
\value{
object of class \code{\link[ggplot2]{ggplot}}
}
\description{
Methods for plotting Time-Weighted DTW objects
returned by dtwSat.
}
\examples{
names(query.list)
alig = dtwSat(query.list[["Soybean"]], template, weight = "logistic", alpha = 0.1, beta = 50, alignments = 4, keep = TRUE)
gp = plot(alig, normalize=TRUE, show.dist = TRUE)
gp
## Crate a list of dtwSat objects and plot them
gp.list = lapply(query.list, function(query){
     alig = dtwSat(query, template, weight = "logistic", alpha = 0.1, beta = 50, alignments = 4, keep = TRUE)
     plot(alig, normalize = TRUE, show.dist = TRUE)
})
grid.arrange(arrangeGrob(gp.list[[1]] + ggtitle(names(query.list)[1]) + theme(axis.title.x=element_blank(), legend.position="none"),
                         gp.list[[2]] + ggtitle(names(query.list)[2]) + theme(axis.title.x=element_blank(), legend.position="none"),
                         gp.list[[3]] + ggtitle(names(query.list)[3]) + theme(legend.position="none"),
                         nrow=3))
}

