% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-dtwSat.R
\docType{methods}
\name{plotPath}
\alias{plotPath}
\title{Plotting Time-Weighted DTW paths}
\usage{
plotPath(x, normalize = FALSE, show.dist = FALSE)
}
\arguments{
\item{x}{\code{\link[dtwSat]{twdtw}} object}

\item{normalize}{Normalized distance. Default is FALSE}

\item{show.dist}{Display dtw distance for each alignment}
}
\value{
object of class \code{\link[ggplot2]{ggplot}}
}
\description{
Methods for plotting Time-Weighted DTW objects
returned by dtwSat.
}
\examples{
names(query.list)
alig = dtwSat(query.list[["Soybean"]], template, weight = "logistic",
       alpha = 0.1, beta = 50, alignments = 4, keep = TRUE)
gp = plotPath(alig, normalize=TRUE, show.dist = TRUE)
gp
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

