% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-dtwSat.R
\docType{methods}
\name{plotPath}
\alias{plotPath}
\title{Plotting paths of dtwSat object}
\usage{
plotPath(x, normalize = FALSE, show.dist = FALSE, shift = c(-4, -1))
}
\arguments{
\item{x}{\code{\link[dtwSat]{dtwSat-class}} object}

\item{normalize}{Normalized distance. Default is FALSE}

\item{show.dist}{Display dtw distance for each alignment}

\item{shift}{A vector whose values shift the position of the text
in the plot. Argument Used with show.dist.}
}
\value{
object of class \code{\link[ggplot2]{ggplot}}

#'
}
\description{
Method for plotting the paths in the cost matrix
of the \code{\link[dtwSat]{dtwSat-class}} objects.
}
\examples{
names(query.list)
alig = dtwSat(query.list[["Soybean"]], template, weight = "logistic",
       alpha = 0.1, beta = 50, alignments = 4, keep = TRUE)
gp = plotPath(alig, normalize=TRUE, show.dist = TRUE)
gp
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{dtwSat}}, \code{\link[dtwSat]{twdtw}},
\code{\link[dtwSat]{plotAlignment}}
}

