% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPath.R
\docType{methods}
\name{plotPath}
\alias{plotPath}
\title{Plotting paths of twdtw alignments}
\usage{
plotPath(x, p.names, n.alignments = NULL)
}
\arguments{
\item{x}{An \code{\link[dtwSat]{twdtw-class}} object}

\item{p.names}{A \link[base]{character} or \link[base]{numeric}
vector with the patterns identification. If not declared the function 
will plot the paths for all patterns}

\item{n.alignments}{An \link[base]{integer} vector. The alignment indices 
to plot. If not declared the function will plot all possible alignments}
}
\value{
A \link[ggplot2]{ggplot} object
}
\description{
Method for plotting the minimum paths in the 
cost matrix of Time-Weighted DTW
}
\examples{

weight.fun = logisticWeight(alpha=-0.1, beta=100)
alig = twdtw(x=template, patterns=patterns.list, weight.fun = weight.fun, 
        normalize.patterns=TRUE, patterns.length=23, keep=TRUE)
       
gp1 = plotPath(alig, n.alignments=1:4)

gp1

gp2 = plotPath(alig, p.names=c("Cotton","Maize"), n.alignments=1:4)

gp2

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, 
\code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{plotCostMatrix}},
\code{\link[dtwSat]{plotAlignment}},
\code{\link[dtwSat]{plotMatch}}, and
\code{\link[dtwSat]{plotGroup}}
}

