% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPatterns.R
\docType{methods}
\name{plotPatterns}
\alias{plotPatterns}
\title{Plotting temporal patterns}
\usage{
plotPatterns(x, p.names, year = 2005)
}
\arguments{
\item{x}{An \code{\link[dtwSat]{twdtw-class}} object, a \code{\link[zoo]{zoo}},
or a list of \code{\link[zoo]{zoo}} objects.}

\item{p.names}{A \link[base]{character} or \link[base]{numeric}
vector with the patterns identification. If not declared the function 
will plot the paths for all patterns.}

\item{year}{An integer. The base year to shift the dates of the time series to. 
If NULL then it does not shif the time series. Default is 2005.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Method for plotting the temporal patterns.
}
\examples{

log_fun = logisticWeight(alpha=-0.1, beta=100)
matches = twdtw(x=example_ts, patterns=patterns.list, weight.fun = log_fun, 
        normalize.patterns=TRUE, patterns.length=23, keep=TRUE)
       
gp1 = plotPatterns(matches)
gp1

gp2 = plotPatterns(patterns.list)
gp2



}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtw-class}}, 
\code{\link[dtwSat]{twdtw}}, 
\code{\link[dtwSat]{plotPaths}}, 
\code{\link[dtwSat]{plotCostMatrix}},
\code{\link[dtwSat]{plotAlignments}}, 
\code{\link[dtwSat]{plotMatches}}, and 
\code{\link[dtwSat]{plotClassification}}.
}

