% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query.list}
\alias{query.list}
\title{Query time series}
\format{\preformatted{List of 3
 $ Soybean:‘zoo’ series from 2000-10-15 to 2001-02-28
  Data: num [1:18, 1:7] 0.365 0.379 0.411 0.467 0.542 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:18] "6" "7" "8" "9" ...
  .. ..$ : chr [1:7] "ndvi" "evi" "red" "nir" ...
  Index:  Date[1:18], format: "2000-10-15" "2000-10-23" ...
 $ Cotton :‘zoo’ series from 2001-01-03 to 2001-07-14
  Data: num [1:25, 1:7] 0.445 0.443 0.45 0.471 0.508 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:25] "16" "17" "18" "19" ...
  .. ..$ : chr [1:7] "ndvi" "evi" "red" "nir" ...
  Index:  Date[1:25], format: "2001-01-03" "2001-01-11" ...
 $ Maize  :‘zoo’ series from 2001-02-20 to 2001-06-28
  Data: num [1:17, 1:7] 0.462 0.479 0.526 0.593 0.669 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:17] "22" "23" "24" "25" ...
  .. ..$ : chr [1:7] "ndvi" "evi" "red" "nir" ...
  Index:  Date[1:17], format: "2001-02-20" "2001-02-28" ...
}}
\usage{
query.list
}
\description{
This dataset has a list of queries form the
crops: Soybean, Cotton, and Maize. The time series are based on
the MODIS product MOD13Q1 250 m 16 days.
The queries were exttracted from more than 100 samples of each class
using Generalized Additive Models (GAM), see
package \link[gam]{gam}.
}
\author{
Victor Maus, 2015-08-06
}
\keyword{datasets}

