% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query.list}
\alias{query.list}
\title{Data: query time series}
\format{A named \code{list} of 3 \link[zoo]{zoo} objects, "Soybean", "Cotton", 
and "Maize", whose indices are \code{\link[base]{Dates}} in the format "yyyy-mm-dd".
Each node has 6 attributes: "ndvi", "evi", "red", "nir", "blue", 
and "mir".}
\usage{
query.list
}
\description{
This dataset has a list of queries form the crops: Soybean,
Cotton, and Maize. These time series are based on the MODIS product 
MOD13Q1 250 m 16 days. The queries were extracted from more than 100 
samples of each class using Generalized Additive Models (GAM), see 
package \link[mgcv]{gam}.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\link[dtwSat]{template}

MOD13Q1 documentation: See 
\url{https://lpdaac.usgs.gov/dataset_discovery/modis/modis_products_table/mod13q1}
}
\keyword{datasets}

