% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{symmetric1}
\alias{symmetric2}
\alias{asymmetric}
\alias{rabinerJuangStepPattern}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dtw}{\code{\link[dtw]{symmetric1}}, \code{\link[dtw]{symmetric2}}, \code{\link[dtw]{asymmetric}}, \code{\link[dtw]{rabinerJuangStepPattern}}}
}}

