% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{shift_dates}
\alias{shift_dates}
\title{Shift dates}
\usage{
shift_dates(x, year)
}
\arguments{
\item{year}{the base year to shift the time series to.}

\item{object}{\code{\link[base]{data.frame}} objects.}
}
\value{
An object of the same class as the input \code{object}.
}
\description{
This function shifts the dates of the time series to a
given base year.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
