% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftDates.R
\docType{methods}
\name{shiftDates}
\alias{shiftDates}
\title{Shift dates of zoo object}
\usage{
shiftDates(x, year)
}
\arguments{
\item{x}{a \link[zoo]{zoo} object.}

\item{year}{An integer. The base year to shift the dates of the time series to.}
}
\value{
a \link[zoo]{zoo} object shifted to the given year.
}
\description{
This function shifts the dates of the time series to a given base year.
}
\examples{
dates = seq(from = as.Date("2014-09-01"), to = as.Date("2015-03-01"), by = "15 day")
x = zoo(seq_along(dates), dates)
y = shiftDates(x, year=2005)
index(x)
index(y)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

