% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{shiftDates}
\alias{shiftDates}
\alias{shiftDates,twdtwTimeSeries-method}
\alias{shiftDates-twdtwTimeSeries}
\alias{shiftDates,list-method}
\alias{shiftDates-list}
\alias{shiftDates,zoo-method}
\alias{shiftDates-zoo}
\title{Shift dates}
\usage{
shiftDates(object, year = NULL)

\S4method{shiftDates}{twdtwTimeSeries}(object, year = NULL)

\S4method{shiftDates}{list}(object, year = NULL)

\S4method{shiftDates}{zoo}(object, year = NULL)
}
\arguments{
\item{object}{\code{\link[dtwSat]{twdtwTimeSeries}} objects, 
\code{\link[zoo]{zoo}} objects or a list of \code{\link[zoo]{zoo}} objects.}

\item{year}{the base year to shift the time series to.}
}
\value{
An object of the same class as the input \code{object}.
}
\description{
This function shifts the dates of the time series to a 
given base year.
}
\examples{
patt = twdtwTimeSeries(MOD13Q1.patterns.list)
npatt = shiftDates(patt, year=2005)
index(patt)
index(npatt)

}
\references{
\insertRef{Maus:2019}{dtwSat}
  
  \insertRef{Maus:2016}{dtwSat}

\insertRef{Maus:2019}{dtwSat}
  
  \insertRef{Maus:2016}{dtwSat}
}
\seealso{
\code{\link[dtwSat]{twdtwTimeSeries-class}}
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
