% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitDataset.R
\docType{methods}
\name{splitDataset}
\alias{splitDataset}
\title{split the time series samples and create patterns}
\usage{
splitDataset(timeseries, ref, times = 1, p = 0.1, mc.cores = 1, ...)
}
\arguments{
\item{timeseries}{A list of \code{\link[zoo]{zoo}} objects with the time series.}

\item{ref}{A vector with the identification of each time series. 
It must have the same length as \code{x}.}

\item{times}{Number of partitions to create.}

\item{p}{the percentage of data that goes to training. 
See \code{\link[caret]{createDataPartition}} for details.}

\item{mc.cores}{The number of cores to use, See \code{\link[parallel]{mclapply}} 
for details.}

\item{...}{Other arguments to be passed to \code{\link[dtwSat]{createPattern}}.}
}
\value{
A \code{\link[base]{list}} of three attributes: the trained \code{pattrens},
the validation dataset \code{x}, and the reference for the validation dataset \code{ref}.
}
\description{
This function splits the set of samples for 
training and validation. The function uses stratified sampling. 
Then a simple random sampling is applied within each stratum.
}
\examples{

\dontrun{
set.seed(1)
data_folder = system.file("lucc_MT/data", package = "dtwSat")
field_samples = read.csv(paste(data_folder,"samples.csv", sep = "/"))
proj_str = scan(paste(data_folder,"samples_projection", sep = "/"), 
           what = "character")
reference = as.character(field_samples[["class"]])
load(system.file("lucc_MT/ts_list.RData", package="dtwSat"))

x = splitted_dataset = splitDataset(timeseries = ts_list, ref = reference, 
    times=2, p=0.1, mc.cores=1, freq=8, from="2007-09-01", to="2008-09-01", 
    formula = y ~ s(time, bs="cc"))

names(x)
names(x$Resample1)
plotPatterns(x$Resample1$patterns)
}

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwAssessment}}
}

