% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template}
\alias{template}
\title{Template time series}
\format{\preformatted{‘zoo’ series from 2009-08-05 to 2013-07-31
  Data: num [1:93, 1:7] 0.317 0.261 0.179 0.279 0.177 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:93] "218" "219" "220" "221" ...
  ..$ : chr [1:7] "ndvi" "evi" "red" "nir" ...
  Index:  Date[1:93], format: "2009-08-05" "2009-08-28" "2009-09-13" "2009-09-28" ...
}}
\usage{
template
}
\description{
This dataset has a template time series based on the
MODIS product MOD13Q1 250 m 16 days. See package pkg{rwtss} that is an
R client for Web Time Series Service
\url{https://github.com/albhasan/rwtss.git}.
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\keyword{datasets}

