% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/twdtw.R
\docType{methods}
\name{twdtw}
\alias{twdtw}
\title{Multidimensional Time-Weighted DTW Alignment}
\usage{
twdtw(query, template, weight = NULL, dist.method = "Euclidean",
  theta = NULL, alpha = NULL, beta = NULL, alignments = NULL,
  step.matrix = symmetric1, window.function = noWindow, keep = FALSE, ...)
}
\arguments{
\item{query}{A \link[zoo]{zoo} object with the multidimensional time series.}

\item{template}{A \link[zoo]{zoo} object with the template time series similar
to \code{query}. The \code{template} must have the same number of attributes
and be equal or longer than the \code{query},
\emph{i.e.} \code{nrow(query)<=nrow(template)}.}

\item{weight}{A character. ''linear'' for linear weight or ''logistic''
for logistic weight. Default is NULL that runs the original dtw method,
\emph{i.e.} without time weight.}

\item{dist.method}{A character. Method to derive the local cost matrix.
Default is ''Euclidean'' See \code{\link[proxy]{dist}} in package
\pkg{proxy}.}

\item{theta}{A number. Parameter for ''linear'' time weight. For \code{theta=1}
the time weight is equal to the number of elapsed days.}

\item{alpha}{A number. The steepness of logistic weight.}

\item{beta}{A number. The midpoint of logistic weight.}

\item{alignments}{An integer. The maximun number of alignments to
perform. NULL will return all possible alignments.}

\item{step.matrix}{see \code{\link[dtw]{stepPattern}} in package \pkg{dtw}}

\item{window.function}{see parameter \code{window.type} in \code{\link[dtw]{dtw}}}

\item{keep}{preserve the cost matrix, inputs, and other internal structures.
Default is FALSE}

\item{...}{other parameters}
}
\value{
object of class \code{\link[dtwSat]{dtwSat-class}}
}
\description{
This function performs a multidimensional Time-Weighted DTW
analysis and retrieves the alignments of a query within a time series.
}
\examples{
names(query.list)
alig = twdtw(query.list[["Soybean"]], template, weight = "logistic",
       alpha = 0.1, beta = 50, alignments=4)
alig
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{mtwdtw}}, \code{\link[dtwSat]{dtwSat-class}}
}

