% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dtwSat.R
\docType{methods}
\name{twdtw}
\alias{twdtw}
\title{Multidimensional Time-Weighted DTW analysis}
\usage{
twdtw(query, template, weight = NULL, dist.method = "Euclidean",
  theta = NULL, alpha = NULL, beta = NULL, alignments = NULL,
  step.matrix = symmetric1, window.function = noWindow, keep = FALSE)
}
\arguments{
\item{query}{A zoo object with the multidimensional time series.}

\item{template}{A zoo object with the template time series.
It must be iguel or be equal or longer than the length of the query and
the same number of dimensions.}

\item{weight}{A character. ''linear'' for linear weight or ''logistic''
for logistic weight. Default is NULL that runs the original dtw method.}

\item{dist.method}{A character. Method to derive the local cost matrix.
Default is ''Euclidean'' See \code{\link[proxy]{dist}} in package
\pkg{proxy}}

\item{theta}{A real. Parameter for linear weight. It is the slope
of the linear TWDTW. For theta equal 1 the time weight is equal to the
number of elapsed days. Default is NULL.}

\item{alpha}{A real. The steepness of logistic method. Default is NULL.}

\item{beta}{A real. The midpoint of logistic method. Default is NULL.}

\item{alignments}{An integer. The maximun number of alignments to
perform. Default is NULL to return all possible alignment.}

\item{step.matrix}{see \code{\link{stepPattern}} in package \pkg{dtw}}

\item{window.function}{see \code{window.type} in package \pkg{dtw}}

\item{keep}{preserve the cost matrix, inputs, and other internal structures.
Default is FALSE}
}
\description{
This function performs a multidimensional Time-Weighted DTW
analysis and retrieves one or more possible alignments of a query within
a time series.
}
\examples{
#alig = mtwdtw(query.list, template, weight = "logistic", alpha = 0.1, beta = 50, alignments=4)
#alig
}

