% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-twdtwMatches.R, R/getTimeSeries.R, R/getPatterns.R, R/is.R, R/joinAlignments.R, R/labels.R, R/show.R, R/utils.R
\docType{methods}
\name{twdtwMatches-class}
\alias{[,twdtwMatches,ANY,ANY,ANY-method}
\alias{[[,twdtwMatches,numeric,ANY-method}
\alias{as.list,twdtwMatches-method}
\alias{as.list,twdtwRaster-method}
\alias{getPatterns,twdtwMatches-method}
\alias{getTimeSeries,twdtwMatches-method}
\alias{is.twdtwMatches}
\alias{is.twdtwMatches,ANY-method}
\alias{joinAlignments,twdtwMatches-method}
\alias{labels,twdtwMatches-method}
\alias{length,twdtwMatches-method}
\alias{show,twdtwMatches-method}
\alias{twdtwMatches,twdtwTimeSeries,twdtwTimeSeries,list-method}
\alias{twdtwMatches-class}
\title{class "twdtwMatches"}
\usage{
\S4method{twdtwMatches}{twdtwTimeSeries,twdtwTimeSeries,list}(timeseries,
  patterns, alignments)

\S4method{getTimeSeries}{twdtwMatches}(object, labels = NULL)

\S4method{getPatterns}{twdtwMatches}(object, labels = NULL)

\S4method{is.twdtwMatches}{ANY}(x)

\S4method{joinAlignments}{twdtwMatches}(...)

\S4method{labels}{twdtwMatches}(object)

\S4method{show}{twdtwMatches}(object)

\S4method{length}{twdtwMatches}(x)

\S4method{as.list}{twdtwMatches}(x)

\S4method{as.list}{twdtwRaster}(x)

\S4method{[}{twdtwMatches,ANY,ANY,ANY}(x, i, drop = TRUE)

\S4method{[[}{twdtwMatches,numeric,ANY}(x, i)
}
\arguments{
\item{timeseries}{an object of class twdtwTimeSeries.}

\item{patterns}{an object of class twdtwTimeSeries.}

\item{alignments}{an object of class list with the TWDTW results. 
For more details see slot alignments of twdtwMatches-class.}

\item{object}{an object of class twdtwMatches.}

\item{labels}{a vector with labels of the time series.}

\item{x}{an object of class twdtwMatches.}

\item{...}{objects of class twdtwMatches}

\item{i}{indices of the alignments.}

\item{drop}{if TRUE returns a data.frame, if FALSE returns a list. 
Default is TRUE.}
}
\description{
Class for Time-Weighted Dynamic Time Warping results.
}
\section{Methods (by generic)}{
\itemize{
\item \code{twdtwMatches}: Create object of class twdtwMatches.

\item \code{getTimeSeries}: Get subset of time series from objects of class twdtwMatches.

\item \code{getPatterns}: Get temporal patterns from objects of class twdtwMatches.

\item \code{is.twdtwMatches}: Check if the object belongs to the class twdtwMatches.

\item \code{joinAlignments}: Join TWDTW alignments from two or more objects of class twdtwMatches
}}
\section{Slots }{

\describe{
 \item{\code{timeseries}:}{An object of class \code{\link[dtwSat]{twdtwTimeSeries-class}} with the satellite time series.}
 \item{\code{pattern}:}{An object of class \code{\link[dtwSat]{twdtwTimeSeries-class}} with the temporal patterns.}
 \item{\code{alignments}:}{A \code{\link[base]{list}} of TWDTW results with the same length as 
 the \code{timeseries}. Each element in this list has the following results for each temporal pattern 
 in \code{patterns}:
      \cr\code{from}: a vector with the starting dates of each match in the format "YYYY-MM-DD",
      \cr\code{to}: a vector with the ending dates of each match in the format "YYYY-MM-DD", 
	     \cr\code{distance}: a vector with TWDTW dissimilarity measure, and
	     \cr\code{K}: the number of matches of the pattern.
 }
	\item{This list might have additional elements:}{ if \code{keep=TRUE} in the \code{twdtw} call 
	the list is extended to include internal structures used during the TWDTW computation: 
      \cr\code{costMatrix}: cumulative cost matrix,
      \cr\code{directionMatrix}: directions of steps that would be taken from each element of matrix,
      \cr\code{startingMatrix}: the starting points of each element of the matrix,
      \cr\code{stepPattern}: \code{\link[dtw]{stepPattern}} used for the 
      computation, see package \code{\link[dtw]{dtw}},
      \cr\code{N}: the length of the \code{pattern}, 
      \cr\code{M}: the length of the time series \code{timeseries}, 
      \cr\code{timeWeight}: time weight matrix,
      \cr\code{localMatrix}: local cost matrix,
	     \cr\code{matching}: A list whose elements have the matching points for 
	     each match between pattern the time series, such that:
         \cr--\code{index1}: a vector with matching points of the pattern, and
         \cr--\code{index2}: a vector with matching points of the time series.
      }
}
}
\examples{
ts = twdtwTimeSeries(timeseries=example_ts.list)
patterns = twdtwTimeSeries(timeseries=patterns.list)
matches = twdtwApply(x = ts, y = patterns, mc.cores=2)
class(matches)
length(matches)
matches[1]
matches[[1]]
matches[]
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwApply}} and 
\code{\link[dtwSat]{twdtwTimeSeries-class}}
}

