% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-twdtwTimeSeries.R, R/dimensions.R, R/getTimeSeries.R, R/is.R, R/joinPatterns.R, R/joinTimeSeries.R, R/labels.R, R/normalizePatterns.R, R/show.R
\docType{methods}
\name{twdtwTimeSeries-class}
\alias{dim,twdtwTimeSeries-method}
\alias{getTimeSeries,twdtwTimeSeries-method}
\alias{is.twdtwTimeSeries}
\alias{is.twdtwTimeSeries,ANY-method}
\alias{joinPatterns,twdtwTimeSeries-method}
\alias{joinTimeSeries,twdtwTimeSeries-method}
\alias{labels,twdtwTimeSeries-method}
\alias{length,twdtwTimeSeries-method}
\alias{ncol,twdtwTimeSeries-method}
\alias{normalizePatterns,twdtwTimeSeries,numeric-method}
\alias{nrow,twdtwTimeSeries-method}
\alias{show,twdtwTimeSeries-method}
\alias{twdtwTimeSeries,ANY-method}
\alias{twdtwTimeSeries-class}
\title{class "twdtwTimeSeries"}
\usage{
\S4method{twdtwTimeSeries}{ANY}(..., timeseries = list(...), labels)

\S4method{dim}{twdtwTimeSeries}(x)

\S4method{nrow}{twdtwTimeSeries}(x)

\S4method{ncol}{twdtwTimeSeries}(x)

\S4method{length}{twdtwTimeSeries}(x)

\S4method{getTimeSeries}{twdtwTimeSeries}(object, labels = NULL)

\S4method{is.twdtwTimeSeries}{ANY}(x)

\S4method{joinPatterns}{twdtwTimeSeries}(...)

\S4method{joinTimeSeries}{twdtwTimeSeries}(...)

\S4method{labels}{twdtwTimeSeries}(object)

\S4method{normalizePatterns}{twdtwTimeSeries,numeric}(object, length = NULL)

\S4method{show}{twdtwTimeSeries}(object)
}
\arguments{
\item{...}{objects of class \code{\link[zoo]{zoo}}.}

\item{timeseries}{a list of \code{\link[zoo]{zoo}} objects.}

\item{labels}{a vector with labels of the time series.}

\item{x}{an object of class twdtwTimeSeries.}

\item{object}{an object of class twdtwTimeSeries.}

\item{length}{a number. If not informed the function will use the 
length of the longest pattern.}
}
\description{
Class for set of irregular time series.
}
\section{Methods (by generic)}{
\itemize{
\item \code{twdtwTimeSeries}: Create object of class twdtwTimeSeries.

\item \code{dim}: Dimensions of time series from objects of class twdtwTimeSeries.

\item \code{nrow}: Number of rows of time series from objects of class twdtwTimeSeries.

\item \code{ncol}: Number of columns of time series from objects of class twdtwTimeSeries.

\item \code{length}: Number of time series in objects of class twdtwTimeSeries.

\item \code{getTimeSeries}: Get time series from objects of class twdtwTimeSeries

\item \code{is.twdtwTimeSeries}: Check if the object belongs to the class twdtwTimeSeries.

\item \code{joinPatterns}: Join patterns (time series) from two or more 
objects of class twdtwTimeSeries. This function removes patterns (time series) 
with duplicated label.

\item \code{joinTimeSeries}: Join time series from two or more objects 
of class twdtwTimeSeries

\item \code{labels}: Labels of the time series in object of class twdtwTimeSeries.

\item \code{normalizePatterns}: Get temporal patterns from objects of class twdtwTimeSeries.

\item \code{show}: Show object of class twdtwTimeSeries
}}
\section{Slots }{

\describe{
 \item{\code{timeseries}:}{A list of \code{\link[zoo]{zoo}} objects.}
 \item{\code{labels}:}{A vector of class \code{\link[base]{factor}} with time series labels.}
}
}
\examples{
# Creating new object of class twdtwTimeSeries  
new("twdtwTimeSeries")

sapply(example_ts.list, class)
new("twdtwTimeSeries", timeseries = example_ts.list)
new("twdtwTimeSeries", timeseries = example_ts.list, labels = c(1,2))

# Showing objects of class twdtwTimeSeries 
ex_ts = twdtwTimeSeries(timeseries = example_ts.list, labels = c(1,2))
show(ex_ts)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

