% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-twdtwTimeSeries.R, R/methods-getTimeSeries.R, R/methods-is.R, R/methods-labels.R, R/methods-length.R, R/methods-show.R
\docType{methods}
\name{twdtwTimeSeries-class}
\alias{getTimeSeries,twdtwTimeSeries-method}
\alias{is.twdtwTimeSeries}
\alias{is.twdtwTimeSeries,twdtwTimeSeries-method}
\alias{labels,twdtwTimeSeries-method}
\alias{length,twdtwTimeSeries-method}
\alias{show,twdtwTimeSeries-method}
\alias{twdtwTimeSeries}
\alias{twdtwTimeSeries,ANY-method}
\alias{twdtwTimeSeries-class}
\title{class "twdtwTimeSeries"}
\usage{
\S4method{twdtwTimeSeries}{ANY}(..., timeseries = list(...), labels)

\S4method{getTimeSeries}{twdtwTimeSeries}(object, labels = NULL)

\S4method{is.twdtwTimeSeries}{twdtwTimeSeries}(x)

\S4method{labels}{twdtwTimeSeries}(object)

\S4method{length}{twdtwTimeSeries}(x)

\S4method{show}{twdtwTimeSeries}(object)
}
\arguments{
\item{...}{objects of class \code{\link[zoo]{zoo}}.}

\item{timeseries}{a list of \code{\link[zoo]{zoo}} objects.}

\item{labels}{a vector with labels of the time series.}

\item{object}{an object of class twdtwTimeSeries.}

\item{x}{an object of class twdtwTimeSeries.}
}
\description{
Class for set of irregular time series.
}
\section{Methods (by generic)}{
\itemize{
\item \code{twdtwTimeSeries}: Create object of class twdtwTimeSeries

\item \code{getTimeSeries}: Get time series from objects of class twdtwTimeSeries

\item \code{is.twdtwTimeSeries}: Check if the object belongs to the class twdtwTimeSeries

\item \code{labels}: Labels of the time series in object of class twdtwTimeSeries

\item \code{length}: Number of time series in objects of class twdtwTimeSeries

\item \code{show}: Show object of class twdtwTimeSeries
}}
\section{Slots }{

\describe{
 \item{\code{timeseries}:}{A list of \code{\link[zoo]{zoo}} objects.}
 \item{\code{labels}:}{A vector of class \code{\link[base]{factor}} with time series labels.}
}
}
\examples{
# Creating new object of class twdtwTimeSeries  
new("twdtwTimeSeries")

sapply(example_ts.list, class)
new("twdtwTimeSeries", timeseries = example_ts.list)
new("twdtwTimeSeries", timeseries = example_ts.list, labels = c(1,2))

# Creating objects of class twdtwTimeSeries 
class(example_ts)
twdtwTimeSeries(example_ts)
sapply(example_ts.list, class)
twdtwTimeSeries(timeseries = example_ts.list)
twdtwTimeSeries(timeseries = example_ts.list, labels = c(1,2))
twdtwTimeSeries(example_ts.list[[1]], example_ts.list[[2]])

# Getting time series from objects of class twdtwTimeSeries
ex_ts = twdtwTimeSeries(timeseries=example_ts.list)
getTimeSeries(ex_ts, 1)

# Checking if object belongs to the class twdtwTimeSeries
ex_ts = twdtwTimeSeries(timeseries = example_ts.list)
is.twdtwTimeSeries(ex_ts)

# Getting labels of time series in objects of class twdtwTimeSeries
ex_ts = twdtwTimeSeries(timeseries = example_ts.list)
length(ex_ts)

# Getting the number of time series in objects of class twdtwTimeSeries
ex_ts = twdtwTimeSeries(timeseries = example_ts.list)
length(ex_ts)

# Showing objects of class twdtwTimeSeries 
ex_ts = twdtwTimeSeries(timeseries = example_ts.list, labels = c(1,2))
show(ex_ts)

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

