% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-twdtwTimeSeries.R, R/getTimeSeries.R, R/is.R, R/joinPatterns.R, R/joinTimeSeries.R, R/labels.R, R/normalizePatterns.R, R/show.R, R/utils.R
\docType{methods}
\name{twdtwTimeSeries-class}
\alias{[,twdtwTimeSeries,ANY,ANY,ANY-method}
\alias{[[,twdtwTimeSeries,ANY,ANY-method}
\alias{as.list,twdtwTimeSeries-method}
\alias{dim,twdtwTimeSeries-method}
\alias{getTimeSeries,twdtwTimeSeries-method}
\alias{is.twdtwTimeSeries}
\alias{is.twdtwTimeSeries,ANY-method}
\alias{joinPatterns,twdtwTimeSeries-method}
\alias{joinTimeSeries,twdtwTimeSeries-method}
\alias{labels,twdtwTimeSeries-method}
\alias{length,twdtwTimeSeries-method}
\alias{ncol,twdtwTimeSeries-method}
\alias{normalizePatterns,twdtwTimeSeries,numeric-method}
\alias{nrow,twdtwTimeSeries-method}
\alias{show,twdtwTimeSeries-method}
\alias{twdtwTimeSeries,ANY-method}
\alias{twdtwTimeSeries-class}
\title{class "twdtwTimeSeries"}
\usage{
\S4method{twdtwTimeSeries}{ANY}(..., timeseries = list(...), labels)

\S4method{getTimeSeries}{twdtwTimeSeries}(object, labels = NULL)

\S4method{is.twdtwTimeSeries}{ANY}(x)

\S4method{joinPatterns}{twdtwTimeSeries}(...)

\S4method{joinTimeSeries}{twdtwTimeSeries}(...)

\S4method{labels}{twdtwTimeSeries}(object)

\S4method{normalizePatterns}{twdtwTimeSeries,numeric}(object, length = NULL)

\S4method{show}{twdtwTimeSeries}(object)

\S4method{dim}{twdtwTimeSeries}(x)

\S4method{nrow}{twdtwTimeSeries}(x)

\S4method{ncol}{twdtwTimeSeries}(x)

\S4method{length}{twdtwTimeSeries}(x)

\S4method{as.list}{twdtwTimeSeries}(x)

\S4method{[}{twdtwTimeSeries,ANY,ANY,ANY}(x, i)

\S4method{[[}{twdtwTimeSeries,ANY,ANY}(x, i)
}
\arguments{
\item{...}{objects of class \code{\link[zoo]{zoo}}.}

\item{timeseries}{a list of \code{\link[zoo]{zoo}} objects.}

\item{labels}{a vector with labels of the time series.}

\item{object}{an object of class twdtwTimeSeries.}

\item{x}{an object of class twdtwTimeSeries.}

\item{length}{a number. If not informed the function will use the 
length of the longest pattern.}

\item{i}{indices of the time series.}
}
\description{
Class for set of irregular time series.
}
\section{Methods (by generic)}{
\itemize{
\item \code{twdtwTimeSeries}: Create object of class twdtwTimeSeries.

\item \code{getTimeSeries}: Get subset of time series from objects of class twdtwTimeSeries.

\item \code{is.twdtwTimeSeries}: Check if the object belongs to the class twdtwTimeSeries.

\item \code{joinPatterns}: Join patterns (time series) from two or more 
objects of class twdtwTimeSeries. This function removes patterns (time series) 
with duplicated label.

\item \code{joinTimeSeries}: Join time series from two or more objects 
of class twdtwTimeSeries

\item \code{normalizePatterns}: Get temporal patterns from objects of class twdtwTimeSeries.
}}
\section{Slots }{

\describe{
 \item{\code{timeseries}:}{A list of \code{\link[zoo]{zoo}} objects.}
 \item{\code{labels}:}{A vector of class \code{\link[base]{factor}} with time series labels.}
}
}
\examples{
# Creating new object of class twdtwTimeSeries  
ptt = new("twdtwTimeSeries", timeseries = patterns.list, labels = names(patterns.list))
class(ptt)
length(ptt)
nrow(ptt)
ncol(ptt)
dim(ptt)
ptt[1]
ptt[[1]]
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}

