% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-twdtwTimeSeries.R, R/methods.R
\name{twdtwTimeSeries-class}
\alias{twdtwTimeSeries-class}
\alias{twdtwTimeSeries}
\alias{twdtwTimeSeries,ANY-method}
\alias{twdtwTimeSeries-create}
\alias{dim,twdtwTimeSeries-method}
\alias{index,twdtwTimeSeries-method}
\alias{nrow,twdtwTimeSeries-method}
\alias{ncol,twdtwTimeSeries-method}
\alias{length,twdtwTimeSeries-method}
\alias{as.list,twdtwTimeSeries-method}
\alias{as.data.frame,twdtwTimeSeries-method}
\alias{[,twdtwTimeSeries,ANY,ANY,ANY-method}
\alias{[[,twdtwTimeSeries,ANY,ANY-method}
\alias{labels,twdtwTimeSeries-method}
\alias{levels,twdtwTimeSeries-method}
\alias{show,twdtwTimeSeries-method}
\alias{as.twdtwTimeSeries,ANY-method}
\alias{as.twdtwTimeSeries}
\alias{is.twdtwTimeSeries,ANY-method}
\alias{is.twdtwTimeSeries}
\title{class "twdtwTimeSeries"}
\usage{
\S4method{twdtwTimeSeries}{ANY}(..., labels = NULL)

\S4method{dim}{twdtwTimeSeries}(x)

\S4method{index}{twdtwTimeSeries}(x)

\S4method{nrow}{twdtwTimeSeries}(x)

\S4method{ncol}{twdtwTimeSeries}(x)

\S4method{length}{twdtwTimeSeries}(x)

\S4method{as.list}{twdtwTimeSeries}(x)

\S4method{as.data.frame}{twdtwTimeSeries}(x)

\S4method{[}{twdtwTimeSeries,ANY,ANY,ANY}(x, i)

\S4method{[[}{twdtwTimeSeries,ANY,ANY}(x, i)

\S4method{labels}{twdtwTimeSeries}(object)

\S4method{levels}{twdtwTimeSeries}(x)

\S4method{show}{twdtwTimeSeries}(object)

\S4method{as.twdtwTimeSeries}{ANY}(x)

\S4method{is.twdtwTimeSeries}{ANY}(x)
}
\arguments{
\item{...}{\code{\link[dtwSat]{twdtwTimeSeries}} objects, 
\code{\link[zoo]{zoo}} objects or a list of \code{\link[zoo]{zoo}} objects.}

\item{labels}{a vector with labels of the time series.}

\item{x}{an object of class twdtwTimeSeries.}

\item{i}{indices of the time series.}

\item{object}{an object of class twdtwTimeSeries.}
}
\description{
Class for setting irregular time series.
}
\section{Methods (by generic)}{
\itemize{
\item \code{twdtwTimeSeries(ANY)}: Create object of class twdtwTimeSeries.

\item \code{as.twdtwTimeSeries(ANY)}: convert list of data.frame to class twdtwTimeSeries.

\item \code{is.twdtwTimeSeries(ANY)}: Check if the object belongs to the class twdtwTimeSeries.

}}
\section{Slots }{

\describe{
 \item{\code{timeseries}:}{A list of \code{\link[zoo]{zoo}} objects.}
 \item{\code{labels}:}{A vector of class \code{\link[base]{factor}} with time series labels.}
}
}

\examples{
# Creating a new object of class twdtwTimeSeries  
ptt = new("twdtwTimeSeries", timeseries = MOD13Q1.patterns.list, 
           labels = names(MOD13Q1.patterns.list))
class(ptt)
labels(ptt)
levels(ptt)
length(ptt)
nrow(ptt)
ncol(ptt)
dim(ptt)
# Creating objects of class twdtwTimeSeries from zoo objects
ts = twdtwTimeSeries(MOD13Q1.ts)
ts 

# Creating objects of class twdtwTimeSeries from list of zoo objects 
patt = twdtwTimeSeries(MOD13Q1.patterns.list)
patt

# Joining objects of class twdtwTimeSeries 
tsA = twdtwTimeSeries(MOD13Q1.ts.list[[1]], labels = "A")
tsB = twdtwTimeSeries(B = MOD13Q1.ts.list[[2]])
ts = twdtwTimeSeries(tsA, tsB, C=MOD13Q1.ts)
ts
 
}
\references{
\insertRef{Maus:2019}{dtwSat}
  
  \insertRef{Maus:2016}{dtwSat}
}
\seealso{
\code{\link[dtwSat]{twdtwMatches-class}}, 
\code{\link[dtwSat]{twdtwRaster-class}}, 
\code{\link[dtwSat]{getTimeSeries}}, and
\code{\link[dtwSat]{twdtwApply}}
}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
