% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-twdtwTimeSeries.R
\name{twdtwTimeSeries}
\alias{twdtwTimeSeries}
\title{Create twdtwTimeSeries object}
\usage{
twdtwTimeSeries(..., timeseries, labels)
}
\arguments{
\item{...}{objects of class \code{\link[zoo]{zoo}}.}

\item{timeseries}{a list of \code{\link[zoo]{zoo}} objects.}

\item{labels}{a vector with labels of the time series.}
}
\description{
Create object of class twdtwTimeSeries.
}
\examples{
# Creating objects of class twdtwTimeSeries 
ptt = twdtwTimeSeries(timeseries = patterns.list)
ptt 

}
\author{
Victor Maus, \email{vwmaus1@gmail.com}
}
\seealso{
\code{\link[dtwSat]{twdtwTimeSeries-class}},
\code{\link[dtwSat]{getTimeSeries}}, and 
\code{\link[dtwSat]{twdtwApply}}
}

