% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{methods}
\name{waveletSmoothing}
\alias{waveletSmoothing}
\title{Wavelet filter}
\usage{
waveletSmoothing(x, timeline = NULL, frequency = NULL, wf = "la8",
  J = 1, boundary = "periodic", ...)
}
\arguments{
\item{x}{A zoo object with the time series}

\item{timeline}{A time line vector for the internalsput}

\item{frequency}{A numeric with the frequency for internalsput time series}

\item{wf}{Name of the wavelet filter to use in the decomposition.
Default is "la8"}

\item{J}{Specifies the depth of the decomposition. This must be a number
less than or equal to log(length(x),2). Default is 1}

\item{boundary}{Character string specifying the boundary condition.
Default is "periodic"}

\item{...}{see parameters of \code{\link[waveslim]{mra}} in the
packege \pkg{waveslim}}
}
\description{
This function performs a smoothing algorithm to
the time series. It computes the a discreat wavelet
smoothing for each dimension in the imput time series.
}
\examples{
# Wavelet filter
sy = waveletSmoothing(x=template, frequency=16, wf = "la8", J=1, boundary = "periodic")
# Plot raw EVI and filtered EVI
df.y = melt(data.frame(Time=index(template), Raw=template$evi), id="Time")
df.sy = melt(data.frame(Time=index(sy), Wavelet=sy$evi), id="Time")
df = rbind(df.y, df.sy)
ggplot(df, aes(x=Time, y=value, group=variable, colour=variable)) +
  geom_line() +
    ylab("EVI")

# Plot all filter bands
df = melt(data.frame(Time=index(sy), sy), id="Time")
ggplot(df, aes(x=Time, y=value, group=variable, colour=variable)) +
  geom_line()
}

